### 2020-09-17 #2684 @whedon
[PRE REVIEW]: The Pencil Code, a modular MPI code for partial differential equations and particles: multipurpose and multiuser-maintained
**Submitting author:** @AxelBrandenburg (<a href="http://orcid.org/0000-0002-7304-021X">Axel Brandenburg</a>)
**Repository:** <a href="https://github.com/pencil-code/pencil-code" target ="_blank">https://github.com/pencil-code/pencil-code</a>
**Version:** v2020.08.06
