### 2020-09-17 #2684 @kyleniemeyer
Hi @AxelBrandenburg, thanks for your interest in JOSS. It looks like our system is having trouble with your paper—I think you need to remove the `paper.tex` from the paper directory, because that might be confusing things.

cc @openjournals/dev
