### 2020-09-18 #2684 @AxelBrandenburg
Thanks for your work on this. I have now removed paper.tex from the paper directory.
Cheers, Axel
