### 2020-09-18 #2684 @harpolea
:wave: @zingale, would you be interested in reviewing this submission for JOSS?
