### 2020-09-18 fc90d83
-      if (linsert_particles_continuously .and. (ncpus .gt. 1)) then +      if (lpersist .and. linsert_particles_continuously .and. (ncpus .gt. 1)) then          if (random_gen .eq. "min_std") then            if (lroot) then              print*,'You can not set random_gen="min_std" when' -            print*,'linsert_particles_continuously=T.....EXITING!' +            print*,'linsert_particles_continuously=T and lpersist=T.....EXITING!'            endif            call fatal_error('initialize_particles','')          endif
