### 2020-09-18 ec6f885
MR: modified get_grid_mn for grid coarsening near poles
