+++ src/grid.f90
@@ -2441,18 +2441,22 @@ module Grid
         dline: if (lspherical_coords) then
           dline_1(:,1) = dx_1(l1:l2)
           dline_1(:,2) = r1_mn * dy_1(m)
-          !if (lpole(2) .and. lcoarse) then
-          !  if (lfirst_proc_y .and. m<m1+1.5*ncoarse .and. m>=m1) then
-          !    nphi = max(mod(int(ncoarse/(m-m1+1)),ncoarse),1)
-          !  elseif (llast_proc_y .and. m>m2-1.5*ncoarse .and. m<=m2) then
-          !    nphi = max(mod(int(ncoarse/(m2-m+1)),ncoarse),1)
-          !  else
-          !    nphi = 1
-          !  endif
-          !  dline_1(:,3) = r1_mn * sin1th(m) * dz_1(n)/nphi
-          !else
-          dline_1(:,3) = r1_mn * sin1th(m) * dz_1(n)
-          !endif
+!
+          if (lpole(2) .and. lcoarse) then
+            if (lfirst_proc_y .and. m<m1+ncoarse .and. m>=m1) then
+              nphi = mod(int(ncoarse/(m-m1+1)),ncoarse+1)
+!if (n==n1) print*, 'first:m,nphi=',m,nphi 
+            elseif (llast_proc_y .and. m>m2-ncoarse .and. m<=m2) then
+              nphi = mod(int(ncoarse/(m2-m+1)),ncoarse+1)
+!if (n==n1) print*, 'last:m,nphi=',m,nphi 
+            else
+              nphi = 1
+            endif
+            dline_1(:,3) = r1_mn * sin1th(m) * dz_1(n)/nphi
+          else
+            dline_1(:,3) = r1_mn * sin1th(m) * dz_1(n)
+          endif
+!
         else if (lcylindrical_coords) then dline
           dline_1(:,1) = dx_1(l1:l2)
           dline_1(:,2) = rcyl_mn1 * dy_1(m)
