+++ src/viscosity.f90
@@ -227,7 +227,7 @@ module Viscosity
 !
       if (any(ivisc=='nu-slope-limited')) then
         lslope_limit_diff = .true.
-        if (dimensionality<3)lisotropic_advection=.true.
+        if (dimensionality<3) lisotropic_advection=.true.
         if (isld_char == 0) then
           call farray_register_auxiliary('sld_char',isld_char,communicated=.true.)
           if (lroot) write(15,*) 'sld_char = fltarr(mx,my,mz)*one'
@@ -1725,7 +1725,7 @@ module Viscosity
         if (lpencil(i_visc_heat)) then  ! Heating term not implemented
           if (headtt) then
             call warning('calc_pencils_viscosity', 'viscous heating term '// &
-              'is not implemented for lvisc_hyper2_simplified')
+                         'is not implemented for lvisc_hyper2_simplified')
           endif
         endif
         if (lfirst .and. ldt) p%diffus_total2=p%diffus_total2+nu_hyper2
@@ -1738,7 +1738,7 @@ module Viscosity
         if (lpencil(i_visc_heat)) then  ! Heating term not implemented
           if (headtt) then
             call warning('calc_pencils_viscosity', 'viscous heating term '// &
-              'is not implemented for lvisc_hyper3_simplified')
+                         'is not implemented for lvisc_hyper3_simplified')
           endif
         endif
         if (lfirst .and. ldt) p%diffus_total3=p%diffus_total3+nu_hyper3
@@ -1752,13 +1752,13 @@ module Viscosity
           do i=1,3
             call der6(f,ju,tmp3,i,IGNOREDX=.true.)
             p%fvisc(:,j) = p%fvisc(:,j) + &
-                nu_hyper3*pi4_1*tmp3*dline_1(:,i)**2
+                           nu_hyper3*pi4_1*tmp3*dline_1(:,i)**2
           enddo
         enddo
         if (lpencil(i_visc_heat)) then
           if (headtt) then
             call warning('calc_pencils_viscosity', 'viscous heating term '//&
-                 'is not implemented for lvisc_hyper3_polar')
+                         'is not implemented for lvisc_hyper3_polar')
           endif
         endif
         if (lfirst .and. ldt) &
@@ -1776,7 +1776,7 @@ module Viscosity
               p%fvisc(:,j) = p%fvisc(:,j) + nu_hyper3_mesh * tmp3 * dline_1(:,i)
             else
               p%fvisc(:,j) = p%fvisc(:,j) + &
-                   nu_hyper3_mesh*pi5_1/60.* tmp3 *dline_1(:,i)
+                             nu_hyper3_mesh*pi5_1/60.* tmp3 *dline_1(:,i)
             endif
           enddo
         enddo
@@ -1801,14 +1801,12 @@ module Viscosity
       if (lvisc_hyper3_mesh_residual) then
         do j=1,3
           ju=j+iuu-1
-          do i=1,3
-            if (ldynamical_diffusion) then
-              p%fvisc(:,j) = p%fvisc(:,j) + nu_hyper3_mesh * p%der6u_res(:,i,j) * dline_1(:,i)
-            else
-              p%fvisc(:,j) = p%fvisc(:,j) + &
-                   nu_hyper3_mesh*pi5_1/60.*p%der6u_res(:,i,j)*dline_1(:,i)
-            endif
-          enddo
+          if (ldynamical_diffusion) then
+            p%fvisc(:,j) = p%fvisc(:,j) + nu_hyper3_mesh * sum(p%der6u_res(:,:,j) * dline_1,2)
+          else
+            p%fvisc(:,j) = p%fvisc(:,j) + &
[…]
