+++ src/entropy.f90
@@ -4499,9 +4499,9 @@ module Energy
       df(l1:l2,m,n,iss) = df(l1:l2,m,n,iss) + thdiff
 !
       if (lfirst.and.ldt) diffus_chi3=diffus_chi3+ &
-           (chi_hyper3_aniso(1)*dx_1(l1:l2)**6 + &
-            chi_hyper3_aniso(2)*dy_1(  m  )**6 + &
-            chi_hyper3_aniso(3)*dz_1(  n  )**6)
+           (chi_hyper3_aniso(1)*dline_1(:,1)**6 + &
+            chi_hyper3_aniso(2)*dline_1(:,2)**6 + &
+            chi_hyper3_aniso(3)*dline_1(:,3)**6)
 !
     endsubroutine calc_heatcond_hyper3_aniso
 !***********************************************************************
@@ -4574,7 +4574,7 @@ module Energy
 !
       if (lfirst.and.ldt) then
         if (ldynamical_diffusion) then
-          diffus_chi3 = diffus_chi3 + chi_hyper3_mesh * (abs(dline_1(:,1)) + abs(dline_1(:,2)) + abs(dline_1(:,3)))
+          diffus_chi3 = diffus_chi3 + chi_hyper3_mesh * sum(abs(dline_1),2)
           advec_hypermesh_ss = 0.0
         else
           advec_hypermesh_ss=chi_hyper3_mesh*pi5_1*sqrt(dxyz_2)
