### 2020-09-19 14b6848
MR: base timestep control entirely on dline_1; removed/added  if (idiag_* where needed
