+++ src/dustvelocity.f90
@@ -1308,9 +1308,7 @@ module Dustvelocity
 !  ``uud/dx'' for timestep
 !
           if (lfirst .and. ldt) then
-            advec_uud=abs(p%uud(:,1,k))*dx_1(l1:l2)+ &
-                      abs(p%uud(:,2,k))*dy_1(  m  )+ &
-                      abs(p%uud(:,3,k))*dz_1(  n  )
+            advec_uud=sum(abs(p%uud(:,:,k))*dline_1,2)
             maxadvec=maxadvec+advec_uud
             if (idiag_dtud(k)/=0) &
                 call max_mn_name(advec_uud/cdt,idiag_dtud(k),l_dt=.true.)
@@ -1340,26 +1338,22 @@ module Dustvelocity
         do k=1,ndustspec
           if ((headtt.or.ldebug) .and. (ip<6)) &
               print*, 'duud_dt: Calculate diagnostic values...'
-          if (idiag_udrms(k)/=0) &
-              call sum_mn_name(p%ud2(:,k),idiag_udrms(k),lsqrt=.true.)
-          if (idiag_udmax(k)/=0) &
-              call max_mn_name(p%ud2(:,k),idiag_udmax(k),lsqrt=.true.)
-          if (idiag_rdudmax(k)/=0) &
-              call max_mn_name(p%rhod(:,k)**2*p%ud2(:,k),idiag_rdudmax(k), &
-              lsqrt=.true.)
-          if (idiag_ud2m(k)/=0) call sum_mn_name(p%ud2(:,k),idiag_ud2m(k))
-          if (idiag_ekintot_dust/=0) &
-             call integrate_mn_name(p%ud2,idiag_ekintot_dust)
-          if (idiag_udxm(k)/=0) call sum_mn_name(p%uud(:,1,k),idiag_udxm(k))
-          if (idiag_udym(k)/=0) call sum_mn_name(p%uud(:,2,k),idiag_udym(k))
-          if (idiag_udzm(k)/=0) call sum_mn_name(p%uud(:,3,k),idiag_udzm(k))
+          call sum_mn_name(p%ud2(:,k),idiag_udrms(k),lsqrt=.true.)
+          call max_mn_name(p%ud2(:,k),idiag_udmax(k),lsqrt=.true.)
+          call max_mn_name(p%rhod(:,k)**2*p%ud2(:,k),idiag_rdudmax(k), &
+                           lsqrt=.true.)
+          call sum_mn_name(p%ud2(:,k),idiag_ud2m(k))
+          call integrate_mn_name(p%ud2,idiag_ekintot_dust)
+          call sum_mn_name(p%uud(:,1,k),idiag_udxm(k))
+          call sum_mn_name(p%uud(:,2,k),idiag_udym(k))
+          call sum_mn_name(p%uud(:,3,k),idiag_udzm(k))
           if (idiag_udx2m(k)/=0) &
               call sum_mn_name(p%uud(:,1,k)**2,idiag_udx2m(k))
           if (idiag_udy2m(k)/=0) &
               call sum_mn_name(p%uud(:,2,k)**2,idiag_udy2m(k))
           if (idiag_udz2m(k)/=0) &
               call sum_mn_name(p%uud(:,3,k)**2,idiag_udz2m(k))
-          if (idiag_udm2(k)/=0) call max_mn_name(p%ud2(:,k),idiag_udm2(k))
+          call max_mn_name(p%ud2(:,k),idiag_udm2(k))
           if (idiag_divud2m(k)/=0) &
               call sum_mn_name(p%divud(:,k)**2,idiag_divud2m(k))
[…]
