+++ src/thermal_energy.f90
@@ -567,8 +567,7 @@ module Energy
           df(l1:l2,m,n,ieth) = df(l1:l2,m,n,ieth) + &
               chi_hyper3_mesh * d6eth * dline_1(:,j)
         enddo
-        if (lfirst .and. ldt) diffus_chi3 = diffus_chi3 + chi_hyper3_mesh* &
-            (abs(dline_1(:,1))+abs(dline_1(:,2))+abs(dline_1(:,3)))
+        if (lfirst .and. ldt) diffus_chi3 = diffus_chi3 + chi_hyper3_mesh*sum(dline_1,2)                  
       endif
 !
 !  Radiative diffusion (Rosseland approximation) through thermal energy diffusion.

[…]
