### 2020-09-19 c65c045
MR: based timestep control entirely on dline_1
