### 2020-09-19 95a2938
MR: base timestep control entirely on dline_1
