### 2020-09-19 ca8dd54
MR: minor in comments; start.f90 avoid double update of ghosts
