### 2020-09-19 564448b
MR: based timestep control entirely on dline_1
