### 2020-09-19 6e4296e
MR: improved use of links to snapshots: link is safely removed after reading
