### 2020-09-19 9a23f39
MR: added missing /=0 in if (idiag*
