+++ src/neutralvelocity.f90
@@ -641,25 +641,25 @@ module NeutralVelocity
         call yzsum_mn_name_x(p%uun(:,1),idiag_unxmx)
         call yzsum_mn_name_x(p%uun(:,2),idiag_unymx)
         call yzsum_mn_name_x(p%uun(:,3),idiag_unzmx)
-        if (idiag_unx2mz) call xysum_mn_name_z(p%uun(:,1)**2,idiag_unx2mz)
-        if (idiag_uny2mz) call xysum_mn_name_z(p%uun(:,2)**2,idiag_uny2mz)
-        if (idiag_unz2mz) call xysum_mn_name_z(p%uun(:,3)**2,idiag_unz2mz)
-        if (idiag_unx2my) call xzsum_mn_name_y(p%uun(:,1)**2,idiag_unx2my)
-        if (idiag_uny2my) call xzsum_mn_name_y(p%uun(:,2)**2,idiag_uny2my)
-        if (idiag_unz2my) call xzsum_mn_name_y(p%uun(:,3)**2,idiag_unz2my)
-        if (idiag_unx2mx) call yzsum_mn_name_x(p%uun(:,1)**2,idiag_unx2mx)
-        if (idiag_uny2mx) call yzsum_mn_name_x(p%uun(:,2)**2,idiag_uny2mx)
-        if (idiag_unz2mx) call yzsum_mn_name_x(p%uun(:,3)**2,idiag_unz2mx)
-        if (idiag_unxunymz) call xysum_mn_name_z(p%uun(:,1)*p%uun(:,2),idiag_unxunymz)
-        if (idiag_unxunzmz) call xysum_mn_name_z(p%uun(:,1)*p%uun(:,3),idiag_unxunzmz)
-        if (idiag_unyunzmz) call xysum_mn_name_z(p%uun(:,2)*p%uun(:,3),idiag_unyunzmz)
-        if (idiag_rnunxunymz) call xysum_mn_name_z(p%rhon*p%uun(:,1)*p%uun(:,2),idiag_rnunxunymz)
-        if (idiag_unxunymy) call xzsum_mn_name_y(p%uun(:,1)*p%uun(:,2),idiag_unxunymy)
-        if (idiag_unxunzmy) call xzsum_mn_name_y(p%uun(:,1)*p%uun(:,3),idiag_unxunzmy)
-        if (idiag_unyunzmy) call xzsum_mn_name_y(p%uun(:,2)*p%uun(:,3),idiag_unyunzmy)
-        if (idiag_unxunymx) call yzsum_mn_name_x(p%uun(:,1)*p%uun(:,2),idiag_unxunymx)
-        if (idiag_unxunzmx) call yzsum_mn_name_x(p%uun(:,1)*p%uun(:,3),idiag_unxunzmx)
-        if (idiag_unyunzmx) call yzsum_mn_name_x(p%uun(:,2)*p%uun(:,3),idiag_unyunzmx)
+        if (idiag_unx2mz/=0) call xysum_mn_name_z(p%uun(:,1)**2,idiag_unx2mz)
+        if (idiag_uny2mz/=0) call xysum_mn_name_z(p%uun(:,2)**2,idiag_uny2mz)
+        if (idiag_unz2mz/=0) call xysum_mn_name_z(p%uun(:,3)**2,idiag_unz2mz)
+        if (idiag_unx2my/=0) call xzsum_mn_name_y(p%uun(:,1)**2,idiag_unx2my)
+        if (idiag_uny2my/=0) call xzsum_mn_name_y(p%uun(:,2)**2,idiag_uny2my)
[…]
