### 2020-09-21 8f8e524
MR: simplified grid-coarsing part
