### 2020-09-22 0ebdd21
Few changes to the ogrid: - option to reduce order for boundary closures at the cylinder - do not allow for explicit boundary for rho when chemistry is solved - interpolatio_method=2 is now 2nd order Lagrangian, not spline interpolation - tried to implement boundary conditions for species without SBP property
