+++ src/solid_cells_ogrid_sub.f90
@@ -555,7 +555,11 @@ public :: der_ogrid_SBP_experimental, der2_ogrid_SBP_experimental
           if(lfirst_proc_x) then
             if(SBP) then
               call der_ogrid_SBP(f(l1_ogrid:l1_ogrid+8,m_ogrid,n_ogrid,k),df(1:6))
-              i=6
+              if (.not. SBP_reduced_to2) then
+		i=6
+	      else
+		i=4
+	      endif
             elseif(BDRY5) then
               call der_ogrid_bdry5(f(l1_ogrid:l1_ogrid+5,m_ogrid,n_ogrid,k),df(1:3))
               i=3
@@ -632,7 +636,11 @@ public :: der_ogrid_SBP_experimental, der2_ogrid_SBP_experimental
           if(lfirst_proc_x) then
             if(SBP) then
               call der_ogrid_SBP(f(l1_ogrid:l1_ogrid+8,m_ogrid,n_ogrid),df(1:6))
-              i=6
+	      if (.not. SBP_reduced_to2) then
+		i=6
+	      else
+		i=4
+	      endif
             elseif(BDRY5) then
               call der_ogrid_bdry5(f(l1_ogrid:l1_ogrid+5,m_ogrid,n_ogrid),df(1:3))
               i=3
@@ -703,7 +711,11 @@ public :: der_ogrid_SBP_experimental, der2_ogrid_SBP_experimental
           if(lfirst_proc_x) then
             if(SBP) then
               call der2_ogrid_SBP(f(l1_ogrid:l1_ogrid+8,m_ogrid,n_ogrid,k),df2(1:6))
-              i=6
+	      if (.not. SBP_reduced_to2) then
+		i=6
+	      else
+		i=4
+	      endif
             elseif(BDRY5) then
               call der2_ogrid_bdry5(f(l1_ogrid:l1_ogrid+6,m_ogrid,n_ogrid,k),df2(1:3))
               !call der2_ogrid_bdry5_alt(f,df2(1:3),k)
@@ -839,7 +851,11 @@ public :: der_ogrid_SBP_experimental, der2_ogrid_SBP_experimental
           if (nxgrid_ogrid/=1.and.nygrid_ogrid/=1) then
             if(lfirst_proc_x) then
               if(SBP) then
-                ii=6
+	        if (.not. SBP_reduced_to2) then
+		  ii=6
+	        else
+		  ii=4
+	        endif
                 call der_ijm_ogrid_SBP(f,df(1:6),k)
               elseif(BDRY5) then
                 ii=3
@@ -906,7 +922,11 @@ public :: der_ogrid_SBP_experimental, der2_ogrid_SBP_experimental
           if (nzgrid_ogrid/=1.and.nxgrid_ogrid/=1) then
             if(lfirst_proc_x) then
               if(SBP) then
-                ii=6
+	        if (.not. SBP_reduced_to2) then
+		  ii=6
+	        else
+		  ii=4
+	        endif
                 call der_ijn_ogrid_SBP(f,df(1:6),k)
               elseif(BDRY5) then
                 ii=3
@@ -1004,7 +1024,11 @@ public :: der_ogrid_SBP_experimental, der2_ogrid_SBP_experimental
 !
         if(lfirst_proc_x) then
           if(SBP) then
-            df(1:6)=0.
+	    if (.not. SBP_reduced_to2) then
+              df(1:6)=0.
+	    else
+              df(1:4)=0.
+	    endif
           elseif(BDRY5) then
             df(1:3)=0.
           endif
[…]
