+++ src/solid_cells_ogrid.f90
@@ -45,7 +45,7 @@ module Solid_Cells
       initsolid_cells, init_uu, r_ogrid, lset_flow_dir,ampl_noise, &
       grid_func_ogrid, coeff_grid_o, xyz_star_ogrid, &
       lcheck_interpolation, lcheck_init_interpolation, SBP, BDRY5, &
-      interpolation_method, lock_dt, lexpl_rho, &
+      interpolation_method, lock_dt, lexpl_rho, SBP_reduced_to2, &
       lshift_origin_ogrid,lshift_origin_lower_ogrid, interpol_filter, &
       lrk_tvd, SBP_optimized, interp_shift, llin_BC, lnonegative_Yk, &
       particle_interpolate, lparticle_uradonly, lsinus_spec_distr, &
@@ -57,7 +57,7 @@ module Solid_Cells
 !  Read run.in file
   namelist /solid_cells_run_pars/ &
       flow_dir_set, lset_flow_dir, interpolation_method, lcheck_interpolation, &
-      SBP, BDRY5, lrk_tvd, SBP_optimized, lexpl_rho, &
+      SBP, BDRY5, lrk_tvd, SBP_optimized, SBP_reduced_to2, lexpl_rho, &
       particle_interpolate, lparticle_uradonly, lfilter_solution, lock_dt, af, &
       lspecial_rad_int, lfilter_rhoonly, lspecial_rad_int_mom, &
       solid_reactions_intro_time, &
@@ -310,8 +310,8 @@ module Solid_Cells
         if(lroot) print*, 'Cylinder boundary condition: Fifth order boundary closures'
         SBP=.false.
       elseif(SBP) then
-        if(lroot) print*, 'Cylinder boundary condition: Third order SBP boundary closures'
-        if(.not. SBP_optimized) then
+        if(.not. SBP_optimized .and. .not. SBP_reduced_to2) then
+          if(lroot) print*, 'Cylinder boundary condition: Third order SBP boundary closures'
           D1_SBP(1,:)=(/ -21600./13649.  , 104009./54596.  , 30443./81894.   , & 
                          -33311./27298.  , 16863./27298.   , -15025./163788. , &
                          0.              , 0.              , 0.              /)
@@ -348,6 +348,44 @@ module Solid_Cells
           D2_SBP(6,:)=(/ 21035./525612.  , -24641./131403. ,  30409./87602.  , & 
                          -54899./131403. ,  820271./525612., -117600./43801. , &
                          64800./43801.   , -6480./43801.   , 480./43801.     /)
+	elseif(SBP_reduced_to2) then
+          if(lroot) print*, 'Cylinder boundary condition: Second order SBP boundary closures'
+          D1_SBP(1,:)=(/-24./17.         , 59./34.         , -4./17.      , &
+                        -3./34.          , 0.              , 0.           , &
+                         0.              , 0.              , 0.              /)
+          D1_SBP(2,:)=(/-1./2.           , 0.              , 1./2.        , &
+                         0.              , 0.              , 0.           , &
+                         0.              , 0.              , 0.              /)
+          D1_SBP(3,:)=(/ 4./43.          , -59./86.        , 0.           , &
+                         59./86.         , -4./43.         , 0.           , &
+                         0.              , 0.              , 0.              /)
+          D1_SBP(4,:)=(/ 3./98.          , 0.              , -59./98.     , &
+                         0.              , 32./49.         , -4./49.      , &
+                         0.              , 0.              , 0.              /)
[…]
