+++ python/pencil/ism_dyn/derived_h5.py
@@ -28,16 +28,51 @@ def is_vector(key):
     """Check if the variable denoted by the label key is a vector.
     """
     vec = False
-    for tag in ('bb', 'jj', 'upot', 'urot', 'vort'):
+    for tag in ('aa', 'uu', 'bb', 'jj', 'upot', 'urot', 'vort'):
         if key in tag:
             vec = True
     return vec
 
-def derive_data(sim_path, src, dst, magic=['pp','tt'], par=[],
-                gd=[], overwrite=False, rank=0, size=1, nghost=3,
-                chunksize = 1000.0, dtype=np.float64, quiet=True  
+def der_limits(n1,n2,m1,m2,l1,l2,nghost):
+    if n1 >= nghost:
+        n1shift = n1 - nghost
+    else:
+        n1shift = n1
+    if m1 >= nghost:
+        m1shift = m1 - nghost
+    else:
+        m1shift = m1
+    if l1 >= nghost:
+        l1shift = l1 - nghost
+    else:
+        l1shift = l1
+    n2shift = n2 + nghost
+    m2shift = m2 + nghost
+    l2shift = l2 + nghost
+    return n1shift,n2shift,m1shift,m2shift,l1shift,l2shift
+
+def under_limits(n1,m1,l1,n1shift,m1shift,l1shift,nghost):
+    if n1shift == n1:
+        n1r = 0
+    else:
+        n1r = nghost
+    if m1shift == m1:
+        m1r = 0
+    else:
+        m1r = nghost
+    if l1shift == l1:
+        l1r = 0
+    else:
+        l1r = nghost
+    return n1r,m1r,l1r
+
+def derive_data(sim_path, src, dst, magic=['pp','tt'], par=[], comm=None,
+                gd=[], overwrite=False, rank=0, size=1, nghost=3,status='a',
+                chunksize = 1000.0, dtype=np.float64, quiet=True, nmin=32  
                ):
 
+    if comm:
+        overwrite = False 
     if isinstance(par, list):
         os.chdir(sim_path)
         par = read.param(quiet=True,conflicts_quiet=True)
@@ -54,104 +89,126 @@ def derive_data(sim_path, src, dst, magic=['pp','tt'], par=[],
     #split data into manageable memory chunks
     dstchunksize = 8*nx*ny*nz/1024*1024
     if dstchunksize > chunksize:
-        nchunks = cpu_optimal(nx,ny,nz,mvar=1,maux=0,MBmin=chunksize,nmin=8)[1]
-        print('nchunks {}'.format(nchunks)) 
+        nchunks = cpu_optimal(nx,ny,nz,quiet=quiet,
+                              mvar=src['settings/mvar'][0],
+                              maux=src['settings/maux'][0],
+                              MBmin=chunksize,nmin=nmin,size=size)[1]
+    else:
+        nchunks = [1,1,1]
+    print('nchunks {}'.format(nchunks)) 
+    allchunks = nchunks[0]*nchunks[1]*nchunks[2]
     # for mpi split chunks across processes
     if size > 1:
-        allchunks = np.arange(nchunks[0]*nchunks[1]*nchunks[2])
-        locchunks = np.array_split(allchunks,size)
-        ichunks = locchunks[rank]
+        #allchunks = np.arange(nchunks[0]*nchunks[1]*nchunks[2])
+        locindx = np.array_split(np.arange(nx)+nghost,nchunks[0]) 
+        locindy = np.array_split(np.arange(ny)+nghost,nchunks[1]) 
+        locindz = np.array_split(np.arange(nz)+nghost,nchunks[2])
+        #ichunks = np.array_split(allchunks,size)
[…]
