### 2020-09-23 138e161
Added functions for computing masks and stats for h5py with mpi
