+++ python/pencil/ism_dyn/get_stats.py
@@ -1,0 +1,204 @@
+# get_stats.py
+#
+# 05-may-20
+# Author: F. Gent (fred.gent.ncl@gmail.com).
+#
+""" Derive auxilliary data and other diagnostics from var.h5 file and 
+    save to new h5 file
+ 
+    uses:
+      compute 'data' arrays of size [nz,ny,nx] as required
+      store 'time' of snapshot
+      compute 'masks' for example by temperature phase
+      compute summary statistics 'stats' 
+      compute 'structure' functions as required
+"""
+import numpy as np
+import scipy
+from . import is_vector
+from scipy.interpolate import interp1d
+from ..math import dot, dot2, natural_sort, helmholtz_fft, cpu_optimal
+from ..math.derivatives import curl, div, curl2, grad
+from ..calc import fluid_reynolds, magnetic_reynolds
+from ..io import open_h5, group_h5, dataset_h5
+from fileinput import input
+from sys import stdout
+import subprocess as sub
+from .. import read 
+import os
+
+def derive_stats(sim_path, src, dst, stat_keys=['Rm', 'uu', 'Ms'], par=[],
+                 comm=None, overwrite=False, rank=0, size=1, nghost=3,
+                 status='a', chunksize = 1000.0, quiet=True, nmin=32,
+                 lmask=False, mask_key = 'hot'
+                ):
+
+    if comm:
+        overwrite = False
+    if isinstance(par, list):
+        os.chdir(sim_path)
+        par = read.param(quiet=True,conflicts_quiet=True)
+    #get data dimensions
+    nx, ny, nz = src['settings']['nx'][0],\
+                 src['settings']['ny'][0],\
+                 src['settings']['nz'][0]
+    mx, my, mz = src['settings']['mx'][0],\
+                 src['settings']['my'][0],\
+                 src['settings']['mz'][0]
+    #split data into manageable memory chunks
+    dstchunksize = 8*nx*ny*nz/1024*1024
+    if dstchunksize > chunksize:
+        nchunks = cpu_optimal(nx,ny,nz,quiet=quiet,
+                              mvar=src['settings/mvar'][0],
+                              maux=src['settings/maux'][0],
+                              MBmin=chunksize,nmin=nmin,size=size)[1]
+    else:
+        nchunks = [1,1,1]
+    print('nchunks {}'.format(nchunks)) 
+    allchunks = nchunks[0]*nchunks[1]*nchunks[2]
+    # for mpi split chunks across processes
+    if size > 1:
+        #allchunks = np.arange(nchunks[0]*nchunks[1]*nchunks[2])
+        locindx = np.array_split(np.arange(nx)+nghost,nchunks[0]) 
+        locindy = np.array_split(np.arange(ny)+nghost,nchunks[1]) 
+        locindz = np.array_split(np.arange(nz)+nghost,nchunks[2])
+        #ichunks = np.array_split(allchunks,size)
+        indx = [locindx[np.mod(rank,nchunks[0])]]
+        indy = [locindy[np.mod(rank,nchunks[1])]]
+        indz = [locindz[np.mod(rank,nchunks[2])]]
+        if rank >= nchunks[2]:
+            indy = [locindy[np.mod(rank+1,nchunks[1])]]
+        if rank >= nchunks[1]:
+            indx = [locindx[np.mod(rank+1,nchunks[0])]]
+            if rank >= nchunks[2]:
+                indx = [locindx[np.mod(rank+2,nchunks[0])]]
+    else:
+        locindx = np.array_split(np.arange(nx)+nghost,nchunks[0]) 
+        locindy = np.array_split(np.arange(ny)+nghost,nchunks[1]) 
[…]
