+++ python/pencil/ism_dyn/__init__.py
@@ -5,4 +5,6 @@
 ################
 
 from .derived_h5 import *
+from .get_masks import *
+from .get_stats import *
 
