+++ python/pencil/ism_dyn/get_masks.py
@@ -1,0 +1,170 @@
+# get_masks.py
+#
+# 05-may-20
+# Author: F. Gent (fred.gent.ncl@gmail.com).
+#
+""" Derive auxilliary data and other diagnostics from var.h5 file and 
+    save to new h5 file
+ 
+    uses:
+      compute 'data' arrays of size [nz,ny,nx] as required
+      store 'time' of snapshot
+      compute 'masks' for example by temperature phase
+      compute summary statistics 'stats' 
+      compute 'structure' functions as required
+"""
+import numpy as np
+from ..math import cpu_optimal
+from ..io import open_h5, group_h5, dataset_h5
+from .. import read 
+import os
+
+def thermal_decomposition(
+    ss, pars, unit_key='unit_entropy', ent_cut=[2.32e9,]):
+    """
+    call signature:
+
+    thermal_decomposition(ss, pars, unit='unit_entropy', ent_cut=[2.32e9,])
+    
+    Keyword arguments:
+        ss:       dataset used for masks, default 'ss', alternate e.g.'tt'
+        pars:     Param() object required for units rescaling
+        unit_key: label of physical units in pars to apply to code values
+        ent_cut:  list of boundary mask values, default see thesis
+                  http://hdl.handle.net/10443/1755 Figure 5.10
+                  may have multiple boundaries
+    """
+    temp_masks = list()
+    hh = np.ma.array(np.copy(ss))
+    for ent in ent_cut:
+        hcut = ent/pars.__getattribute__(unit_key)
+        hh[np.where(ss<hcut)] = np.ma.masked
+        temp_masks.append(hh.mask)
+    print('thermal_decomposition', temp_masks[0].shape,len(temp_masks))
+    return temp_masks
+
+def derive_masks(sim_path, src, dst, data_key='data/ss', par=[], comm=None,
+                overwrite=False, rank=0, size=1, nghost=3, status='a',
+                chunksize = 1000.0, quiet=True, nmin=32,
+                ent_cuts=[2.32e9,], mask_keys=['hot',], unit_key='unit_entropy'
+               ):
+    if comm:
+        overwrite = False
+    if isinstance(par, list):
+        os.chdir(sim_path)
+        par = read.param(quiet=True,conflicts_quiet=True)
+    #get data dimensions
+    nx, ny, nz = src['settings']['nx'][0],\
+                 src['settings']['ny'][0],\
+                 src['settings']['nz'][0]
+    mx, my, mz = src['settings']['mx'][0],\
+                 src['settings']['my'][0],\
+                 src['settings']['mz'][0]
+    #split data into manageable memory chunks
+    dstchunksize = 8*nx*ny*nz/1024*1024
+    if dstchunksize > chunksize:
+        nchunks = cpu_optimal(nx,ny,nz,quiet=quiet,
+                              mvar=src['settings/mvar'][0],
+                              maux=src['settings/maux'][0],
+                              MBmin=chunksize,nmin=nmin,size=size)[1]
+    else:
+        nchunks = [1,1,1]
+    print('nchunks {}'.format(nchunks)) 
+    allchunks = nchunks[0]*nchunks[1]*nchunks[2]
+    # for mpi split chunks across processes
+    if size > 1:
+        #allchunks = np.arange(nchunks[0]*nchunks[1]*nchunks[2])
+        locindx = np.array_split(np.arange(nx)+nghost,nchunks[0]) 
+        locindy = np.array_split(np.arange(ny)+nghost,nchunks[1]) 
+        locindz = np.array_split(np.arange(nz)+nghost,nchunks[2])
+        #ichunks = np.array_split(allchunks,size)
[…]
