### 2020-09-23 6bc3f12
Print statements optional
