+++ python/pencil/math/derivatives/der_6th_order_w_ghosts.py
@@ -33,6 +33,8 @@ def xder_6th(f, dx):
         dfdx[..., l1:l2] = dx2*(45.*(f[..., l1+1:l2+1] - f[..., l1-1:l2-1])
                                 -9.*(f[..., l1+2:l2+2] - f[..., l1-2:l2-2])
                                 +(f[..., l1+3:l2+3] - f[..., l1-3:l2-3]))
+        dfdx[...,:l1] = dfdx[...,l2-3:l2]
+        dfdx[...,l2:] = dfdx[...,l1:l1+3]
     else:
         dfdx = 0.
 
@@ -63,6 +65,8 @@ def yder_6th(f, dy):
         dfdy[..., m1:m2, :] = dy2*(45.*(f[..., m1+1:m2+1, :] - f[..., m1-1:m2-1, :])
                                    -9.*(f[..., m1+2:m2+2, :] - f[..., m1-2:m2-2, :])
                                    +(f[..., m1+3:m2+3, :] - f[..., m1-3:m2-3, :]))
+        dfdy[..., :m1, :] = dfdy[..., m2-3:m2, :]
+        dfdy[..., m2:, :] = dfdy[..., m1:m1+3, :]
     else:
         dfdy = 0.
 
@@ -101,6 +105,8 @@ def zder_6th(f, dz, run2D=False):
                                        -9.*(f[..., n1+2:n2+2, :]
                                             -f[..., n1-2:n2-2, :])
                                        +(f[..., n1+3:n2+3, :]-f[..., n1-3:n2-3, :]))
+            dfdz[..., :n1, :] = dfdz[..., n2-3:n2, :]
+            dfdz[..., n2:, :] = dfdz[..., n1:n1+3, :]
         else:
             # Case f[...,z,y,x].
             dfdz[..., n1:n2, :, :] = dz2*(45.*(f[..., n1+1:n2+1, :, :]
@@ -108,6 +114,8 @@ def zder_6th(f, dz, run2D=False):
                                           -9.*(f[..., n1+2:n2+2, :, :]
                                                -f[..., n1-2:n2-2, :, :])
                                           +(f[..., n1+3:n2+3, :, :] - f[..., n1-3:n2-3, :, :]))
+            dfdz[..., :n1, :, :] = dfdz[..., n2-3:n2, :, :]
+            dfdz[..., n2:, :, :] = dfdz[..., n1:n1+3, :, :]
     else:
         dfdz = 0
 
@@ -139,6 +147,8 @@ def xder2_6th(f, dx):
                                 +270.*(f[..., l1-1:l2-1]+f[..., l1+1:l2+1])
                                 -27.*(f[..., l1-2:l2-2]+f[..., l1+2:l2+2])
                                 +2.*(f[..., l1-3:l2-3]+f[..., l1+3:l2+3]))
+        dfdx[...,:l1] = dfdx[...,l2-3:l2]
+        dfdx[...,l2:] = dfdx[...,l1:l1+3]
     else:
         dfdx = 0.
 
@@ -170,6 +180,8 @@ def yder2_6th(f, dy):
                                    +270.*(f[..., m1-1:m2-1, :] + f[..., m1+1:m2+1, :])
                                    -27.*(f[..., m1-2:m2-2, :] + f[..., m1+2:m2+2, :])
                                    +2.*(f[..., m1-3:m2-3, :] + f[..., m1+3:m2+3, :]))
+        dfdy[..., :m1, :] = dfdy[..., m2-3:m2, :]
+        dfdy[..., m2:, :] = dfdy[..., m1:m1+3, :]
     else:
         dfdy = 0.
 
@@ -200,6 +212,8 @@ def zder2_6th(f, dz):
[…]
