### 2020-09-23 d3442ff
Revised primes to use mpi size for proc layout
