### 2020-09-24 b594518
Debug non-mpi data distribution
