+++ python/pencil/ism_dyn/get_masks.py
@@ -70,23 +70,17 @@ def derive_masks(sim_path, src, dst, data_key='data/ss', par=[], comm=None,
     else:
         nchunks = [1,1,1]
     print('nchunks {}'.format(nchunks)) 
-    allchunks = nchunks[0]*nchunks[1]*nchunks[2]
+    # for mpi split chunks across processes
     # for mpi split chunks across processes
     if size > 1:
-        #allchunks = np.arange(nchunks[0]*nchunks[1]*nchunks[2])
         locindx = np.array_split(np.arange(nx)+nghost,nchunks[0]) 
         locindy = np.array_split(np.arange(ny)+nghost,nchunks[1]) 
         locindz = np.array_split(np.arange(nz)+nghost,nchunks[2])
-        #ichunks = np.array_split(allchunks,size)
-        indx = [locindx[np.mod(rank,nchunks[0])]]
-        indy = [locindy[np.mod(rank,nchunks[1])]]
+        indx = [locindx[np.mod(rank+int(rank/nchunks[2])
+                                   +int(rank/nchunks[1]),nchunks[0])]]
+        indy = [locindy[np.mod(rank+int(rank/nchunks[2]),nchunks[1])]]
         indz = [locindz[np.mod(rank,nchunks[2])]]
-        if rank >= nchunks[2]:
-            indy = [locindy[np.mod(rank+1,nchunks[1])]]
-        if rank >= nchunks[1]:
-            indx = [locindx[np.mod(rank+1,nchunks[0])]]
-            if rank >= nchunks[2]:
-                indx = [locindx[np.mod(rank+2,nchunks[0])]]
+        allchunks = 1
     else:
         locindx = np.array_split(np.arange(nx)+nghost,nchunks[0]) 
         locindy = np.array_split(np.arange(ny)+nghost,nchunks[1]) 
@@ -94,6 +88,7 @@ def derive_masks(sim_path, src, dst, data_key='data/ss', par=[], comm=None,
         indx = np.array_split(np.arange(nx)+nghost,nchunks[0]) 
         indy = np.array_split(np.arange(ny)+nghost,nchunks[1]) 
         indz = np.array_split(np.arange(nz)+nghost,nchunks[2])
+        allchunks = nchunks[0]*nchunks[1]*nchunks[2]
     # ensure derived variables are in a list
     if isinstance(mask_keys, list):
         mask_keys = mask_keys
@@ -108,63 +103,66 @@ def derive_masks(sim_path, src, dst, data_key='data/ss', par=[], comm=None,
                       comm=comm, size=size, rank=rank,
                       overwrite=overwrite, dtype=np.bool_)
         print('writing '+key+' shape {}'.format([ne,mz,my,mx]))
-        for iz in indz:
-            n1, n2 = iz[ 0]-nghost,\
-                     iz[-1]+nghost+1
-            n1out = n1+nghost
-            n2out = n2-nghost
-            varn1 =  nghost
-            varn2 = -nghost
-            if iz[0] == locindz[0][0]:
-                n1out = 0
-                varn1 = 0
-            if iz[-1] == locindz[-1][-1]:
-                n2out = n2
-                varn2 = n2
-            for iy in indy:
-                m1, m2 = iy[ 0]-nghost,\
-                         iy[-1]+nghost+1
-                m1out = m1+nghost
-                m2out = m2-nghost
-                varm1 =  nghost
-                varm2 = -nghost
-                if iy[0] == locindy[0][0]:
-                    m1out = 0
-                    varm1 = 0
-                if iy[-1] == locindy[-1][-1]:
-                    m2out = m2
-                    varm2 = m2
-                for ix in indx:
-                    l1, l2 = ix[ 0]-nghost,\
-                             ix[-1]+nghost+1
-                    l1out = l1+nghost
[…]

[…]
