### 2020-09-25 cfa8b57
Overwrite stat data True
