### 2020-09-25 f870de6
Experimental implementation of hyperdiffusion for ss_run_aver.
