### 2020-09-25 8bb0c7e
Compute and print hcond0_kramers from the initial condition.
