### 2020-09-29 f889f41
MR: new diagnostics lnrhorms
