+++ src/density.f90
@@ -195,6 +195,7 @@ module Density
   integer :: idiag_lnrhomin=0   ! DIAG_DOC: $\min(\log\rho)$
   integer :: idiag_lnrhomax=0   ! DIAG_DOC: $\max(\log\rho)$
   integer :: idiag_rhorms=0     ! DIAG_DOC:
+  integer :: idiag_lnrhorms=0   ! DIAG_DOC:
   integer :: idiag_ugrhom=0     ! DIAG_DOC: $\left<\uv\cdot\nabla\varrho\right>$
   integer :: idiag_uglnrhom=0   ! DIAG_DOC: $\left<\uv\cdot\nabla\ln\varrho\right>$
   integer :: idiag_lnrhomphi=0  ! PHIAVG_DOC: $\left<\ln\varrho\right>_\varphi$
@@ -256,6 +257,7 @@ module Density
 !
   logical :: lupdate_mass_source
   real, dimension(nx) :: diffus_diffrho
+!$omp THREADPRIVATE(diffus_diffrho)
 
   contains
 !***********************************************************************
@@ -2051,7 +2053,7 @@ module Density
       if (idiag_rhom/=0 .or. idiag_rho2m/=0 .or. idiag_rhof2m/=0 .or. idiag_rhomy/=0 .or. &
            idiag_rhomx/=0 .or. idiag_rho2mx/=0 .or. idiag_rhomz/=0 .or. idiag_rho2mz/=0 .or. &
            idiag_rhomin/=0 .or.  idiag_rhomax/=0 .or. idiag_rhomxy/=0 .or. idiag_rhomxz/=0 .or. &
-           idiag_totmass/=0 .or. idiag_mass/=0 .or. idiag_drho2m/=0 .or. idiag_rhorms/=0 .or. &
+           idiag_totmass/=0 .or. idiag_mass/=0 .or. idiag_drho2m/=0 .or. idiag_rhorms/=0 .or. idiag_lnrhorms/=0 .or. &
            idiag_inertiaxx/=0 .or. idiag_inertiayy/=0 .or. idiag_inertiazz/=0 .or. &
            idiag_inertiaxx_car/=0 .or. idiag_inertiayy_car/=0 .or. idiag_inertiazz_car/=0 .or. &
            idiag_drhom/=0 .or. idiag_rhomxmask/=0 .or. idiag_sigma/=0 .or. idiag_rhomzmask/=0 .or. &
@@ -2879,6 +2881,7 @@ module Density
         if (idiag_rho2m/=0)    call sum_mn_name(p%rho**2,idiag_rho2m)
         if (idiag_rhof2m/=0.and.lrho_flucz_as_aux) call sum_mn_name(f(l1:l2,m,n,irho_flucz)**2,idiag_rhof2m)
         if (idiag_rhorms/=0)   call sum_mn_name(p%rho**2,idiag_rhorms,lsqrt=.true.)
+        if (idiag_lnrhorms/=0) call sum_mn_name(p%rho**2,idiag_lnrhorms,lsqrt=.true.)
         if (idiag_lnrho2m/=0)  call sum_mn_name(p%lnrho**2,idiag_lnrho2m)
         if (idiag_drho2m/=0)   call sum_mn_name((p%rho-rho0)**2,idiag_drho2m)
         if (idiag_drhom/=0)    call sum_mn_name(p%rho-rho0,idiag_drhom)
@@ -3412,7 +3415,7 @@ module Density
 !
       if (lreset) then
         idiag_rhom=0; idiag_rho2m=0; idiag_rhof2m=0; idiag_lnrho2m=0
-        idiag_drho2m=0; idiag_drhom=0; idiag_rhorms=0
+        idiag_drho2m=0; idiag_drhom=0; idiag_rhorms=0; idiag_lnrhorms=0
[…]
