### 2020-09-29 1714745
modified sample to use kx_aa=100 instead of 1
