+++ samples/GravitationalWaves/print.in
@@ -6,7 +6,9 @@ dt(e10.3)
 !urms(e11.4)
 !oum(e11.4)
 !brms(e11.4)
-!EEK(e11.4)
+drho2m(f11.7)
+rhof2m(f11.7)
+EEK(e11.4)
 EEM(e11.4)
 EEGW(e11.4)
 !o2m(e11.4)
