### 2020-09-29 e106345
MR: new diagnostics idiag_etaanisoBB; made initialize_magnetic RELOAD-proof
