### 2020-09-30 4f43963
Added random phase initial condition for muS.
