### 2020-09-30 d7245d1
Chemistry on the ogrid: impose grad(species)=0 condition at the cylinder surface when no surface reactions
