### 2020-09-30 65f3f3b
add point vortex initial condition, and the Li-Ingersoll (2020) initial condition for jupiter's polar vortices
