### 2020-10-02 ebbcc21
Added timescale/fraction to the removal of zaveraged mean-field
