### 2020-10-03 61f54fa
added lpenc_requested(i_glnrho)=.true. for nu-const
