### 2020-10-03 bc3dbd7
initializing Qrecv, Qrad, and Qsend to zero (was unnoticed when not initializing with NaNs)
