### 2020-10-04 28aee96
in powerscl, nk wasn't correct for xparallelization
