### 2020-10-05 3e74a0a
MR: || -> bitwise or |
