### 2020-10-05 efb6c0a
MR: dissolved ylm with entry ylm_other into three functions ylm_core, ylm, ylm_other     (flang doesn't like entries)
