### 2020-10-06 aa3425f
Added second affiliation and a reference to the special issue paper on HDF5 and solar-atmospheric boundary conditions.
