+++ doc/paper/paper.md
@@ -24,101 +24,101 @@ authors:
    affiliation: 4
    orcid: 0000-0002-5893-6165
  - name: Philippe A. Bourdin
-   affiliation: 5
+   affiliation: "5, 6"
    orcid: 0000-0002-6793-601X
  - name: Wolfgang Dobler
-   affiliation: 6
- - name: Wladimir Lyra
    affiliation: 7
+ - name: Wladimir Lyra
+   affiliation: 8
    orcid: 0000-0002-3768-7542
  - name: Matthias Rheinhardt
-   affiliation: 8
- - name: Sven Bingert
    affiliation: 9
+ - name: Sven Bingert
+   affiliation: 10
    orcid: 0000-0001-9547-1582
  - name: Nils Erland L. Haugen
-   affiliation: "10, 11, 1"
+   affiliation: "11, 12, 1"
    orcid: 0000-0002-9184-8722
  - name: Antony Mee
-   affiliation: 12
+   affiliation: 13
  - name: Frederick Gent
-   affiliation: "8, 13"
+   affiliation: "9, 14"
    orcid: 0000-0002-1331-2260
  - name: Natalia Babkovskaia
-   affiliation: 14
- - name: Chao-Chin Yang
    affiliation: 15
+ - name: Chao-Chin Yang
+   affiliation: 16
    orcid: 0000-0003-2589-5034
  - name: Tobias Heinemann
-   affiliation: 16
+   affiliation: 17
    orcid: 0000-0002-2476-9733
  - name: Boris Dintrans
-   affiliation: 17
+   affiliation: 18
  - name: Dhrubaditya Mitra
    affiliation: 1
    orcid: 0000-0003-4861-8152
  - name: Simon Candelaresi
-   affiliation: 18
+   affiliation: 19
    orcid: 0000-0002-7666-8504
  - name: Jörn Warnecke
-   affiliation: 19
+   affiliation: 20
    orcid: 0000-0002-9292-4600
  - name: Petri J. Käpylä
-   affiliation: 20
+   affiliation: 21
    orcid: 0000-0001-9619-0053
  - name: Andreas Schreiber
-   affiliation: 14
+   affiliation: 15
  - name: Piyali Chatterjee
-   affiliation: 21
+   affiliation: 22
    orcid: 0000-0002-0181-2495
  - name: Maarit J. Käpylä
-   affiliation: "8, 19"
+   affiliation: "9, 20"
    orcid: 0000-0002-9614-2200
  - name: Xiang-Yu Li
    affiliation: 1
    orcid: 0000-0002-5722-0018 
  - name: Jonas Krüger
-   affiliation: "10, 11"
+   affiliation: "11, 12"
    orcid: 0000-0001-8036-0695
  - name: Jørgen R. Aarnes
-   affiliation: 11
+   affiliation: 12
    orcid: 0000-0002-5899-2597
  - name: Graeme R. Sarson
-   affiliation: 13
+   affiliation: 14
    orcid: 0000-0001-6774-9372
  - name: Jeffrey S. Oishi
-   affiliation: 22
+   affiliation: 23
    orcid: 0000-0001-8531-6570
  - name: Jennifer Schober
-   affiliation: 23
+   affiliation: 24
    orcid: 0000-0001-7888-6671
  - name: Raphaël Plasson
-   affiliation: 24
+   affiliation: 25
    orcid: 0000-0003-2319-1463
  - name: Christer Sandin
    affiliation: 1
    orcid: 0000-0002-6370-5505
  - name: Ewa Karchniwy
-   affiliation: "11, 25"
+   affiliation: "12, 26"
    orcid: 0000-0001-6709-1160
  - name: Luiz Felippe S. Rodrigues
[…]
