### 2020-10-08 0887367
Revise functions notanumber().
Reference: https://stackoverflow.com/questions/17389958/is-there-a-standard-way-to-check-for-infinite-and-nan-in-fortran-90-95

NOTE: Since Fortran 2000, the standard intrinsic module IEEE_ARITHMETIC
    can be used to check for Infs and NaNs.
