### 2020-10-08 b5af6a8
removed debug output
