### 2020-10-10 858ccf4
SLD: add different treatment of non-linear diffusivity reduction following Matthias Rempels earlier work
