### 2020-10-11 69d805a
ourms diagnostic added (analogously to jbrms)
