### 2020-10-12 ec77c99
MR: generic copy_addr function
