### 2020-10-12 fd4b2e5
MR: copy_addr_c -> copy_addr
