+++ src/eos_temperature_ionization.f90
@@ -2364,10 +2364,12 @@ module EquationOfState
 !***********************************************************************
     subroutine pushpars2c(p_par)
 
+    use Syscalls, only: copy_addr
+
     integer, parameter :: n_pars=1
     integer(KIND=ikind8), dimension(n_pars) :: p_par
 
-    call copy_addr_c(cs20,p_par(1))
+    call copy_addr(cs20,p_par(1))
 
     endsubroutine pushpars2c
 !***********************************************************************
