+++ src/entropy.f90
@@ -8117,10 +8117,12 @@ module Energy
 !***********************************************************************
     subroutine pushpars2c(p_par)
 
+    use Syscalls, only: copy_addr
+
     integer, parameter :: n_pars=1
     integer(KIND=ikind8), dimension(n_pars) :: p_par
 
-    call copy_addr_c(chi,p_par(1))
+    call copy_addr(chi,p_par(1))
 
     endsubroutine pushpars2c
 !***********************************************************************
