+++ src/eos_idealgas.f90
@@ -4930,15 +4930,17 @@ module EquationOfState
 !***********************************************************************
     subroutine pushpars2c(p_par)
 !
+    use Syscalls, only: copy_addr
+
     integer, parameter :: n_pars=6
     integer(KIND=ikind8), dimension(n_pars) :: p_par
 !
-    call copy_addr_c(cs20,p_par(1))
-    call copy_addr_c(gamma,p_par(2))
-    call copy_addr_c(cv,p_par(3))
-    call copy_addr_c(cp,p_par(4))
-    call copy_addr_c(lnrho0,p_par(5))
-    call copy_addr_c(lnTT0,p_par(6))
+    call copy_addr(cs20,p_par(1))
+    call copy_addr(gamma,p_par(2))
+    call copy_addr(cv,p_par(3))
+    call copy_addr(cp,p_par(4))
+    call copy_addr(lnrho0,p_par(5))
+    call copy_addr(lnTT0,p_par(6))
 !
     endsubroutine pushpars2c
 !***********************************************************************
