### 2020-10-12 8c05a7f
MR: added missing copy_addr variant
