+++ src/special/lucky_droplet.f90
@@ -37,7 +37,8 @@ module Special
    integer, target :: ispecial=0
 !
   ! input parameters
-  real :: gam_lucky=fourthird, efficiency_exponent=0., efficiency_prefactor=1., tstar=0.
+  real :: gam_lucky=fourthird, efficiency_exponent=0., efficiency_prefactor=1.
+  real :: rcrit=0., lambda_star1=1, runit=1.
   character(len=50) :: init_qq='zero'
   namelist /special_init_pars/ &
     gam_lucky
@@ -45,10 +46,13 @@ module Special
   ! run parameters
   logical :: lMFT=.false., lrA=.false., lrB=.false.
   namelist /special_run_pars/ &
-    gam_lucky, lMFT, lrA, lrB, efficiency_exponent, efficiency_prefactor, tstar
+    gam_lucky, lMFT, lrA, lrB, efficiency_exponent, efficiency_prefactor, &
+    rcrit, lambda_star1, runit
 !
 ! other variables (needs to be consistent with reset list below)
 !
+  integer :: idiag_rad =0 ! DIAG_DOC: $r/r_\ast$
+  integer :: idiag_tauk=0 ! DIAG_DOC: $\tau_k$
   integer :: idiag_tt1m=0 ! DIAG_DOC: $\langle T \rangle$
   integer :: idiag_qq1m=0 ! DIAG_DOC: $\langle \ln T \rangle$
   integer :: idiag_qq2m=0 ! DIAG_DOC: $\langle \ln T^2 \rangle$
@@ -182,7 +186,8 @@ module Special
       real, dimension (mx,my,mz,mfarray) :: f
       real, dimension (mx,my,mz,mvar) :: df
       real, dimension (nx) :: rrr, tauk, lamk
-      real, dimension (nx) :: tt, qq
+      real, dimension (nx) :: tt, qq, rad, ttauk
+      real :: r, r2, rrunit
       type (pencil_case) :: p
 !
       intent(in) :: f,p
@@ -204,15 +209,20 @@ module Special
         qq=alog(tt)
       endif
 !
+!  particle radius, and radius squared
+!
+      r=t**onethird
+      r2=r**2
+!
 !  Selection of different combinations of rA and rB
 !  Black line in paper corresponds to: lrA=T, lrB=T
 !
       if (lrA.and.lrB) then
-        lamk=(t**onethird+1.)**2*(t**twothird-1.)
+        lamk=(r+1.)**2*(r2-1.)
       elseif (lrA.and..not.lrB) then
-        lamk=(t**onethird+1.)**2*t**twothird
+        lamk=(r+1.)**2* r2
       elseif (.not.lrA.and.lrB) then
-        lamk=t**twothird*(t**twothird-1.)
+        lamk= r2      *(r2-1.)
       else
         lamk=t**gam_lucky
       endif
@@ -221,11 +231,14 @@ module Special
 !  for t>tstar with exponent efficiency_exponent.
 !  Allow also efficiency_prefactor, just to demonstrate that
 !  it has no effect on the final P(T) curve, as expected.
+!  Entering the inverse, lambda_star1, is also useful and more
+!  intuitive (lambda_star1=1123.18s for standard parameters)
 !
-      if (efficiency_exponent/=0. .and. t>tstar) then
-        lamk=lamk*(t/tstar)**efficiency_exponent
+      rrunit=r*runit
+      if (efficiency_exponent/=0. .and. rrunit>rcrit) then
+        lamk=lamk*(rrunit/rcrit)**efficiency_exponent
       endif
-      lamk=lamk*efficiency_prefactor
[…]
