### 2020-10-14 34a8e74
added oxurms and o2u2m diagnostics
