+++ src/hydro.f90
@@ -13,7 +13,7 @@
 ! MVAR CONTRIBUTION 3
 ! MAUX CONTRIBUTION 0
 !
-! PENCILS PROVIDED divu; oo(3); o2; ou; u2; uij(3,3); uu(3); curlo(3)
+! PENCILS PROVIDED divu; oo(3); o2; ou; oxu2; oxu(3); u2; uij(3,3); uu(3); curlo(3)
 ! PENCILS PROVIDED sij(3,3); sij2; uij5(3,3); ugu(3); ugu2; oij(3,3)
 ! PENCILS PROVIDED d2uidxj(3,3), uijk(3,3,3); ogu(3)
 ! PENCILS PROVIDED u3u21; u1u32; u2u13; del2u(3); del4u(3); del6u(3)
@@ -422,15 +422,16 @@ module Hydro
                                 ! DIAG_DOC:  \quad(time step relative to
                                 ! DIAG_DOC:   advective time step;
                                 ! DIAG_DOC:   see \S~\ref{time-step})
-  integer :: idiag_oum=0        ! DIAG_DOC: $\left<\boldsymbol{\omega}
-                                ! DIAG_DOC:   \cdot\uv\right>$
-  integer :: idiag_ourms=0      ! DIAG_DOC: $\left<(\boldsymbol{\omega}
-                                ! DIAG_DOC:   \cdot\uv^2\right>^{1/2}$
+  integer :: idiag_oum=0        ! DIAG_DOC: $\left<\boldsymbol{\omega}\cdot\uv\right>$
+  integer :: idiag_oxum=0       ! DIAG_DOC: $\left<\boldsymbol{\omega}\times\uv\right>$
+  integer :: idiag_ourms=0      ! DIAG_DOC: $\left<(\boldsymbol{\omega}\cdot\uv)^2\right>^{1/2}$
+  integer :: idiag_oxurms=0     ! DIAG_DOC: $\left<(\boldsymbol{\omega}\times\uv)^2\right>^{1/2}$
   integer :: idiag_ou_int=0     ! DIAG_DOC: $\int_V\boldsymbol{\omega}\cdot\uv\,dV$
   integer :: idiag_fum=0        ! DIAG_DOC: $\left<\fv\cdot\uv\right>$
   integer :: idiag_odel2um=0    ! DIAG_DOC: $\left<\boldsymbol{\omega}\nabla^2\uv\right>$
   integer :: idiag_o2m=0        ! DIAG_DOC: $\left<\boldsymbol{\omega}^2\right>
                                 ! DIAG_DOC:   \equiv \left<(\curl\uv)^2\right>$
+  integer :: idiag_o2u2m=0      ! DIAG_DOC: $\left<\uv^2\boldsymbol{\omega}^2\right>$
   integer :: idiag_o2sphm=0     ! DIAG_DOC: $\int_{r=0}^{r=1} \boldsymbol{\omega}^2 dV$,
                                 ! DIAG_DOC:   where $r=\sqrt{x^2+y^2+z^2}$
   integer :: idiag_orms=0       ! DIAG_DOC: $\left<\boldsymbol{\omega}^2\right>^{1/2}$
@@ -1503,9 +1504,9 @@ module Hydro
         case ('xjump')
           call jump(f,iux,uu_left,uu_right,widthuu,'x')
           call jump(f,iuy,uy_left,uy_right,widthuu,'x')
-        case ('Beltrami-x'); call beltrami(ampluu(j),f,iuu,kx=kx_uu)
-        case ('Beltrami-y'); call beltrami(ampluu(j),f,iuu,ky=ky_uu)
-        case ('Beltrami-z'); call beltrami(ampluu(j),f,iuu,kz=kz_uu)
[…]
