### 2020-10-14 40ed18b
Added initial condition of a solar atmosphere with magnetic field.
