+++ src/initial_condition/solar_atmosphere_Omega_bfield.f90
@@ -1,0 +1,314 @@
+! $Id$
+!
+! Initial condition for a stably tratified solar atmosphere
+! from the convection zone to the corona, with a Omega shaped
+! magnetic field loop.
+!
+!** AUTOMATIC CPARAM.INC GENERATION ****************************
+! Declare (for generation of cparam.inc) the number of f array
+! variables and auxiliary variables added by this module
+!
+! CPARAM logical, parameter :: linitial_condition = .true.
+!
+!***************************************************************
+
+module InitialCondition
+!
+  use Cparam
+  use Cdata
+  use General, only: keep_compiler_quiet
+  use Messages
+
+!
+  implicit none
+!
+  include '../initial_condition.h'
+!
+  ! g = surface gravity
+  ! mu = mean atomic weight
+  ! z_ss = photosphereic height
+  ! z_tr = start of the transition region
+  ! z_cr = start of the corona
+  ! p_ss = pressure ont he solar surface
+  ! rho_ss = density on the solar surface
+  ! T_ss = temperature on the solar surface
+  ! T_cr = temperature in the corona
+  ! H_ss = photosphere scale height 
+  ! H_cr = corona scale height
+  ! xi = mu*g/R
+
+  real :: z_ss = 0, z_tr = 10, z_cr = 20
+  real :: p_ss = 1, rho_ss = 1, T_ss = 1, T_cr = 150
+  real :: H_ss = 1, H_cr = 150
+  real :: xi = 1
+
+  namelist /initial_condition_pars/ &
+          z_ss, z_tr, z_cr, &
+          p_ss, rho_ss, T_ss, T_cr, &
+          H_ss, H_cr, &
+          xi
+!
+  contains
+!***********************************************************************
+    subroutine register_initial_condition()
+!
+!  Register variables associated with this module; likely none.
+!
+!  20-may-11/simon: coded
+!
+      if (lroot) call svn_id( &
+           "$Id$")
+!
+    endsubroutine register_initial_condition
+!***********************************************************************
+    subroutine initialize_initial_condition(f)
+!
+!  Initialize any module variables which are parameter dependent.
+!
+!  07-may-09/wlad: coded
+!
+      real, dimension (mx,my,mz,mfarray) :: f
+!
+      call keep_compiler_quiet(f)
+!
+    endsubroutine initialize_initial_condition
+!***********************************************************************
+    subroutine initial_condition_uu(f)
+!
+!  Initialize the velocity field.
+!
+!  07-may-09/wlad: coded
+!
+      real, dimension (mx,my,mz,mfarray), intent(inout) :: f
+!
+!  SAMPLE IMPLEMENTATION
+!
+      call keep_compiler_quiet(f)
+!
+    endsubroutine initial_condition_uu
+!***********************************************************************
+    subroutine initial_condition_lnrho(f)
+!
+!  Initialize logarithmic density. init_lnrho 
+!  will take care of converting it to linear 
+!  density if you use ldensity_nolog
+!
+!  07-sep-11/simon: coded
+!
+      use SharedVariables
+      use EquationOfState
+      
+      real, dimension (mx,my,mz,mfarray), intent(inout) :: f
+      integer :: l, m, n
+!
+    do n = n1, n2, 1
+      do m = m1, m2, 1
+        do l = l1, l2, 1
+          if (z(n) < z_ss) then
[…]
