### 2020-10-14 bedf704
Corrected factor 2 in vortex braid initial condition.
