### 2020-10-14 659d113
Added a subroutine ``corrfunc_3d(f)'' to compute autocorrelations in Fourier space, e.g., <u_i^* u_j>, which has 3*3 components and full 3d dependence on the vector k. Specify corr_3d=T in run.in to use.
