+++ src/power_spectrum.f90
@@ -3728,4 +3728,196 @@ endsubroutine pdf
   !
   endsubroutine anisoq_diag
 !***********************************************************************
+  subroutine corrfunc_3d(f)
+!
+!  Calculate correlation functions <u_i^* u_j>, for i,j=1,2,3
+!  as functions of the vector k.
+!  I call the symmetric, real part of it vij,
+!  and (the antisymmetric, imaginary part / i) = wij.
+!  For the moment only u=velocity field
+!
+!  2020-Oct-14 added this subroutine
+    use Fourier, only: fft_xyz_parallel
+    use Mpicomm, only: mpireduce_sum
+    use Sub, only: del2vi_etc, del2v_etc, cross, grad, curli, curl, dot2
+    use Chiral, only: iXX_chiral, iYY_chiral
+!
+  integer, parameter :: nk=nxgrid/2
+  integer :: i, k, ikx, iky, ikz, jkx,jky,jkz, im, in, ivec, ivec_jj
+  real, dimension (mx,my,mz,mfarray) :: f
+  real, dimension(nx,ny,nz) :: ux_re, ux_im
+  real, dimension(nx,ny,nz) :: uy_re, uy_im
+  real, dimension(nx,ny,nz) :: uz_re, uz_im
+  real, dimension(nxgrid,nygrid,nzgrid) :: vxx, vxy, vyy, vxz, vyz, vzz
+  real, dimension(nxgrid,nygrid,nzgrid) :: wxy, wyz, wzx
+  real, dimension(nxgrid,nygrid,nzgrid) :: vxx_sum, vxy_sum, vyy_sum
+  real, dimension(nxgrid,nygrid,nzgrid) :: vxz_sum, vyz_sum, vzz_sum
+  real, dimension(nxgrid,nygrid,nzgrid) :: wxy_sum, wyz_sum, wzx_sum
+  real, dimension(nk) :: nks=0.,nks_sum=0.
+  real, dimension(nk) :: k2m=0.,k2m_sum=0.,krms
+  real, dimension(nxgrid) :: kx
+  real, dimension(nygrid) :: ky
+  real, dimension(nzgrid) :: kz
+  character (len=3) :: sp
+  logical, save :: lwrite_krms=.true., lwrite_krms_GWs=.false.
+!
+!  passive scalar contributions (hardwired for now)
+!
+  real, dimension(nx,3) :: gtmp1,gtmp2
+!
+!  identify version
+!
+  if (lroot .AND. ip<10) call svn_id("$Id$")
+  !
+  !  Define wave vector, defined here for the *full* mesh.
+  !  Each processor will see only part of it.
+  !  Ignore *2*pi/Lx factor, because later we want k to be integers
+  !
+  kx=cshift((/(i-(nxgrid+1)/2,i=0,nxgrid-1)/),+(nxgrid+1)/2) !*2*pi/Lx
+  ky=cshift((/(i-(nygrid+1)/2,i=0,nygrid-1)/),+(nygrid+1)/2) !*2*pi/Ly
+  kz=cshift((/(i-(nzgrid+1)/2,i=0,nzgrid-1)/),+(nzgrid+1)/2) !*2*pi/Lz
+  !
+  !  initialize power spectrum to zero
+  !
+  k2m=0.
+  nks=0.
+  vxx=0.
+  vxx_sum=0.
+  vxy=0.
+  vxy_sum=0.
+  vyy=0.
+  vyy_sum=0.
+  vxz=0.
+  vxz_sum=0.
+  vyz=0.
[…]
