### 2020-10-14 22bcef1
Storms are now the same strength - cyclone and anticyclones are still mixed
