### 2020-10-14 2601a24
Added temporarily the output of coefficients in the zaveranges
