### 2020-10-15 4522471
Corrected the storm list that caused auto-test to fail.
