### 2020-10-15 64c741b
minor: added cylindrically averaged autocorrelation
