+++ src/power_spectrum.f90
@@ -3754,6 +3754,10 @@ endsubroutine pdf
   real, dimension(nxgrid,nygrid,nzgrid) :: vxx_sum, vxy_sum, vyy_sum
   real, dimension(nxgrid,nygrid,nzgrid) :: vxz_sum, vyz_sum, vzz_sum
   real, dimension(nxgrid,nygrid,nzgrid) :: wxy_sum, wyz_sum, wzx_sum
+  real :: k2
+  real, dimension(nk,nzgrid) :: cyl_vxx, cyl_vxx_sum
+  real, dimension(nk,nzgrid) :: cyl_vzz, cyl_vzz_sum
+  real, dimension(nk,nzgrid) :: cyl_wxy, cyl_wxy_sum
   real, dimension(nk) :: nks=0.,nks_sum=0.
   real, dimension(nk) :: k2m=0.,k2m_sum=0.,krms
   real, dimension(nxgrid) :: kx
@@ -3782,24 +3786,33 @@ endsubroutine pdf
   !
   k2m=0.
   nks=0.
-  vxx=0.
-  vxx_sum=0.
-  vxy=0.
-  vxy_sum=0.
-  vyy=0.
-  vyy_sum=0.
-  vxz=0.
-  vxz_sum=0.
-  vyz=0.
-  vyz_sum=0.
-  vzz=0.
-  vzz_sum=0.
-  wxy=0.
-  wxy_sum=0.
-  wyz=0.
-  wyz_sum=0.
-  wzx=0.
-  wzx_sum=0.
+  if (lcylindrical_spectra) then
+    cyl_vxx=0.
+    cyl_vxx_sum=0.
+    cyl_vzz=0.
+    cyl_vzz_sum=0.
+    cyl_wxy=0.
+    cyl_wxy_sum=0.
+  else
+    vxx=0.
+    vxx_sum=0.
+    vxy=0.
+    vxy_sum=0.
+    vyy=0.
+    vyy_sum=0.
+    vxz=0.
+    vxz_sum=0.
+    vyz=0.
+    vyz_sum=0.
+    vzz=0.
+    vzz_sum=0.
+    wxy=0.
+    wxy_sum=0.
+    wyz=0.
+    wyz_sum=0.
+    wzx=0.
+    wzx_sum=0.
+  endif
   !
   !  calculate each components
   !
@@ -3819,59 +3832,91 @@ endsubroutine pdf
   !
   !  calculate correlation functions
   !
-  if (lroot .AND. ip<10) print*,'fft done; now computing correlation functions...'
-  do ikz=1,nz
-    do iky=1,ny
-      do ikx=1,nx
-        jkx=nint(kx(ikx+ipx*nx))+nxgrid/2+1 ! runs from 1 to nxgrid
-        jky=nint(ky(iky+ipy*ny))+nygrid/2+1
-        jkz=nint(kz(ikz+ipz*nz))+nzgrid/2+1
-        vxx(jkx,jky,jkz)=vxx(jkx,jky,jkz)+ &
-          ux_re(ikx,iky,ikz)*ux_re(ikx,iky,ikz) &
-          +ux_im(ikx,iky,ikz)*ux_im(ikx,iky,ikz)
-        vxy(jkx,jky,jkz)=vxy(jkx,jky,jkz)+ &
-          ux_re(ikx,iky,ikz)*uy_re(ikx,iky,ikz) &
-          +ux_im(ikx,iky,ikz)*uy_im(ikx,iky,ikz)
-        vxz(jkx,jky,jkz)=vxz(jkx,jky,jkz)+ &
-          ux_re(ikx,iky,ikz)*uz_re(ikx,iky,ikz) &
[…]
