### 2020-10-15 7e42b6b
Fix for the storm list for the auto-test
