### 2020-10-20 098a75b
fixed pc_varcontent.pro and heatflux.f90 that pc_read_var works now for the heatflux module
