### 2020-10-20 d7e7640
added boris correction to characteristic speed.
