### 2020-10-21 60f1215
Add a height floor (similar to density floor).
