### 2020-10-21 6bb7c3e
Added the totKE and totAPE diagnostic variables to monitor the system energy.
