### 2020-10-22 f6ffb25
MR: outsourced doubled code into io_dist.h;     added a self-rescue mechanism in case of corrupted snapshot files:     if a proc*/var.dat cannot be read, reading from (max 26) neighboring processors is tried
