+++ src/io_dist.h
@@ -1,0 +1,245 @@
+!
+! Code fragment to be used in both read_snap_single and read_snap_double - type agnostic
+!
+      real :: t_test   ! t in single precision for backwards compatibility
+      logical :: ltest,lok,l0
+      integer :: len1d,ios,iosr,ii,iia,jj,kk,iip,ind,ind1
+      character(LEN=fnlen) :: readdir
+!
+      if (lserial_io) call start_serialize
+
+      readdir = directory_snap
+      lok=.false.; l0=.true.
+!
+kloop:do kk=max(0,ipz-1),min(ipz+1,nprocz-1)
+jloop:  do jj=max(0,ipy-1),min(ipy+1,nprocy-1)
+          iia=max(0,ipx-1)
+          if (l0) iia=iia-1             ! add one iteration for the original path
+iloop:    do ii=iia,min(nprocx-1,ipx+1)
+            if (.not.l0) then
+              iip=find_proc(ii,jj,kk)
+              if (iip/=iproc) then
+                ind=index(directory_snap,'proc')
+                ind1=index(directory_snap(ind:),'/')
+                if (ind1>0) then
+                  readdir=directory_snap(:ind+3)//trim(itoa(iip))//directory_snap(ind1+ind-1:)
+                else
+                  readdir=directory_snap(:ind+3)//trim(itoa(iip))
+                endif
+              else
+                cycle iloop
+              endif
+            endif
+            open (lun_input, FILE=trim(readdir)//'/'//trim(file), FORM='unformatted', status='old', iostat=ios)
+            if (ios==0) then
+
+              if (islink(trim(readdir)//'/'//trim(file))) &
+                  snap_is_link = islink(trim(readdir)//'/'//trim(file))
+!      if (ip<=8) print *, 'read_snap: open, mx,my,mz,nv=', mx, my, mz, nv
+              if (lwrite_2d) then
+                if (nx == 1) then
+                  if (ivar_omit1>0) allocate(tmp(1,my,mz,ivar_omit1:ivar_omit2))
+                  if (ivar_omit1==1) then
+                    read (lun_input,iostat=iosr) tmp,a(nghost+1,:,:,:)
+                  elseif (ivar_omit1>1) then
+                    read (lun_input,iostat=iosr) a(nghost+1,:,:,:ivar_omit1-1),tmp,a(nghost+1,:,:,ivar_omit1:)
+                  else
+                    read (lun_input,iostat=iosr) a(nghost+1,:,:,:)
+                  endif
+                elseif (ny == 1) then
+                  if (ivar_omit1>0) allocate(tmp(mx,1,mz,ivar_omit1:ivar_omit2))
+                  if (ivar_omit1==1) then
+                    read (lun_input,iostat=iosr) tmp,a(:,nghost+1,:,:)
+                  elseif (ivar_omit1>1) then
+                    read (lun_input,iostat=iosr) a(:,nghost+1,:,:ivar_omit1-1),tmp,a(:,nghost+1,:,ivar_omit1:)
+                  else
+                    read (lun_input,iostat=iosr) a(:,nghost+1,:,:)
+                  endif
+                elseif (nz == 1) then
+                  if (ivar_omit1>0) allocate(tmp(mx,my,1,ivar_omit1:ivar_omit2))
+                  if (ivar_omit1==1) then
+                    read (lun_input,iostat=iosr) tmp,a(:,:,nghost+1,:)
+                  elseif (ivar_omit1>1) then
+                    read (lun_input,iostat=iosr) a(:,:,nghost+1,:ivar_omit1-1),tmp,a(:,:,nghost+1,ivar_omit1:)
+                  else
+                    read (lun_input,iostat=iosr) a(:,:,nghost+1,:)
+                  endif
[…]
