+++ src/io_dist.f90
@@ -545,218 +545,30 @@ module Io
 !                                > 0, tstart specified: use this value
 !  13-feb-20/MR: added possibility to omit a range of variables when reading the snapshot,
 !                range is defined as [ivar_omi1,ivar_omi2]
+!  22-oct-20/MR: outsourced code shared with read_snap_double into io_dist.h
+!                Added "rescue mechanism": if snapshot is not properly readable,
+!                reding is tried from the (min. 11, max. 26) neighboring
+!                processors' snaphots.
 !
       use Mpicomm, only: start_serialize, end_serialize, mpibcast_real, mpiallreduce_or, &
                          stop_it, mpiallreduce_min, mpiallreduce_max, MPI_COMM_WORLD
       use Syscalls, only: islink
+      use General, only: itoa, find_proc
 !
       character (len=*), intent(in) :: file
       integer, intent(in) :: nv, mode
       real(KIND=rkind4), dimension (mx,my,mz,nv), intent(out) :: a
 !
-      real(KIND=rkind4) :: t_sp, t_sgl
+      real(KIND=rkind4) :: t_sp, t_red
 
       real(KIND=rkind4),                 intent(out) :: dx, dy, dz, deltay
       real(KIND=rkind4), dimension (mx), intent(out) :: x
       real(KIND=rkind4), dimension (my), intent(out) :: y
       real(KIND=rkind4), dimension (mz), intent(out) :: z
 
-      real :: t_test   ! t in single precision for backwards compatibility
-      logical :: ltest
       real(KIND=rkind4), dimension(:,:,:,:), allocatable :: tmp
-      integer :: len1d
 !
-      if (lserial_io) call start_serialize
-      open (lun_input, FILE=trim(directory_snap)//'/'//file, FORM='unformatted', status='old')
-      if (islink(trim(directory_snap)//'/'//trim(file))) &
-        snaplink=trim(directory_snap)//'/'//trim(file)
-!      if (ip<=8) print *, 'read_snap_single: open, mx,my,mz,nv=', mx, my, mz, nv
-      if (lwrite_2d) then
-!
-        if (nx == 1) then
-          if (ivar_omit1>0) allocate(tmp(1,my,mz,ivar_omit1:ivar_omit2))
-          if (ivar_omit1==1) then
-            read (lun_input) tmp,a(nghost+1,:,:,:)
-          elseif (ivar_omit1>1) then
-            read (lun_input) a(nghost+1,:,:,:ivar_omit1-1),tmp,a(nghost+1,:,:,ivar_omit1:)
-          else
-            read (lun_input) a(nghost+1,:,:,:)
-          endif
-        elseif (ny == 1) then
-          if (ivar_omit1>0) allocate(tmp(mx,1,mz,ivar_omit1:ivar_omit2))
-          if (ivar_omit1==1) then
-            read (lun_input) tmp,a(:,nghost+1,:,:)
-          elseif (ivar_omit1>1) then
-            read (lun_input) a(:,nghost+1,:,:ivar_omit1-1),tmp,a(:,nghost+1,:,ivar_omit1:)
-          else
-            read (lun_input) a(:,nghost+1,:,:)
-          endif
-        elseif (nz == 1) then
-          if (ivar_omit1>0) allocate(tmp(mx,my,1,ivar_omit1:ivar_omit2))
-          if (ivar_omit1==1) then
-            read (lun_input) tmp,a(:,:,nghost+1,:)
-          elseif (ivar_omit1>1) then
-            read (lun_input) a(:,:,nghost+1,:ivar_omit1-1),tmp,a(:,:,nghost+1,ivar_omit1:)
-          else
-            read (lun_input) a(:,:,nghost+1,:)
-          endif
-        else
[…]
