### 2020-10-24 fccb673
MR: removed outdated branch
