### 2020-10-24 6139c90
MR: changes to make index.pro work with pc_read_var.pro for testfields/flows etc.
