### 2020-10-27 cbdce6a
added omega_fourier to compute Fourier transform in uut and oot
