### 2020-10-27 81b0c68
SLD: prevent the large diffusivities for very small slopes
