### 2020-10-28 ba06d1d
Added option to remove the mean mu5 for the initial condition power_randomphase.
