### 2020-10-29 1c23699
Added a subroutine k_omega_spectra, which uses auxiliary variables uut and uust to compute spectra with both space and time Fourier transformed. To use it, call ou_omega=T, luut_as_aux=T, and also specify omega_fourier in run.in. Under development.
