+++ src/param_io.f90
@@ -185,7 +185,8 @@ module Param_IO
       ldirect_access, lproper_averages, lmaximal_cdt, lmaximal_cdtv, &
       pipe_func, glnCrossSec0, CrossSec_x1, CrossSec_x2, CrossSec_w, &
       cyinyang_intpol_type, yy_biquad_weights, lcutoff_corners, nycut, nzcut, rel_dang, &
-      lignore_nonequi, tag_foreign, tau_aver1, fmt_avgs
+      lignore_nonequi, tag_foreign, tau_aver1, fmt_avgs, &
+      ou_omega
 !
   namelist /IO_pars/ &
       lcollective_IO, IO_strategy
