+++ src/power_spectrum.f90
@@ -3491,7 +3491,6 @@ endsubroutine pdf
   !
   endsubroutine power_vec
 !***********************************************************************
-! hongzhe developing
   subroutine anisoq_diag(f)
 !
 !  Anisotropic alpha2 dynamo diagostics.
@@ -3725,7 +3724,7 @@ endsubroutine pdf
 !  For the moment only u=velocity field,
 !  but possible to extend using 'sp'
 !
-!  2020-Oct-14 Hongzhe added this subroutine
+!  2020-Oct-14/hongzhe added this subroutine
     use Fourier, only: fft_xyz_parallel
     use Mpicomm, only: mpireduce_sum
     use Sub, only: del2vi_etc, del2v_etc, cross, grad, curli, curl, dot2
@@ -3980,5 +3979,260 @@ endsubroutine pdf
   endif
     
   endsubroutine corrfunc_3d
+!***********************************************************************
+  subroutine k_omega_spectra(f,sp)
+!
+!  Calculate spectra with both (x,y,z) and t Fourier-transformed
+!  For the moment only works with kinetic helicity
+!  Use luut_as_aux=T and specify omega_fourier
+!  in run.in under &hydro_run_pars
+!
+!  29-oct-20/hongzhe: added this subroutine and coding
+!
+    use Fourier, only: fft_xyz_parallel
+    use Mpicomm, only: mpireduce_sum
+    use Sub, only: del2vi_etc, del2v_etc, cross, grad, curli, curl, dot2
+    use Chiral, only: iXX_chiral, iYY_chiral
+!
+  integer, parameter :: nk=nxgrid/2
+  integer :: i, k, ikx, iky, ikz, jkz, im, in, ivec, ivec_jj
+  real :: k2
+  real, dimension (mx,my,mz,mfarray) :: f
+  real, dimension(nx,ny,nz) :: a_re,a_im,b_re,b_im
+  real, dimension(nx) :: bbi, jji, b2, j2
+  real, dimension(nx,3) :: bb, bbEP, jj
+  real, dimension(nk) :: nks=0.,nks_sum=0.
+  real, dimension(nk) :: k2m=0.,k2m_sum=0.,krms
+  real, dimension(nk) :: spectrum,spectrum_sum
+  real, dimension(nk) :: spectrumhel,spectrumhel_sum
+  real, dimension(nk,nzgrid) :: cyl_spectrum, cyl_spectrum_sum
+  real, dimension(nk,nzgrid) :: cyl_spectrumhel, cyl_spectrumhel_sum
+  real, dimension(nxgrid) :: kx
+  real, dimension(nygrid) :: ky
+  real, dimension(nzgrid) :: kz
+  character (len=3) :: sp
+  logical, save :: lwrite_krms=.true., lwrite_krms_GWs=.false.
+!
+!  passive scalar contributions (hardwired for now)
+!
+  real, dimension(nx,3) :: gtmp1,gtmp2
+!
+!  identify version
+!
+  if (lroot .AND. ip<10) call svn_id("$Id$")
+  !
+  !  Define wave vector, defined here for the *full* mesh.
+  !  Each processor will see only part of it.
+  !  Ignore *2*pi/Lx factor, because later we want k to be integers
+  !
+  kx=cshift((/(i-(nxgrid+1)/2,i=0,nxgrid-1)/),+(nxgrid+1)/2) !*2*pi/Lx
+  ky=cshift((/(i-(nygrid+1)/2,i=0,nygrid-1)/),+(nygrid+1)/2) !*2*pi/Ly
+  kz=cshift((/(i-(nzgrid+1)/2,i=0,nzgrid-1)/),+(nzgrid+1)/2) !*2*pi/Lz
+  !
[…]
