### 2020-10-29 bec7279
Coded subroutine k_omega_spectra
