+++ src/power_spectrum.f90
@@ -4003,8 +4003,6 @@ endsubroutine pdf
   real, dimension(nx,3) :: bb, bbEP, jj
   real, dimension(nk) :: nks=0.,nks_sum=0.
   real, dimension(nk) :: k2m=0.,k2m_sum=0.,krms
-  real, dimension(nk) :: spectrum,spectrum_sum
-  real, dimension(nk) :: spectrumhel,spectrumhel_sum
   real, dimension(nk,nzgrid) :: cyl_spectrum, cyl_spectrum_sum
   real, dimension(nk,nzgrid) :: cyl_spectrumhel, cyl_spectrumhel_sum
   real, dimension(nxgrid) :: kx
@@ -4033,10 +4031,6 @@ endsubroutine pdf
   !
   k2m=0.
   nks=0.
-  spectrum=0.
-  spectrum_sum=0.
-  spectrumhel=0.
-  spectrumhel_sum=0.
   !
   if (lcylindrical_spectra) then
     cyl_spectrum=0.
@@ -4054,18 +4048,15 @@ endsubroutine pdf
     !  For "mag", calculate spectra of <bk^2> and <ak.bk>
     !
     if (sp=='kin') then
-      if (iuu==0) call fatal_error('powerhel','iuu=0')
-      do n=n1,n2
-        do m=m1,m2
-          call curli(f,iuu,bbi,ivec)
-          im=m-nghost
-          in=n-nghost
-          a_re(:,im,in)=bbi  !(this corresponds to vorticity)
-        enddo
-      enddo
-      b_re=f(l1:l2,m1:m2,n1:n2,iuu+ivec-1)  !(this corresponds to velocity)
-      a_im=0.
-      b_im=0.
+      if (iuu==0) call fatal_error('k_omega_spectra','iuu=0')
+      if (iuut==0) call fatal_error('k_omega_spectra','iuut=0')
+      if (iuust==0) call fatal_error('k_omega_spectra','iuust=0')
+      if (ioot==0) call fatal_error('k_omega_spectra','ioot=0')
+      if (ioost==0) call fatal_error('k_omega_spectra','ioost=0')
+      b_re=f(l1:l2,m1:m2,n1:n2,iuut+ivec-1)    ! the real part of u(\vec x,\omega)
+      b_im=f(l1:l2,m1:m2,n1:n2,iuust+ivec-1)   ! the imaginary part of u(\vec x,\omega)
+      a_re=f(l1:l2,m1:m2,n1:n2,ioot+ivec-1)    ! the real part of omega(\vec x,\omega)
+      a_im=f(l1:l2,m1:m2,n1:n2,ioost+ivec-1)   ! the imaginary part of omega(\vec x,\omega)
     endif
 !
 !  Doing the Fourier transform
@@ -4073,40 +4064,7 @@ endsubroutine pdf
     call fft_xyz_parallel(a_re,a_im)
     call fft_xyz_parallel(b_re,b_im)
 !
-!  integration over shells
-!
-    if (lroot .AND. ip<10) print*,'fft done; now integrate over shells...'
-    do ikz=1,nz
-      do iky=1,ny
-        do ikx=1,nx
-          k2=kx(ikx+ipx*nx)**2+ky(iky+ipy*ny)**2+kz(ikz+ipz*nz)**2
-          k=nint(sqrt(k2))
-          if (k>=0 .and. k<=(nk-1)) then
-!
-!  sum energy and helicity spectra
-!
-            spectrum(k+1)=spectrum(k+1) &
-               +b_re(ikx,iky,ikz)**2 &
-               +b_im(ikx,iky,ikz)**2
[…]
