+++ src/diagnostics.f90
@@ -801,6 +801,7 @@ module Diagnostics
       if (iuut/=0) f(:,:,:,iuxt:iuzt)=0.0
       if (iuust/=0) f(:,:,:,iuxst:iuzst)=0.0
       if (ioot/=0) f(:,:,:,ioxt:iozt)=0.0
+      if (ioost/=0) f(:,:,:,ioxst:iozst)=0.0
       if (ibbt/=0) f(:,:,:,ibxt:ibzt)=0.0
       if (ijjt/=0) f(:,:,:,ijxt:ijzt)=0.0
 !
