### 2020-10-30 dca4b15
small modifications in the python read/param and read/var files
