### 2020-10-30 23e379c
Added magnetic field to intial condition.
