+++ src/initial_condition/solar_atmosphere_Omega_bfield.f90
@@ -36,17 +36,24 @@ module InitialCondition
   ! H_ss = photosphere scale height 
   ! H_cr = corona scale height
   ! xi = mu*g/R
+  ! x0, z0 = center of the magnetic flux tube
+  ! B0 = magnetic field strength
+  ! d = diameter of the flux tube with Gaussian profile
+  ! alpha = twist parameter for the magnetic field
+  ! lam = exponent for Gaussian y-profile for rising flux tube
 
   real :: z_ss = 0, z_tr = 10, z_cr = 20
   real :: p_ss = 1, rho_ss = 1, T_ss = 1, T_cr = 150
   real :: H_ss = 1, H_cr = 150
   real :: xi = 1
+  real :: x_0 = 0, z_0 = 0, B0 = 1, d = 1, alpha = 0.5, lam = 1
 
   namelist /initial_condition_pars/ &
           z_ss, z_tr, z_cr, &
           p_ss, rho_ss, T_ss, T_cr, &
           H_ss, H_cr, &
-          xi
+          xi, &
+          x_0, z_0, B0, d, alpha, lam
 !
   contains
 !***********************************************************************
@@ -100,18 +107,31 @@ module InitialCondition
       
       real, dimension (mx,my,mz,mfarray), intent(inout) :: f
       integer :: l, m, n
+      real :: r, p_ex, p_b
 !
     do n = n1, n2, 1
       do m = m1, m2, 1
         do l = l1, l2, 1
           if (z(n) < z_ss) then
             f(l,m,n,ilnrho) = (1/(gamma-1))*log((1-z(n)*(gamma-1)/gamma))
+            f(l,m,n,ilnTT) = log(1 - z(n)*(gamma-1)/gamma)
           elseif (z(n) < z_tr) then
             f(l,m,n,ilnrho) = -z(n)
+            f(l,m,n,ilnTT) = 0
           else if (z(n) < z_cr) then
             f(l,m,n,ilnrho) = -z_tr - (z(n) - z_tr)/z_tr*log(T_cr) + z_tr/log(T_cr)*(1/T_cr**((z(n)-z_tr)/z_tr) - 1)
+            f(l,m,n,ilnTT) = (z(n)-z_tr)/z_tr * log(T_cr)
           else
             f(l,m,n,ilnrho) = -z_tr - log(T_cr) + z_tr/log(T_cr)*(1/T_cr-1) - (z(n)-z_cr)/T_cr
+            f(l,m,n,ilnTT) = log(T_cr)
+          end if
+          ! Introduce density deficit.
+          r = sqrt((x(l)-x_0)**2 + (z(n)-z_0)**2)
+          if (r < d) then
+            p_ex = B0**2 * exp(-r**2/d**2)*((alpha*d)**2 - 2 - 2*(alpha*r)**2)/4
+            ! 0.4 = (cv - cp); gamma = cp/cv; cv = 1
+            p_b = 0.4*exp(f(l,m,n,ilnrho) + f(l,m,n,ilnTT))
+            f(l,m,n,ilnrho) = log(exp(f(l,m,n,ilnrho)) - p_ex/p_b*exp(-y(m)**2/lam**2))
           end if
         enddo
       enddo
@@ -132,95 +152,39 @@ module InitialCondition
 
       real, dimension (mx,my,mz,mfarray) :: f      
       integer :: l, j, ju
-      real :: log_T_b, log_T_0
-      real :: j0, j1, R, R2d, theta, gg, gg_p, BB_r, BB_theta, BB_phi ! auxiliary variables
+      real :: r, theta
[…]
