### 2020-10-30 #2684 @arfon
:wave: @jmstone @bpbrown - would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The software under review is [_The Pencil Code, a modular MPI code for partial differential equations and particles: multipurpose and multiuser-maintained_](https://github.com/pencil-code/pencil-code) and the associated short paper [here](https://raw.githubusercontent.com/openjournals/joss-papers/joss.02684/joss.02684/10.21105.joss.02684.pdf).
