### 2020-10-31 #2684 @bpbrown
@arfon and JOSS: I would be very happy to help in this review.

I have two potential COI's however: I work very closely with co-author
Jeffrey Oishi as part of the Dedalus project (both as co-authors on papers
and as collaborators in funded research), and I am the supervising manager
who is hiring a second co-author (Illa Losada) to a position at CU
Boulder.  I am not involved in the Pencil code itself in any way
myself, and personally I think I can be an effective and impartial reviwer;
please let me know if you would like me to help in the review process.

On Fri, Oct 30, 2020 at 1:56 PM Arfon Smith <notifications@github.com>
wrote:

> Reopened #2684 <https://github.com/openjournals/joss-reviews/issues/2684>.
>
