+++ license/CODE_OF_CONDUCT.md
@@ -55,10 +55,9 @@ further defined and clarified by project maintainers.
 ## Enforcement
 
 Instances of abusive, harassing, or otherwise unacceptable behavior may be
-reported by contacting all 5 members of the Pencil Code Steering Committee 
-[https://www.nordita.org/~brandenb/pencil-code/PCSC/],
-<matthias.rheinhardt@aalto.fi> , <brandenb@nordita.org>
-<wdobler@gmail.com>, <wladimir.lyra@gmail.com>, <Nils.E.Haugen@sintef.no>.
+reported by contacting all 5 members of the [Pencil Code Steering Committee][PCSC]
+(<matthias.rheinhardt@aalto.fi> , <brandenb@nordita.org>, <wdobler@gmail.com>,
+<wladimir.lyra@gmail.com>, <Nils.E.Haugen@sintef.no>).
 All complaints will be reviewed and investigated and will result in a response that
 is deemed necessary and appropriate to the circumstances. The project team is
 obligated to maintain confidentiality with regard to the reporter of an incident.
@@ -75,3 +74,4 @@ available at [http://contributor-covenant.org/version/1/4][version]
 
 [homepage]: http://contributor-covenant.org
 [version]: http://contributor-covenant.org/version/1/4/
+[PCSC]: https://www.nordita.org/~brandenb/pencil-code/PCSC/
