### 2020-10-31 d8e389f
included fact1=.5*(1.+relhel**2) in ampl_ff(i)/sqrt(fact1*(ABC_A(i)**2+ABC_B(i)**2+ABC_C(i)**2)). This does not influence the default for which relhel=1.
