### 2020-10-31 #2684 @arfon
Thanks for the fast reply here and the COI disclosures @bpbrown. I would like to try and avoid this conflict if at all possible so I'll keep looking for a few more days to find an alternative second reviewer. If I fail, I may return to your offer - does that sound OK?
