### 2020-10-31 #2684 @jmstone
Dear Afron,

Thanks for the invitation and great to see the pencil code team is publishing in JOSS, but I am afraid I am not going to be able to provide a review at the moment.  You might also try my former graduate student Chris White (c-white) who has a lot of experience developing MHD codes.

Cheers
Jim

> On Oct 30, 2020, at 3:55 PM, Arfon Smith <notifications@github.com> wrote:
>
>
> 👋 @jmstone <https://github.com/jmstone> @bpbrown <https://github.com/bpbrown> - would either of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>
> The software under review is The Pencil Code, a modular MPI code for partial differential equations and particles: multipurpose and multiuser-maintained <https://github.com/pencil-code/pencil-code> and the associated short paper here <https://raw.githubusercontent.com/openjournals/joss-papers/joss.02684/joss.02684/10.21105.joss.02684.pdf>.
>
