### 2020-10-31 #2684 @rtfisher
Hi Arfon :

  I would be happy to do this, but I am fully committed for the next couple
of weeks. It would need to happen in the second half of the month.

  Best wishes,

  Bob

On Sat, Oct 31, 2020 at 10:08 AM Arfon Smith <notifications@github.com>
wrote:

> @rtfisher <https://github.com/rtfisher> @c-white
> <https://github.com/c-white> - would either of you be willing to review
> this submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
>
> The software under review is *The Pencil Code, a modular MPI code for
> partial differential equations and particles: multipurpose and
> multiuser-maintained* <https://github.com/pencil-code/pencil-code> and
> the associated short paper here
> <https://raw.githubusercontent.com/openjournals/joss-papers/joss.02684/joss.02684/10.21105.joss.02684.pdf>
> .
>
