### 2020-10-31 #2807 @arfon
@zingale, @rtfisher - This is the review thread for the paper. All of our communications will happen here from now on.

We aim for the review process to be _completed_ within about 4-6 weeks but please make a start well ahead of this as JOSS reviews are by their nature iterative and any early feedback you may be able to provide to the author will be very helpful in meeting this schedule.
