### 2020-10-31 #2807 @rtfisher
Thank you! I'll be in touch in a couple of weeks.

-B

On Sat, Oct 31, 2020 at 10:56 AM Arfon Smith <notifications@github.com>
wrote:

> @rtfisher <https://github.com/rtfisher> - @whedon
> <https://github.com/whedon> will automatically remind you in a week about
> starting your review, please ignore as I'm noting here that you're
> expecting to supply your review towards the end of November.
>
