### 2020-11-01 #2807 @AxelBrandenburg
Hi Whedon,
The supplied bibtex file contains all references that quote the Pencil
Code, but those with missing DOIs are not quoted in the paper, I believe.
Axel

PS: thanks, Arfon, for reviving the process!
