### 2020-11-02 dbefa1e
MR: testfield_z, testfield_general: write test variables indices in "indgen form" into index.pro;     testfield_nonlin_z: new diagnostics ub0, uj0; new set of linear testfields, by 'B11-B22-lin'; testfield_general: initial conditions of the form <type>[-<number>]                    no -<number>: ICs for all test variales the same, only in effect for 'gaussian-noise', 'zero'                                  otherwise ICs set individually                         number=0: IC for *last* test problem
