### 2020-11-03 1a1a5b4
MR: cparam: declared new datatype "16-byte real"     messages: safer handling of svnid.dat
