+++ license/developers.txt
@@ -22,7 +22,7 @@ Simon Candelaresi,    iomsn1/gmail.com,              iomsn,                +44 1
 Chi-kwan Chan,        ckch/nordita.org,              chanchikwan,          +46 8 5537 8573,   +46 7273 91600,           ,            Steward Observatory Arizona
 Piyali Chatterjee,    mppiyali/gmail.com,            mppiyali,             +47 228 57529,     +47 9845 2367,            mppiyali,    University of Oslo
 Fabio Del Sordo,      fabio/nordita.org,             fadiesis,             +46 8 5537 8277,   +46 700629700,            fabiodelsordo, NORDITA
-Boris Dintrans,       dintrans/cines.fr,    	     boris.dintrans,       +33 4 67141400,                 , bdintrans,   CNRS/CINES
+Boris Dintrans,       boris.dintrans/gmail.com,    	     boris.dintrans,       +33 4 67141400,                 , bdintrans,   CNRS/CINES
 Wolfgang Dobler,      wdobler/gmail.com,             wdobler,              +49 33201 43791,   +49 151 28123472,         wolfgangdobler,
 Thomas Gastine,       gastine/mps.mpg.de,            ,                     +49 5556 979 456,  ,                         ,            Max-Planck-Institut für Sonnensystemforschung Göttingen
 Fred A. Gent,         frederick.gent@aalto.fi,       fred.gent.ncl,        +358 50 344 9122,  +44 7779 364 893 (mob),   fred_gent,   Aalto University
