+++ license/authors.txt
@@ -84,7 +84,7 @@ atalayas = atalayas <atalayas@mycompany.com>
 atefeh.barekat = Atefeh Barekat <atefeh.barekat@gmail.com>
 bicanski = Andrej Bicanski <bicanski@mycompany.com>
 bingert = Sven Bingert <sven.bingert@gwdg.de>
-boris.dintrans@gmail.com = Boris Dintrans <dintrans@cines.fr>
+boris.dintrans@gmail.com = Boris Dintrans <boris.dintrans@gmail.com>
 brandenb = Axel Brandenburg <brandenb@nordita.org>
 brooke = John Brooke <brooke@mycompany.com>
 ccyang@astro.lu.se = Chao-Chin Yang <ccyang@unlv.edu>
@@ -97,8 +97,8 @@ daniel.o.gomez = Daniel O. Gomez <daniel.o.gomez@mycompany.com>
 dcarrera = Daniel Carrera <danielc@astro.lu.se>
 dhruba = Dhrubaditya Mitra <dhruba.mitra@gmail.com>
 dhruba.mitra = Dhrubaditya Mitra <dhruba.mitra@gmail.com>
-dintrans = Boris Dintrans <dintrans@cines.fr>
-bdintrans = Boris Dintrans <dintrans@cines.fr>
+dintrans = Boris Dintrans <boris.dintrans@gmail.com>
+bdintrans = Boris Dintrans <boris.dintrans@gmail.com>
 dobler = Wolfgang Dobler <wdobler@gmail.com>
 dorch = Bertil Dorch <dorch@mycompany.com>
 easy = Edward Spiegel <easy@mycompany.com>
