### 2020-11-03 bf443db
-        call finalize_aver(nprocxz,23,aatestmy) +        call finalize_aver(nprocxz,13,aatestmy) corrected
