### 2020-11-03 5802739
MR: more precise: committee where appropriate     -is deemed necessary and appropriate to the circumstances. The project team is     +is deemed necessary and appropriate to the circumstances. The committee is
    -faith may face temporary or permanent repercussions as determined by other
    -members of the project's leadership.
    +faith may face temporary or permanent repercussions as determined by 
    +the committee.

    A remark:
    "The committee is
     obligated to maintain confidentiality with regard to the reporter of an incident."
    This is not possible with respect to the accu
[…]
