+++ license/CODE_OF_CONDUCT.md
@@ -59,13 +59,13 @@ reported by contacting all 5 members of the [Pencil Code Steering Committee][PCS
 (<matthias.rheinhardt@aalto.fi> , <brandenb@nordita.org>, <wdobler@gmail.com>,
 <wladimir.lyra@gmail.com>, <Nils.E.Haugen@sintef.no>).
 All complaints will be reviewed and investigated and will result in a response that
-is deemed necessary and appropriate to the circumstances. The project team is
+is deemed necessary and appropriate to the circumstances. The committee is
 obligated to maintain confidentiality with regard to the reporter of an incident.
 Further details of specific enforcement policies may be posted separately.
 
 Project maintainers who do not follow or enforce the Code of Conduct in good
-faith may face temporary or permanent repercussions as determined by other
-members of the project's leadership.
+faith may face temporary or permanent repercussions as determined by 
+the committee.
 
 ## Attribution
 
