### 2020-11-03 0c3f8b8
adding a new flow equation
