### 2020-11-03 100e26a
corrected the python script to calculated transport coefficients
