+++ python/pencil/calc/aver2h5.py
@@ -13,6 +13,8 @@
 #
 #where:
 #
+# for 10 time steps between reseting, and do cut 30% first and 40% at the end (like idl routine)
+# choose: rmfzeros=3, rmbzeros=6
 #
 #  zav2h5:   function included in aver2h5.py
 #  folder:   location of the data you want to store
