### 2020-11-04 5ae66c1
fixed the sign for the b-tensor add more output
