+++ src/special/meanfield_e_tensor.f90
@@ -194,6 +194,11 @@ module Special
   integer :: idiag_emfxmxy=0, idiag_emfymxy=0, idiag_emfzmxy=0
   integer :: idiag_emfcoefxmxy=0, idiag_emfcoefymxy=0, idiag_emfcoefzmxy=0
   integer :: idiag_alphaxxmxy=0, idiag_alphayymxy=0, idiag_alphazzmxy=0
+  integer :: idiag_alphaxymxy=0, idiag_alphaxzmxy=0, idiag_alphayzmxy=0
+  integer :: idiag_betaxxmxy=0, idiag_betayymxy=0, idiag_betazzmxy=0
+  integer :: idiag_betaxymxy=0, idiag_betaxzmxy=0, idiag_betayzmxy=0
+  integer :: idiag_gammaxmxy=0, idiag_gammaymxy=0, idiag_gammazmxy=0
+  integer :: idiag_deltaxmxy=0, idiag_deltaymxy=0, idiag_deltazmxy=0
   integer :: idiag_umeanxmxy=0, idiag_umeanymxy=0, idiag_umeanzmxy=0
 !
   ! Interpolation parameters
@@ -851,21 +856,21 @@ module Special
           if (lreconstruct_tensors.and.lacoef.and.lbcoef) then                 ! beta, delta, kappa supposed to be correct
             if (lroot) print*, 'Reconstruct alpha and gamma from raw tensors.'
             do mm=1,ny
-              alpha_data(1,:,mm,1,1,1)=     acoef_data(1,:,mm,1,1,1)+bcoef_data(1,:,mm,1,1,2,2)/x(l1:l2)
-              alpha_data(1,:,mm,1,1,2)=0.5*(acoef_data(1,:,mm,1,1,2)-bcoef_data(1,:,mm,1,1,1,2)/x(l1:l2)  &
-                                           +acoef_data(1,:,mm,1,2,1)+bcoef_data(1,:,mm,1,2,2,2)/x(l1:l2))
-              alpha_data(1,:,mm,1,2,2)=     acoef_data(1,:,mm,1,2,2)-bcoef_data(1,:,mm,1,2,1,2)/x(l1:l2)
+              alpha_data(1,:,mm,1,1,1)=     acoef_data(1,:,mm,1,1,1)-bcoef_data(1,:,mm,1,1,2,2)/x(l1:l2)
+              alpha_data(1,:,mm,1,1,2)=0.5*(acoef_data(1,:,mm,1,1,2)+bcoef_data(1,:,mm,1,1,1,2)/x(l1:l2)  &
+                                           +acoef_data(1,:,mm,1,2,1)-bcoef_data(1,:,mm,1,2,2,2)/x(l1:l2))
+              alpha_data(1,:,mm,1,2,2)=     acoef_data(1,:,mm,1,2,2)+bcoef_data(1,:,mm,1,2,1,2)/x(l1:l2)
               alpha_data(1,:,mm,1,1,3)=0.5*(acoef_data(1,:,mm,1,1,3)+acoef_data(1,:,mm,1,3,1) & 
-                                           +bcoef_data(1,:,mm,1,3,2,2)/x(l1:l2))
+                                           -bcoef_data(1,:,mm,1,3,2,2)/x(l1:l2))
               alpha_data(1,:,mm,1,2,3)=0.5*(acoef_data(1,:,mm,1,2,3)+acoef_data(1,:,mm,1,3,2) &
-                                           -bcoef_data(1,:,mm,1,3,1,2)/x(l1:l2))
+                                           +bcoef_data(1,:,mm,1,3,1,2)/x(l1:l2))
             
               gamma_data(1,:,mm,1,1)=0.5*(acoef_data(1,:,mm,1,3,2)-acoef_data(1,:,mm,1,2,3) &
[…]
