### 2020-11-04 0e98b5e
Cosmetics for python module PencilCode.read.
Cut the line length of read.proc_var() down to 80.
