### 2020-11-04 01bb15f
Fix pc_collect.
If the Shear module is present, deltay is also needed to be written.
