+++ src/pre_and_post_processing/pc_collect.f90
@@ -385,7 +385,11 @@ subroutine read_and_combine(filename,f,mvar_in,lonly_farray)
     close (lun_output)
     open (lun_output, FILE=trim(directory_out)//'/'//filename, FORM='unformatted', position='append', status='old')
     t_sp = t
-    write (lun_output) t_sp, gx, gy, gz, dx, dy, dz
+    if (lshear) then
+      write (lun_output) t_sp, gx, gy, gz, dx, dy, dz, deltay
+    else
+      write (lun_output) t_sp, gx, gy, gz, dx, dy, dz
+    endif
   endif
 
   close (lun_output)
