### 2020-11-04 e8bf06f
Fix pc_collect.
The local grid does not need to be reconstructed when the IO strategy is
distributed.
