### 2020-11-04 febc947
Add function to Python module PencilCode.read.
Add read.allprocs_var() to read one snapshot under allprocs/.
