### 2020-11-04 4597094
Add keyword to function PencilCode.read.var().
Add keyword allprocs.  If allprocs is True, the snapshot is read from
under allprocs/.  Otherwise, the snapshot is read and combined from
individual process directories.
