### 2020-11-04 10f6361
Fix pc_collect.
When particles modules are on, they also need to be registered.
Otherwise, data/varname.dat will be overriden without all registered
auxiliary variables from particles.
