### 2020-11-04 8419b9e
Fix pc_collect.
Do not collect ghost cells if they are inside the active domain, because
auxiliary variables are not necessarily updating ghost cells.
