### 2020-11-04 f6c27a5
Follow commit e8bf06f4.
Commentted out seemingly unnecessary section, which actually overrides
good grid data with wrong one.
