+++ src/pre_and_post_processing/pc_collect.f90
@@ -340,39 +340,38 @@ subroutine read_and_combine(filename,f,mvar_in,lonly_farray)
 !
           if (ip<=6.and.lroot) print*, 'reading grid coordinates'
           call rgrid ('grid.dat')
+!03-nov-20/ccyang: Is the following section necessary?
+!                  It actually overrides good grid from grid.dat with wrong one.
+!!
+!! Size of box at local processor. The if-statement is for
+!! backward compatibility.
+!!
+!          if (all(lequidist)) then
+!            Lxyz_loc(1)=Lxyz(1)/nprocx
+!            Lxyz_loc(2)=Lxyz(2)/nprocy
+!            Lxyz_loc(3)=Lxyz(3)/nprocz
+!            xyz0_loc(1)=xyz0(1)+ipx*Lxyz_loc(1)
+!            xyz0_loc(2)=xyz0(2)+ipy*Lxyz_loc(2)
+!            xyz0_loc(3)=xyz0(3)+ipz*Lxyz_loc(3)
+!            xyz1_loc(1)=xyz0_loc(1)+Lxyz_loc(1)
+!            xyz1_loc(2)=xyz0_loc(2)+Lxyz_loc(2)
+!            xyz1_loc(3)=xyz0_loc(3)+Lxyz_loc(3)
+!          else
+!            xyz0_loc(1)=x(l1)
+!            xyz0_loc(2)=y(m1)
+!            xyz0_loc(3)=z(n1)
+!            xyz1_loc(1)=x(l2)
+!            xyz1_loc(2)=y(m2)
+!            xyz1_loc(3)=z(n2)
+!            Lxyz_loc(1)=xyz1_loc(1) - xyz0_loc(1)
+!            Lxyz_loc(2)=xyz1_loc(2) - xyz0_loc(3)
+!            Lxyz_loc(3)=xyz1_loc(3) - xyz0_loc(3)
+!          endif
+!!
+!!  Need to re-initialize the local grid for each processor.
+!!
+!          call construct_grid(x,y,z,dx,dy,dz)
 !
-          recgrid: if (IO_strategy /= "dist") then
-!
-! Size of box at local processor. The if-statement is for
-! backward compatibility.
-!
-            if (all(lequidist)) then
-              Lxyz_loc(1)=Lxyz(1)/nprocx
-              Lxyz_loc(2)=Lxyz(2)/nprocy
-              Lxyz_loc(3)=Lxyz(3)/nprocz
-              xyz0_loc(1)=xyz0(1)+ipx*Lxyz_loc(1)
-              xyz0_loc(2)=xyz0(2)+ipy*Lxyz_loc(2)
-              xyz0_loc(3)=xyz0(3)+ipz*Lxyz_loc(3)
-              xyz1_loc(1)=xyz0_loc(1)+Lxyz_loc(1)
-              xyz1_loc(2)=xyz0_loc(2)+Lxyz_loc(2)
-              xyz1_loc(3)=xyz0_loc(3)+Lxyz_loc(3)
-            else
-              xyz0_loc(1)=x(l1)
-              xyz0_loc(2)=y(m1)
-              xyz0_loc(3)=z(n1)
-              xyz1_loc(1)=x(l2)
-              xyz1_loc(2)=y(m2)
-              xyz1_loc(3)=z(n2)
-              Lxyz_loc(1)=xyz1_loc(1) - xyz0_loc(1)
-              Lxyz_loc(2)=xyz1_loc(2) - xyz0_loc(3)
-              Lxyz_loc(3)=xyz1_loc(3) - xyz0_loc(3)
-            endif
-!
-!  Need to re-initialize the local grid for each processor.
-!
[…]
