### 2020-11-04 82c4c39
Added sixth derivative der6_pencil to the 2nd order derivative method. Added a warning that it's still calculated in 6th order accuracy though.
Changed the fatal_error call to der6 by a warning.
