+++ src/deriv_2nd.f90
@@ -20,7 +20,7 @@ module Deriv
   public :: initialize_deriv, finalize_deriv
   public :: der, der2, der3, der4, der5, der6, derij, der5i1j,der3i3j,der3i2j1k,der4i1j1k
   public :: der4i2j, der2i2j2k
-  public :: der6_other, der_pencil, der2_pencil
+  public :: der6_other, der_pencil, der2_pencil, der6_pencil
   public :: deri_3d_inds
   public :: der_upwind1st
   public :: der_onesided_4_slice
@@ -732,11 +732,15 @@ module Deriv
       integer :: j,k
       logical, optional :: ignoredx,upwind
       logical :: igndx,upwnd
+      logical, save :: lfirstcall=.true.
 !
       intent(in)  :: f,k,j,ignoredx
       intent(out) :: df
 !
-      call fatal_error('deriv_2nd','der6 not implemented yet')
+      if (lfirstcall.and.headtt) then
+         call warning('deriv_2nd','der6 not implemented yet -- using 6th order')
+         lfirstcall=.false.
+      endif
 !
 !debug      if (loptimise_ders) der_call_count(k,icount_der6,j,1) = & !DERCOUNT
 !debug                          der_call_count(k,icount_der6,j,1) + 1 !DERCOUNT
@@ -755,7 +759,7 @@ module Deriv
         endif
         if ((.not.lcartesian_coords).and.(.not.igndx)) then
           call fatal_error('der6','in non-cartesian coordinates '//&
-               'just works if upwinding is used')
+               'only works if upwinding is used')
         endif
      endif
 !
@@ -854,7 +858,7 @@ module Deriv
              'non equidistant grid')
         if (.not.lcartesian_coords) &
              call fatal_error('der6_other','in non-cartesian coordinates '//&
-             'just works if upwiding is used')
+             'only works if upwiding is used')
       endif
 !
       if (j==1) then
@@ -908,6 +912,110 @@ module Deriv
       endif
 !
     endsubroutine der6_other
+!***********************************************************************
+    subroutine der6_pencil(j,pencil,df6,ignoredx,upwind)
+!
+!  Calculate 6th derivative of any x, y or z pencil.
+!
+!  20-jul-20/wlyra: adapted from der2_pencil
+!
+      real, dimension (:) :: pencil,df6
+      real, dimension (nx) :: facx
+      real, dimension (ny) :: facy
+      real, dimension (nz) :: facz
+      integer :: j
+      logical, optional :: ignoredx,upwind
+      logical :: igndx,upwnd
+      logical, save :: lfirstcall
+!
+      intent(in)  :: j, pencil,ignoredx,upwind
+      intent(out) :: df6
+!
+      if (lfirstcall.and.headtt) then
+         call warning('deriv_2nd','der6_pencil not implemented yet -- using 6th order')
+         lfirstcall=.false.
+      endif
+!
+      if (present(ignoredx)) then
+        igndx = ignoredx
+      else
+        igndx = .false.
+      endif
+!     
+      if (present(upwind)) then
+        if (.not. lequidist(j).and..not.lignore_nonequi) then
+          call fatal_error('der6','upwind cannot be used with '//&
+              'non-equidistant grid.')
+        endif
+        upwnd = upwind
+      else
+        upwnd = .false.
+      endif     
+!
+!  x-derivative
+!
+      if (j==1) then
+        if (size(pencil)/=mx) then
+          if (lroot) print*, 'der6_pencil: pencil must be of size mx for x derivative'
[…]
