### 2020-11-05 c5c130c
- Make the sign change of bcoef_mf at the right place - add comment to tensor.py
