### 2020-11-06 a0e2956
Add 2D histogram
