### 2020-11-06 0f39f36
Correct temperature calc
