### 2020-11-06 0d3c989
added cooling functions and sedov-taylor object
