+++ python/pencil/ism_dyn/ism_sedov_taylor.py
@@ -1,0 +1,227 @@
+# sedov_taylor.py
+#
+# 30-oct-20
+# Author: F. Gent (fred.gent.ncl@gmail.com).
+#
+import numpy as np
+import os
+from . import read
+
+def sedov_taylor(*args, **kwargs):
+    """
+    Compute analytic radial time evolution of SN blast waves for 
+    comparison with numerical results
+    
+    *t_sedov*:
+      Time_series object read from the simulation sn_series.dat
+    
+    *par*:
+      Param object containing the simulation parameters
+    
+    *time*:
+      list of time in code units
+    
+    *nt*:
+      Integer size of analytic arrays
+    
+    *endt*
+      Real end time in code units for the time series
+    
+    *dims*:
+      Dimension of the simulation default 3D
+    
+    *rho0*:
+      Ambient ISM density 
+    """
+    st_tmp = SedovTaylor()
+    st_tmp.get_st(*args, **kwargs)
+    return st_tmp
+
+class SedovTaylor(object):
+    """
+    SedovTaylor -- holds blast wave parameters and radial profiles.
+    """
+
+    def __init__(self):
+        """
+        Fill members with default values.
+        """
+
+        self.t = []
+        self.keys = []
+        
+        
+
+
+    def get_st(self,
+           t_sedov=0, par=list(), time=list(), nt=2000, startt=0., endt=0.005,
+           dims=3, quiet=True, rho0=1.6728e-24, M0=10, lsnowplough=True,
+           lcioffi=True
+        ):
+        """
+        Compute analytic radial time evolution of SN blast waves for 
+        comparison with numerical results
+        
+        *t_sedov*:
+          Time_series object read from the simulation sn_series.dat
+        
+        *par*:
+          Param object containing the simulation parameters
+        
+        *time*:
+          list of time in code units
+        
+        *nt*:
+          Integer size of analytic arrays
+        
+        *endt*
+          Real end time in code units for the time series
+        
+        *dims*:
+          Dimension of the simulation default 3D
+        
+        *rho0*:
+          Ambient ISM density 
+ 
+        *lsnowplough:
+          Include original snowplough profile
+
+        *lcioffi:
+          Include Cioffi et al profile
+        """
+        
+        if isinstance(par, list):
+            unit_length = 3.086e21 # 1pc in cm
+            unit_time = 3.1557e16 # Myr in s
+            unit_velocity = unit_length/unit_time # ~km/s in cm/s
+            unit_density = 1e-24 # g/cm3
+            unit_energy_density = unit_density*unit_velocity**2 # erg/cm3
+            unit_energy = unit_energy_density*unit_length**3 # erg
+            E0 = 10e51/unit_energy # erg
+            M0 = 10
+        else:
+            unit_length =         par.unit_length      
+            unit_time =           par.unit_time        
+            unit_velocity =       par.unit_velocity    
+            unit_density =        par.unit_density     
+            unit_energy_density = par.unit_energy_density
+            unit_energy =         par.unit_energy      
+            E0 = par.ampl_sn
+            if par.lsn_mass:
+                M0=par.mass_sn 
+            else:
+                M0 = 10
+        if len(time) > 0:
+            time = np.array(time)
+        else:
+            time = np.linspace(startt, endt, nt) + t_sedov
+        xi=2.026 #McKee/Ostriker 1988
[…]
