+++ python/pencil/ism_dyn/__init__.py
@@ -7,4 +7,5 @@
 from .derived_h5 import *
 from .get_masks import *
 from .get_stats import *
-
+from .ism_cooling import *
+from .ism_sedov_taylor import *
