### 2020-11-06 835171f
added gij calculation of gradient tensor (Cartesian)
