+++ src/param_io.f90
@@ -2,7 +2,6 @@
 !
 !  IO of init and run parameters. Subroutines here are `at the end of the
 !  food chain', i.e. depend on all physics modules plus possibly others.
-
 !  Using this module is also a compact way of referring to all physics
 !  modules at once.
 !
