### 2020-11-06 236348b
MR: enabled slices from hydro_kinematic
