+++ doc/citations/ref.bib
@@ -10376,10 +10376,13 @@ archivePrefix = {arXiv},
 @ARTICLE{2020arXiv200203638B,
        author = {{Brandenburg}, Axel},
         title = "{Piecewise quadratic growth during the 2019 novel coronavirus epidemic}",
-      journal = {$\!\!\!$},
+      journal = {Infectious Disease Modelling},
      keywords = {Quantitative Biology - Populations and Evolution},
          year = 2020,
         month = feb,
+       volume = {5},
+        pages = {681-690},
+          doi = {10.1016/j.idm.2020.08.014},
           eid = {arXiv:2002.03638},
         pages = {arXiv:2002.03638},
 archivePrefix = {arXiv},
@@ -10987,6 +10990,147 @@ archivePrefix = {arXiv},
        eprint = {1911.13093},
  primaryClass = {astro-ph.IM},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2020MNRAS.tmp.2507R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{2020PhRvD.102h3512R,
+       author = {{Roper Pol}, Alberto and {Mandal}, Sayan and {Brandenburg}, Axel and
+         {Kahniashvili}, Tina and {Kosowsky}, Arthur},
+        title = "{Numerical simulations of gravitational waves from early-universe turbulence}",
+      journal = {\prd},
+     keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics},
+         year = 2020,
+        month = oct,
+       volume = {102},
+       number = {8},
+          eid = {083512},
+        pages = {083512},
+          doi = {10.1103/PhysRevD.102.083512},
+archivePrefix = {arXiv},
+       eprint = {1903.08585},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020PhRvD.102h3512R},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@INPROCEEDINGS{2020DPS....5210301H,
+       author = {{Hyder}, A. and {Lyra}, W. and {Chanover}, N. and
+         {Morales-Juberias}, R. and {Jackiewicz}, J.},
+        title = "{Vortex Dynamics in the Polar Atmosphere of Jupiter}",
+    booktitle = {AAS/Division for Planetary Sciences Meeting Abstracts},
+         year = 2020,
+       series = {AAS/Division for Planetary Sciences Meeting Abstracts},
+       volume = {52},
+        month = oct,
+          eid = {103.01},
+        pages = {103.01},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020DPS....5210301H},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{2020arXiv201007046B,
+       author = {{Brandenburg}, Axel and {Das}, Upasana},
+        title = "{Turbulent radiative diffusion and turbulent Newtonian cooling}",
+      journal = {arXiv e-prints},
+     keywords = {Physics - Fluid Dynamics, Astrophysics - Solar and Stellar Astrophysics},
+         year = 2020,
+        month = oct,
+          eid = {arXiv:2010.07046},
+        pages = {arXiv:2010.07046},
+archivePrefix = {arXiv},
+       eprint = {2010.07046},
+ primaryClass = {physics.flu-dyn},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv201007046B},
[…]
