+++ doc/citations/notes.tex
@@ -96,6 +96,10 @@ In addition, 80 papers reference it for code comparison or other purposes %(2020
 \begin{description}
 \item %YEAR
 41 times in 2020 \citep{
+2020DPS....5210301H,%Hyder+ "Vortex Dynamics in the Polar Atmosphere of Jupiter"
+2020arXiv201007046B,%Brandenburg+Das "Turbulent radiative diffusion and turbulent Newtonian cooling"
+2020arXiv201001833G,%Gent+ "Small-Scale Dynamo in Supernova-Driven Interstellar Turbulence"
+2020arXiv200908231B,%Brandenburg+ "{The Pencil Code, a modular MPI code for partial differential ..."
 2020ApJ...898..112P,%Park "Negative Magnetic Diffusivity {\ensuremath{\beta}} Replacing the {\ensuremath{\alpha}} Effect in the Helical Dynamo"
 2020arXiv200801119Z,%Zhu+Yang "Streaming Instability with Multiple Dust Species: I. Favourable Conditions for the Linear Growth"
 2020arXiv200710696K,%Klahr+Schreiber "Turbulence sets the length scale for planetesimal formation: Local 2D simulations of ..."
@@ -107,7 +111,7 @@ In addition, 80 papers reference it for code comparison or other purposes %(2020
 Zhang_etal_2020comb,%Zhang+ "Drag force for a burning particle"
 2020MNRAS.491.1043N,%Navarrete+ "Magneto-hydrodynamical origin of eclipsing time variations in post-common-envelope binaries..."
 2020arXiv200111929P,%Pusztai+ "Dynamo in weakly collisional non-magnetized plasmas impeded by Landau damping of magneti"
-2020arXiv200612984B,%Brandenburg "Hall cascade with fractional magnetic helicity in neutron star crusts"
+2020ApJ...901...18B,%Brandenburg "Hall cascade with fractional magnetic helicity in neutron star crusts"
 2020arXiv200203638B,%Brandenburg "Piecewise quadratic growth during the 2019 novel coronavirus epidemic"
 2020arXiv200604426V,%Viviani+ "Physically motivated heat conduction treatment in simulations of solar-like stars: ..."
 2020arXiv200602971J,%Jakab+Brandenburg "The effect of a dynamo-generated field on the Parker wind"
@@ -130,6 +134,7 @@ Zhang_etal_2020comb,%Zhang+ "Drag force for a burning particle"
 2020GApFD.114...77G,%Gent "Pencil-Code special edition: modelling supernova driven turbulence" 
 2020GApFD.114..106S,%Schober+ "Chiral fermion asymmetry in high-energy plasma simulations"
 2020GApFD.114..130R,%RoperPol+ "The timestep constraint in solving the gravitational wave equations sourced by hydromagnetic ..."
+2020PhRvD.102h3512R,%Roper Pol+ "Numerical simulations of gravitational waves from early-universe ..."
 2020GApFD.114..162B,%Brandenburg+Das "The time step constraint in radiation hydrodynamics"
 2020GApFD.114..196S,%Singh+ "$f$-mode strengthening from a localized bipolar subsurface magnetic field"
 2020GApFD.114..213C,%Chatterjee+ "Testing Alfv\textbackslash'en wave propagation in a realistic set-up of the solar atmosphere"
@@ -151,7 +156,6 @@ Zhang_etal_2020comb,%Zhang+ "Drag force for a burning particle"
 2019arXiv191006896W,%Warnecke+Peter "On the influence of magnetic helicity on X-rays emission of solar and stellar coronae"
 2019A&A...624L..12W,%Warnecke+Peter "Data-driven model of the solar corona above an active region"
[…]
