### 2020-11-07 #2807 @rtfisher
As mentioned previously, I have some pressing prior obligations to attend
to. I will be able to focus on the review after 11/17.

Best wishes,

Bob
