### 2020-11-08 b139f1a
Added read_pstalk to pencil_old python initializer.
