### 2020-11-08 7c7d2b6
added blobHS_random (HS blobs in entropy and density with random location and time
