### 2020-11-08 a4484ad
Cosmetics in function PencilCode.read.proc_pvar().
Make the lines fit in 80 characters.
