### 2020-11-08 bc695bd
Revise function PencilCode.read.proc_pvar().
Use a cleaner way of converting a dictionary to a namedtuple.
