### 2020-11-08 5688a77
Add function to python module PencilCode.read.
Add allprocs_pvar() to read one particle snapshot under allprocs/.
