### 2020-11-08 8338e72
More cosmetics in PencilCode.read.proc_pvar().
