### 2020-11-08 d5763b0
Add stub keyword to PencilCode.read.pvar().
* Add stub keyword allprocs to read.pvar().

* Reorganize how the consistency of time is checked, considering the
  possibility that some process(s) may not have particles.

* Simplify the construction of the returned namedtuple.
