### 2020-11-08 2d99e28
Add keyword allprocs to PencilCode.read.pvar().
When allprocs is True, the particle data is read under allprocs/.
Otherwise, the data are collected from individual process directories.
