### 2020-11-09 7d794ac
Add program pc_collect_pvar.
This program combines one set of PVAR files from individual processor
directories into one file under data/allprocs/.
