+++ src/pre_and_post_processing/pc_collect_pvar.f90
@@ -1,0 +1,326 @@
+! $Id$
+!***********************************************************************
+!
+! This program combines one set of PVAR files from individual processor
+! directories into one file under data/allprocs/.
+!
+! 07-nov-20/ccyang: coded
+!
+!***********************************************************************
+program pc_collect_pvar
+!
+  use Cparam
+!
+  implicit none
+!
+  integer, parameter :: lun_in = 1, lun_out = 2
+!
+  integer, dimension(ncpus) :: npar_proc = 0
+  character(len=fnlen) :: pvar = "pvar.dat"
+  integer :: ierr
+!
+! Sanity check.
+!
+  if (sanity_check()) stop "failed sanity check. "
+!
+! Prompt for the name of the data file.
+!
+  print *, "Enter the name of the particle data file (e.g., PVAR0): "
+  read *, pvar
+!
+! Get number of particles in each process.
+!
+  call get_npar(pvar, npar_proc)
+!
+! Read and combine particle data files.
+!
+  call read_and_combine(pvar, npar_proc, ierr)
+  if (ierr /= 0) stop "unable to combine the data files. "
+!
+contains
+!***********************************************************************
+  subroutine get_npar(pvar, npar_proc)
+!
+! Reads number of particles in each process from the corresponding data
+! file.
+!
+! 07-nov-20/ccyang: coded
+!
+! Input Arguments
+!     pvar
+!         Name of the particle data file
+!
+! Output Arguments
+!     npar_proc
+!         Number of particles in each process
+!
+    character(len=*), intent(in) :: pvar
+    integer, dimension(ncpus), intent(out) :: npar_proc
+!
+    character(len=fnlen) :: fpath
+    integer :: i, n
+!
+! Name the data file under each process directory.
+!
+    proc: do i = 1, ncpus
+!
+! Open the data file.
+!
+      call open_pvar(i - 1, pvar, lun_in, fpath)
+!
+! Read the number of particles.
+!
+      read(unit=lun_in) n
+      print *, trim(fpath), ":", n, "particles"
+      npar_proc(i) = n
+!
+! Close the file.
+!
+      close(unit=lun_in)
+    enddo proc
+!
+  endsubroutine get_npar
+!***********************************************************************
+  subroutine open_pvar(iproc, pvar, lun, fpath)
+!
+! Open one particle data file under one process directory.
+!
+! 07-nov-20/ccyang: coded
+!
+! Input Arguments
+!     iproc
+!         Process ID
+!     pvar
+!         Name of the particle data file
+!     lun
+!         I/O unit assigned to the file
+!
+! Output Arguments
+!     fpath
+!         Path to the data file
+!
+    use Cdata, only: datadir, datadir_snap, directory, iproc_world
+    use General, only: directory_names_std
+!
+    character(len=*), intent(in) :: pvar
+    integer, intent(in) :: iproc, lun
+    character(len=*), intent(out) :: fpath
+!
+! Compose the path to the data file.
+!
+    datadir_snap = datadir
+    iproc_world = iproc
+    call directory_names_std()
+    fpath = trim(directory) // "/" // pvar
+!
+! Open the file.
+!
+    open(unit=lun, file=fpath, status="old", action="read", form="unformatted")
+!
+  endsubroutine open_pvar
+!***********************************************************************
+  subroutine read_and_combine(pvar, npar_proc, ierr)
+!
+! Reads particle data from individual process directories, combines
+! them, and then writes one single combined file under data/allprocs/.
+!
+! 07-nov-20/ccyang: coded
+!
+! Input Arguments
[…]
