### 2020-11-09 739985b
added coskx dependence on kinematic t-dep flow
