### 2020-11-09 f756a07
Added power1d_plane, to compute 1d power spectra of xy-averaged fields. If this conicides with power_1d I will remove it.
