+++ src/power_spectrum.h
@@ -7,4 +7,4 @@
  public :: power_phi,powerhel_phi, power_vec
  public :: read_power_spectrum_run_pars,  write_power_spectrum_run_pars
  public :: n_spectra
- public :: anisoq_diag, corrfunc_3d, k_omega_spectra
+ public :: anisoq_diag, corrfunc_3d, k_omega_spectra, power1d_plane
