+++ src/param_io.f90
@@ -186,7 +186,7 @@ module Param_IO
       pipe_func, glnCrossSec0, CrossSec_x1, CrossSec_x2, CrossSec_w, &
       cyinyang_intpol_type, yy_biquad_weights, lcutoff_corners, nycut, nzcut, rel_dang, &
       lignore_nonequi, tag_foreign, tau_aver1, fmt_avgs, &
-      ou_omega
+      ou_omega, ou_kzspec, ab_kzspec
 !
   namelist /IO_pars/ &
       lcollective_IO, IO_strategy
