+++ src/power_spectrum.f90
@@ -3987,7 +3987,7 @@ endsubroutine pdf
 !  Use luut_as_aux=T and specify omega_fourier
 !  in run.in under &hydro_run_pars
 !
-!  29-oct-20/hongzhe: added this subroutine and coding
+!  29-oct-20/hongzhe: added this subroutine
 !
     use Fourier, only: fft_xyz_parallel
     use Mpicomm, only: mpireduce_sum
@@ -4157,4 +4157,166 @@ endsubroutine pdf
   !
   endsubroutine k_omega_spectra
 !***********************************************************************
+  subroutine power1d_plane(f,sp)
+!
+!  Calculate power and helicity spectra of planar-averaged 
+!  variable specified by `sp', i.e. either the spectra of uu and kinetic
+!  helicity, or those of bb and magnetic helicity..
+!  Since this routine is only used at the end of a time step,
+!  one could in principle reuse the df array for memory purposes.
+!
+!   9-nov-20/hongzhe: if this coincides with power_1d I will remove it
+!
+    use Fourier, only: fft_xyz_parallel
+    use Mpicomm, only: mpireduce_sum
+    use Sub, only: del2vi_etc, del2v_etc, cross, grad, curli, curl, dot2
+    use Chiral, only: iXX_chiral, iYY_chiral
+!
+  integer, parameter :: nk=nxgrid/2
+  integer :: i, ikx, iky, ikz, jkz, im, in, ivec
+  integer :: k3, k
+  real, dimension (mx,my,mz,mfarray) :: f
+  real, dimension(nx,ny,nz) :: a_re,a_im,b_re,b_im
+  real, dimension(nx) :: bbi
+  real, dimension(nk) :: spectrum,spectrum_sum
+  real, dimension(nk) :: spectrumhel,spectrumhel_sum
+  real, dimension(nxgrid) :: kx
+  real, dimension(nygrid) :: ky
+  real, dimension(nzgrid) :: kz
+  character (len=3) :: sp
+!
+!  passive scalar contributions (hardwired for now)
+!
+  real, dimension(nx,3) :: gtmp1,gtmp2
+!
+!  identify version
+!
+  if (lroot .AND. ip<10) call svn_id("$Id$")
+  !
+  !  Define wave vector, defined here for the *full* mesh.
+  !  Each processor will see only part of it.
+  !  Ignore *2*pi/Lx factor, because later we want k to be integers
+  !
+  kx=cshift((/(i-(nxgrid+1)/2,i=0,nxgrid-1)/),+(nxgrid+1)/2) !*2*pi/Lx
+  ky=cshift((/(i-(nygrid+1)/2,i=0,nygrid-1)/),+(nygrid+1)/2) !*2*pi/Ly
+  kz=cshift((/(i-(nzgrid+1)/2,i=0,nzgrid-1)/),+(nzgrid+1)/2) !*2*pi/Lz
+  !
+  !  initialize power spectrum to zero
+  !
+  spectrum=0.
+  spectrum_sum=0.
+  spectrumhel=0.
+  spectrumhel_sum=0.
+  !
+  !  loop over all the components
+  !
+  do ivec=1,3
+    !
+    !  In fft, real and imaginary parts are handled separately.
+    !  For "kin", calculate spectra of <uk^2> and <ok.uk>
+    !  For "mag", calculate spectra of <bk^2> and <ak.bk>
+    !
+    if (sp=='kin') then
+      if (iuu==0) call fatal_error('powerhel','iuu=0')
+      do n=n1,n2
+        do m=m1,m2
+          call curli(f,iuu,bbi,ivec)
+          im=m-nghost
+          in=n-nghost
[…]
