### 2020-11-09 bebe790
update power1d_plane
