### 2020-11-09 763fcd4
Corrected minus sign in solar Omega-loop initial condition.
