+++ python/tutorials/mps/mps_workshop_Mon.py
@@ -1,291 +1,0 @@
-#!/usr/bin/env python
-# coding: utf-8
-
-# In[ ]:
-
-
-"""Install anaconda and customise python """
-#wget http://fagent.wikidot.com/local--files/python/mps_workshop_Mon.ipynb
-#https://www.anaconda.com/distribution/
-#wget https://repo.anaconda.com/archive/Anaconda3-2019.10-Linux-x86_64.sh
-#wget https://repo.anaconda.com/archive/Anaconda2-2019.10-Linux-x86_64.sh
-#sh Anaconda3-2019.10-Linux-x86_64.sh
-
-#also try conda create --name python=3.7
-"""options to run: python, ipython, jupyter-notebook, etc., and command line, batch script submission"""
-
-#differences with IDL - open source, optional modules need to be loaded, functions in modules, format
-
-#It can be convenient to include a pythonstart file, which will load some default modules of your choice
-#when launching python, ipython and jupyter-notebook, but not batch or command line scripts.
-#If python path aleady contains Pencil Code, then only the last two following lines need be included
-#in your .bashrc file.
-"""
-PENCIL_HOME=$HOME/codes/pencil-code
-export PENCIL_HOME
-export PATH=+$PATH:${PENCIL_HOME}/bin:+${PENCIL_HOME}/utils:+${PENCIL_HOME}/axel:${PENCIL_HOME}/{bin,utils{,/axel},remesh/bin}
-if [ -z $PYTHONPATH ]; then
-    PYTHONPATH="$PENCIL_HOME/python"
-else
-    PYTHONPATH="$PYTHONPATH:$PENCIL_HOME/python"
-fi
-export PYTHONSTARTUP=$HOME/pythonstart
-export PYTHONPATH="$PYTHONPATH:$PYTHONSTARTUP"
-"""
-get_ipython().run_line_magic('pwd', '')
-
-
-# In[ ]:
-
-
-#Replace these directory paths with paths accessible to yourself with pencil simulations and data.
-import pencil as pc
-
-get_ipython().run_line_magic('cd', '')
-
-
-# In[ ]:
-
-
-#Alternative navigation to shell command line tools, python command line calls can be included in scripts
-#to be executed from the command line or batch file
-import os
-PC_HOME = os.getenv("PENCIL_HOME")
-os.chdir(PC_HOME+'/python/tutorials')
-dir1=os.getcwd()
-os.chdir(os.path.join(dir1,'mps/ism'))
-dir2=os.getcwd()
-dir1,dir2
-
-
-# In[ ]:
-
-
-#reading and plotting time series objects, also read parameters files data/param.nml and data/param2.nml
-import matplotlib.pyplot as plt
-get_ipython().run_line_magic('ls', '')
-ts = pc.read.ts()
-par=pc.read.param(quiet=True)
-print(ts.__dict__.keys())
-plt.plot(ts.t,ts.brms)
-#if plot does not display try:
-plt.show()
-
-
-# In[ ]:
-
-
-#To keep plots open in ipython plt.ion()
-
-
-# In[ ]:
-
-
-#Also available semilogy, semilogx and loglog
-plt.semilogy(ts.t,ts.brms,'g:')
-plt.semilogy(ts.t,ts.urms,'-.')
-plt.figure()
-
-
-# In[ ]:
-
-
-# various option for subplot or subplots permit combining plots
[…]
