+++ python/tutorials/mps/fort2h5-call.py
@@ -1,198 +1,0 @@
-#!/usr/bin/env python
-# coding: utf-8
-
-# In[8]:
-
-
-import pencil as pc
-import numpy as np
-import os
-from os.path import join, exists
-import matplotlib.pyplot as plt
-from matplotlib import colors
-from matplotlib import cm
-import glob
-import fileinput
-import sys
-import subprocess as sub
-#TODO rvid_box plots
-#TODO sim objects and hdf5
-""" Tools to copy old binary format simulations into new hdf5 format for protability and continuation
-    Remeshing of old simulations in binary and/or hdf5 format shall follow
-"""
-
-#pencil_home = os.getenv("PENCIL_HOME")
-#home = os.getenv("SCRATCH")
-#os.chdir('/scratch/project_2001062')
-wrkdir=os.getcwd()
-print(wrkdir)
-
-#os.chdir(pencil_home+'/python/tutorials/mps')
-#os.chdir(join(wrkdir,'ism_binary'))
-#identify the top directory for sim_path and list dir contents
-sim_path = os.getcwd()
-sim_list = np.array(glob.glob('*'))
-#par_sim = join(sim_path,'restart')
-print(sim_path)
-print(sim_list)
-
-""" An old simulation which has been executed with fortran binary
-    files distributed across a processor tree can be converted to
-    hdf5, redundant data removed and the run continued in hdf5.
-    This script loops through all items in the sim_path and 
-    identifies which objects are existing simulation objects.
-    If a simulation directory contains the file 'data/grid.h5'
-    it is assumed already to be in hdf5 format so is skipped.
-    For some very old simulations errors in loading data may occur
-    for obsolete 'data/param.nml' files. If this occurs a dummy
-    run with the necessary setup can be compiled started and the
-    new param.nml replace the existing file. This can be automated
-    see comment (c) below.
-"""
-#process dir subdirectories if simulation and if not already hdf5
-for sim in sim_list:
-    print(sim)
-    sim_ = join(sim_path,sim)
-    h5_already=exists(join(sim_,'data/grid.h5'))
-    if pc.sim.is_sim_dir(sim_) and not h5_already:
-        #if simulation is to be processed move to directory
-        os.chdir(sim_)
-        cmd = 'du -csh '+sim_
-        os.system(cmd)
-        #edit Makefile.local from io_dist to io_hdf5 pre-compiling
-        for i, line in enumerate(fileinput.input('src/Makefile.local',
-                                                                    inplace=1)):
-            sys.stdout.write(line.replace('io_dist','io_hdf5'))
-        #print('cleaning '+sim_)
-
-        ##uncomment next two lines if video slices do not exist
-        #cmd = 'touch src/read_all_videofiles.x'
-        #os.system(cmd)
-        #create a bash script to compile pc with src/read_all_videoslices.x
-        #may be commented out if executable already present (a) or data/slice_* complete (b)
-        #amend/comment the lines as required for your fortran compiler and python envcomment
-        #if not exists(join(sim_,'src/read_all_videofiles.x')):
-        #    f = open('pc_compiler','w')
-        #    f.write('#!/bin/bash\n')
-        #    f.write('\n')
-        #    f.write('module purge\n')
-        #    f.write('module load intel\n module load intel-mpi\n module load hdf5/1.10.4-mpi\n')
-        #    f.write('module list\n')
-        #    f.write('\n')
-        #    #minimal next 4 lines required to add executables depending on compiler
-        #    f.write('pc_setupsrc\n')
-        #    #f.write('make distclean\n')
-        #    f.write('pc_build -f Intel_MPI.conf\n')
-        #    f.write('pc_build -f Intel_MPI.conf -t ALL\n')
-        #    f.write('\n')
[…]
