+++ python/tutorials/mps/ism/video.in
@@ -1,4 +1,0 @@
-rho
-lnTT
-uu
-b2

[…]
