### 2020-11-10 455d073
Fix a bug in src/io_mpi2.
In output_snap(), everytime a snippet of data (t, x, y, ...) was
appended to a file.  When the file was the same, like var.dat, it would
leave previous snippets alone and keep growing in size.

Now an output file is truncated to the right size before appended.
