### 2020-11-10 0701ee0
Further change in src/io_mpi2.
For consistency, deltay should also be written when Shear module is on.
