### 2020-11-10 dfeeaa7
Add keep_compiler_quiet_c1d() to General.
Keep compiler quiet with 1D array(s) of character strings.
