### 2020-11-10 3f9388d
Keep compiler quiet in src/io_mpi2.
Keep compiler quiet for unused variables in dummy routines.

TODO: There are still some warnings about uninitialized variables
(buffer).
