### 2020-11-10 b10250f
Fix "used uninitialized" warnings in src/io_mpi2.
In write_persist_*(), the allocatable array buffer could be deallocated
when passed into mpirecv_*(), where the argument is attributed
intent(out).

See for further detail at
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=96312.
