### 2020-11-10 9531b79
Check if snaplink is used in src/io_mpi2.
Put a stop when the module variable snaplink is used.
