### 2020-11-11 2fd902c
Add subroutine to Mpicomm.
Add subroutine mpiallreduce_sum_int_arr_inplace() to reduce the sum of
an integer array in-place.
