+++ python/pencil/math/derivatives/__init__.py
@@ -2,6 +2,6 @@
 Differential operators in different coordinate systems.
 '''
 
-from .der import xder, yder, zder, xder2, yder2, zder2, xder3, yder3, zder3, xder6, yder6, zder6
-from .div_grad_curl import div, curl, grad, curl2, del2, curl3, del6
+from .der import xder, yder, zder, xder2, yder2, zder2, xder3, yder3, zder3, xder5, yder5, zder5, xder6, yder6, zder6
+from .div_grad_curl import div, curl, grad, curl2, del2, curl3, del6, gij, traceless_strain
 from .simple_centered import simple_centered
