### 2020-11-11 5f2755e
Implement output_part_snap() in src/io_mpi2.
NOTE: Only time is written as additional data.  Grid data x, y, z, dx,
dy, and dz are excluded.  This might break backward compatibility, but I
feel it is ok to do so.
