+++ src/io_mpi2.f90
@@ -26,12 +26,12 @@ module Io
   use Cparam, only: intlen, fnlen, max_int
   use File_io, only: file_exists
   use Messages, only: fatal_error, svn_id, warning
+  use mpi
 !
   implicit none
 !
   include 'io.h'
   include 'record_types.h'
-  include 'mpif.h'
 !
 ! Indicates if IO is done distributed (each proc writes into a procdir)
 ! or collectively (eg. by specialized IO-nodes or by MPI-IO).
@@ -295,6 +295,20 @@ module Io
 !
     endsubroutine check_success
 !***********************************************************************
+    subroutine check_success_local(routine, message)
+!
+!  Check success locally of MPI2 file access and issue error if necessary.
+!
+!  11-nov-20/ccyang: coded
+!
+      use Messages, only: fatal_error_local
+!
+      character(len=*), intent(in) :: routine, message
+!
+      if (mpi_err /= MPI_SUCCESS) call fatal_error_local(routine // "_snap", "could not " // message)
+!
+    endsubroutine check_success_local
+!***********************************************************************
     subroutine output_snap(a, nv1, nv2, file, mode)
 !
 !  Write snapshot file, always write mesh and time, could add other things.
@@ -413,26 +427,118 @@ module Io
 !***********************************************************************
     subroutine output_part_snap(ipar, a, mv, nv, file, label, ltruncate)
 !
-!  Write particle snapshot file, always write mesh and time.
+!  Write particle snapshot file.
 !
-!  23-Oct-2018/PABourdin: adapted from output_snap
+!  23-Oct-2018/PABourdin: stub
+!  11-nov-2020/ccyang: coded
 !
       use General, only: keep_compiler_quiet
+      use Mpicomm, only: mpi_precision, mpiallreduce_sum_int
 !
       integer, intent(in) :: mv, nv
-      integer, dimension (mv), intent(in) :: ipar
-      real, dimension (mv,mparray), intent(in) :: a
-      character (len=*), intent(in) :: file
-      character (len=*), optional, intent(in) :: label
-      logical, optional, intent(in) :: ltruncate
+      integer, dimension(mv), intent(in) :: ipar
+      real, dimension(mv,mparray), intent(in) :: a
+      character(len=*), intent(in) :: file
+      character(len=*), intent(in), optional :: label
+      logical, intent(in), optional :: ltruncate
 !
-      call fatal_error ('output_part_snap', 'not implemented for "io_mpi2"', .true.)
-      call keep_compiler_quiet(ipar)
-      call keep_compiler_quiet(a)
-      call keep_compiler_quiet(mv, nv)
-      call keep_compiler_quiet(file)
+      integer, dimension(ncpus) :: npar_proc
+      character(len=fnlen) :: fpath
+      integer :: handle, ftype, npar_tot, ip0
+      integer(KIND=MPI_OFFSET_KIND) :: offset, disp
+!
       if (present(label)) call warning("output_part_snap", "The argument label has no effects. ")
       if (present(ltruncate)) call warning("output_part_snap", "The argument ltruncate has no effects. ")
+!
+!  Broadcast number of particles from each process.
+!
+      npar_proc = 0
+      npar_proc(iproc+1) = nv
+      call mpiallreduce_sum_int(npar_proc, ncpus)
+      npar_tot = sum(npar_proc)
+!
+!  Open snapshot file for write.
+!
+      fpath = trim(directory_snap) // '/' // file
[…]
