### 2020-11-11 05041ed
add output for kappa
