### 2020-11-12 34e41b1
Fix a bug in commit 5f2755ed.
MPI_FILE_SET_VIEW() and MPI_FILE_WRITE_ALL() are collective operations.
