### 2020-11-12 5450958
Clean up output_part_snap() in src/io_mpi2.
