### 2020-11-12 fd13c5d
Revise src/io_mpi2.
Extract code snippet from subroutine output_part_snap() into subroutine
get_disp_to_par_real().  The latter finds the displacements in bytes to
the beginning of the particle real data.
