### 2020-11-12 3e5f673
improved the initialisation of some variables for all cases
