### 2020-11-13 cfeaff2
Implement input_part_snap() in src/io_mpi2.
TODO(ccyang): There still is a bug that proc?_bounds are not written and
read properly.
