+++ src/io_mpi2.f90
@@ -25,7 +25,7 @@ module Io
   use Cdata
   use Cparam, only: intlen, fnlen, max_int
   use File_io, only: file_exists
-  use Messages, only: fatal_error, svn_id, warning
+  use Messages, only: fatal_error, fatal_error_local, svn_id, warning
   use mpi
 !
   implicit none
@@ -300,8 +300,6 @@ module Io
 !  Check success locally of MPI2 file access and issue error if necessary.
 !
 !  11-nov-20/ccyang: coded
-!
-      use Messages, only: fatal_error_local
 !
       character(len=*), intent(in) :: routine, message
 !
@@ -730,27 +728,140 @@ module Io
 !
     endsubroutine input_snap_finalize
 !***********************************************************************
-    subroutine input_part_snap(ipar, ap, mv, nv, npar_total, file, label)
+    subroutine input_part_snap(ipar, ap, mv, nv, npar_tot, file, label)
 !
-!  Read particle snapshot file, mesh and time are read in 'input_snap'.
+!  Read particle snapshot file.
 !
-!  24-Oct-2018/PABourdin: apadpted and moved to IO module
+!  24-Oct-2018/PABourdin: stub
+!  12-nov-20/ccyang: coded
 !
       use General, only: keep_compiler_quiet
+      use Mpicomm, only: mpi_precision
+      use Particles_cdata, only: ixp, iyp, izp
 !
       integer, intent(in) :: mv
-      integer, dimension (mv), intent(out) :: ipar
-      real, dimension (mv,mparray), intent(out) :: ap
-      integer, intent(out) :: nv, npar_total
-      character (len=*), intent(in) :: file
-      character (len=*), optional, intent(in) :: label
+      integer, dimension(mv), intent(out) :: ipar
+      real, dimension(mv,mparray), intent(out) :: ap
+      integer, intent(out) :: nv, npar_tot
+      character(len=*), intent(in) :: file
+      character(len=*), intent(in), optional :: label
 !
-      call fatal_error ('input_part_snap', 'not implemented for "io_mpi2"', .true.)
-      call keep_compiler_quiet(ipar)
-      call keep_compiler_quiet(ap)
-      call keep_compiler_quiet(mv, nv, npar_total)
-      call keep_compiler_quiet(file)
+      logical, allocatable, dimension(:) :: lpar_loc
+      integer, allocatable, dimension(:) :: indices
+!
+      real, dimension(mv*ncpus) :: rbuf
+      character(len=fnlen) :: fpath
+      integer(KIND=MPI_OFFSET_KIND) :: offset, esize
+      integer :: handle, ftype, ftype1, i
+!
       if (present(label)) call warning("input_part_snap", "The argument label has no effects. ")
+!
+!  Open snapshot file for read.
+!
+      fpath = trim(directory_snap) // '/' // file
+      call MPI_FILE_OPEN(mpi_comm, fpath, MPI_MODE_RDONLY, io_info, handle, mpi_err)
+      call check_success("input_part", "open", fpath)
+!
+!  Read total number of particles.
+!
+      call MPI_FILE_SET_VIEW(handle, 0_MPI_OFFSET_KIND, MPI_INT, MPI_INT, "native", io_info, mpi_err)
+      call check_success("input_part", "set view of", fpath)
+!
+      call MPI_FILE_READ_ALL(handle, npar_tot, 1, MPI_INT, status, mpi_err)
+      call check_success("input_part", "read total number of particles", fpath)
+!
+      cknp: if (npar_tot > mv * ncpus) then
[…]
