### 2020-11-13 915c71f
Revise get_disp_to_par_real() in src/io_mpi2.
The displacement to the beginning of the real data can be computed
without accessing the file handle.
