### 2020-11-13 960f123
Revise output_part_snap() in src/io_mpi2.
Use collective MPI write instead of non-collective.
