### 2020-11-13 fe91744
Centralize some common constants in src/io_mpi2.
Consolidate the use of some MPI constants and the sizes of some MPI
elementary data types.
