+++ src/io_mpi2.f90
@@ -26,6 +26,7 @@ module Io
   use Cparam, only: intlen, fnlen, max_int
   use File_io, only: file_exists
   use Messages, only: fatal_error, fatal_error_local, svn_id, warning
+  use Mpicomm, only: mpi_precision
   use mpi
 !
   implicit none
@@ -49,7 +50,8 @@ module Io
   integer, parameter :: order=MPI_ORDER_FORTRAN, io_info=MPI_INFO_NULL
 !  integer, dimension(io_dims) :: local_size, local_start, global_size, global_start, subsize
   integer, dimension (:), allocatable :: local_size, local_start, global_size, global_start, subsize
-
+!
+  integer(kind=MPI_OFFSET_KIND) :: intsize = 0, realsize = 0
 !
   contains
 !***********************************************************************
@@ -168,7 +170,15 @@ module Io
 !
       if (lread_from_other_prec) &
         call warning('register_io','Reading from other precision not implemented')
-
+!
+!  Remeber the sizes of some MPI elementary types.
+!
+      call MPI_TYPE_SIZE_X(MPI_INT, intsize, mpi_err)
+      if (mpi_err /= MPI_SUCCESS) call fatal_error_local("register_io", "could not find MPI_INT size")
+!
+      call MPI_TYPE_SIZE_X(mpi_precision, realsize, mpi_err)
+      if (mpi_err /= MPI_SUCCESS) call fatal_error_local("register_io", "could not find MPI real size")
+!
     endsubroutine register_io
 !***********************************************************************
     subroutine finalize_io
@@ -307,7 +317,7 @@ module Io
 !
     endsubroutine check_success_local
 !***********************************************************************
-    subroutine get_disp_to_par_real(npar_tot, disp)
+    pure integer(KIND=MPI_OFFSET_KIND) function get_disp_to_par_real(npar_tot) result(disp)
 !
 !  Gets the displacement in bytes to the beginning of particle real
 !  data.
@@ -315,17 +325,10 @@ module Io
 !  12-nov-20/ccyang: coded
 !
       integer, intent(in) :: npar_tot
-      integer(KIND=MPI_OFFSET_KIND), intent(out) :: disp
-!
-      integer(KIND=MPI_OFFSET_KIND) :: intsize
-      integer :: ierr
-!
-      call MPI_TYPE_SIZE_X(MPI_INT, intsize, ierr)
-      if (ierr /= MPI_SUCCESS) call fatal_error_local("get_disp_to_par_real", "could not find MPI_INT size")
 !
       disp = int(1 + npar_tot, KIND=MPI_OFFSET_KIND) * intsize
 !
-    endsubroutine get_disp_to_par_real
+    endfunction get_disp_to_par_real
 !***********************************************************************
     subroutine output_snap(a, nv1, nv2, file, mode)
 !
@@ -335,7 +338,7 @@ module Io
 !  13-feb-2014/MR: made file optional (prep for downsampled output)
 !
       use General, only: ioptest
-      use Mpicomm, only: globalize_xy, collect_grid, mpi_precision
+      use Mpicomm, only: globalize_xy, collect_grid
 !
       integer, optional, intent(in) :: nv1,nv2
       real, dimension (:,:,:,:), intent(in) :: a
@@ -450,7 +453,7 @@ module Io
 !  23-Oct-2018/PABourdin: stub
 !  11-nov-2020/ccyang: coded
 !
-      use Mpicomm, only: mpi_precision, mpiallreduce_sum_int
[…]
