### 2020-11-13 c90fcf2
Change the interface to read_snapshot_particles().
The original argument snap_directory is no longer needed.
