### 2020-11-13 1f2bffc
Fix subroutine output_part_snap() in src/io_mpi2.
Make sure the output file is properly truncated before adding additional
data.
