### 2020-11-13 9e3e2db
enhanced diagnostics
