### 2020-11-13 d4b6d3f
Fix a bug in input_part_snap() of src/io_mpi2.
It did not properly take into account 1D and 2D cases when identifying
local particles.
