### 2020-11-14 8fce9fa
Redistribute slice IO from HDF5_IO back to IO - 1.
This is in preparation for src/io_mpi2 (as well as other IO module in
the future) to be able to control how to write slices file by itself.

Step 1. Redistribute subroutine input_slice().
