### 2020-11-14 b416d9e
Redistribute slice IO from HDF5_IO back to IO - 2.
Follow commit 8fce9fac.

This is in preparation for src/io_mpi2 (as well as other IO module in
the future) to be able to control how to write slices file by itself.

Step 2. Redistribute subroutine output_slice().
