### 2020-11-14 a850937
Redistribute slice IO from HDF5_IO back to IO - 3.
Follow commit 87abeeae.

This is in preparation for src/io_mpi2 (as well as other IO module in
the future) to be able to control how to write slices file by itself.

Step 3. Redistribute subroutine output_slice().

This completes the redistribution.
