+++ src/hdf5_io_parallel.f90
@@ -14,6 +14,11 @@ module HDF5_IO
 !
   implicit none
 !
+  interface hdf5_input_slice
+    module procedure input_slice_arr
+    module procedure input_slice_scat
+  endinterface
+!
   interface input_hdf5
     module procedure input_hdf5_string
     module procedure input_hdf5_int_0D
