### 2020-11-14 26a9baa
Fix a missing interface in src/hdf5_io_parallel.
NOTES: It appears that no one is calling the subroutine
input_slice_arr() as well as input_slice_real_arr() in general.
