### 2020-11-14 06b8bc8
reduced MAUX CONTRIBUTION to 2 from 4 in particles_dust.f90. Thank chao-CHin for pointing this out.
