### 2020-11-14 eced47a
Initiate development of slice I/O in src/io_mpi2.
Make input_slice(), output_slice(), and output_slice_position() stub
subroutines and hence remove the dependence of io_mpi2 on HDF5_io.
