### 2020-11-14 66e5394
Revise input_part_snap() of src/io_mpi2.
Directly use an array section to read instead of a loop.
