### 2020-11-14 ee11384
Implement output_slice_position() of src/io_mpi2.
Make it a dummy subroutine; there is no need to record slice positions.
