### 2020-11-15 05c3b29
Cosmetics in python/.
Make function read.slices() of the Python module PencilCode fit with the
line length of 80 characters.
