### 2020-11-15 519909b
Add two subroutines to src/mpicomm.f90.
Add in-place versions for subroutines mpiallreduce_sum() and
mpiallreduce_or().
