### 2020-11-15 b0e30b3
in cdata, added +  logical :: lproc_pt=.false., lproc_p2=.false. +  integer :: iproc_pt=0, iproc_p2=0
in cparam, changed
-  integer, parameter :: ilabel_max=-1,ilabel_sum=1,ilabel_save=0
+  integer, parameter :: ilabel_max=-1,ilabel_sum=1,ilabel_save=11

in diagnostics, added
+              if (itype==ilabel_save)            &
+                  vname(iname)=fsum(isum_count)
and changed
-                 if (lsphere_in_a_box)    vol=1.333333*pi*(r_ext**3-r_int**3)
+                 if (lsphere_in_a_box)    vol=four_pi_over_three*(r_ext**3-r_int**3)

in grid, now do lpoint=min(max(l1,lpoint),l2
[…]
