### 2020-11-15 3d7c01a
Revise Python function PencilCode.read.slices().
* Remove preprocessing the time points.
* Also read in the slice positions (as a functino of time) of each
  plane.
* Add keyword argument return_pos to turn on/off the return of the slice
  positions.
