### 2020-11-15 42b0e1f
Adapt python function PencilCode.read.slices().
Make it also be able to read slice data with IO=io_mpi2.
