### 2020-11-15 2106af3
Implement output_slices() of src/io_mpi2.
This should complete the implementation of IO=io_mpi2 for most of the
applications.
