+++ src/io_mpi2.f90
@@ -292,6 +292,53 @@ module Io
 !
     endsubroutine distribute_grid
 !***********************************************************************
+    subroutine check_consistency(x, mask)
+!
+!  Check if a scalar value is consistent between active processes.
+!
+!  15-nov-20/ccyang: coded
+!
+!  Input Arguments
+!      xlocal
+!          Local value.
+!  Optional Input Arguments
+!      mask
+!          If present, processes to be included in the check.
+!
+      use Mpicomm, only: mpiallreduce_sum
+!
+      real, intent(in) :: x
+      logical, dimension(ncpus), intent(in), optional :: mask
+!
+      real, dimension(ncpus) :: buf
+      real :: xmin, xmax
+!
+!  Collect the local values.
+!
+      buf = 0.0
+      buf(iproc+1) = x
+      call mpiallreduce_sum(buf, ncpus)
+!
+!  Find the range of the values.
+!
+      val: if (present(mask)) then
+        if (.not. any(mask)) return
+        xmin = minval(buf, mask=mask)
+        xmax = maxval(buf, mask=mask)
+      else val
+        xmin = minval(buf)
+        xmax = maxval(buf)
+      endif val
+!
+!  Determine if the local values are consistent.
+!
+      cons: if (xmin /= xmax) then
+        print *, "check_consistency: xmin, xmax = ", xmin, xmax
+        call fatal_error("check_consistency", "local values are not consistent")
+      endif cons
+!
+    endsubroutine check_consistency
+!***********************************************************************
     subroutine check_success(routine, message, file)
 !
 !  Check success of MPI2 file access and issue error if necessary.
@@ -315,6 +362,37 @@ module Io
       if (mpi_err /= MPI_SUCCESS) call fatal_error_local(trim(routine) // "_snap", "could not " // trim(message))
 !
     endsubroutine check_success_local
+!***********************************************************************
+    subroutine get_dimensions(dir, nxyzgrid, nxyz, ipxyz)
+!
+!  Get n[xyz]grid, n[xyz], and ip[xyz] according to dir.
+!
+!  14-nov-20/ccyang: coded
+!
+      character, intent(in) :: dir
+      integer, intent(out) :: nxyzgrid, nxyz, ipxyz
+!
+      chdir: if (dir == 'x') then
+        nxyzgrid = nxgrid
+        nxyz = nx
+        ipxyz = ipx
+!
+      elseif (dir == 'y') then chdir
+        nxyzgrid = nygrid
+        nxyz = ny
+        ipxyz = ipy
+!
+      elseif (dir == 'z') then chdir
+        nxyzgrid = nzgrid
+        nxyz = nz
+        ipxyz = ipz
+!
+      else chdir
+        call fatal_error("get_dimensions", "unknown direction " // dir)
+!
+      endif chdir
+!
+    endsubroutine get_dimensions
 !***********************************************************************
     pure integer(KIND=MPI_OFFSET_KIND) function get_disp_to_par_real(npar_tot) result(disp)
 !
@@ -457,10 +535,10 @@ module Io
 !
 !  Append to a slice file.
 !
-!  14-nov-20/ccyang: stub
+!  15-nov-20/ccyang: coded
 !
       use General, only: keep_compiler_quiet
-      use Messages, only: not_implemented
+      use Mpicomm, only: mpiallreduce_or
 !
       real, dimension(:,:), pointer :: data
       character(len=*), intent(in) :: label, suffix
[…]
