### 2020-11-15 ef612e0
Revise Python function PencilCode.read.slices().
Use namedtuple instead of numpy record array to construct the slice
data.
