### 2020-11-15 b3df6d0
Cosmetics of function PencilCode.read.proc_dim().
Also simply the construction of the namedtuple.
