### 2020-11-15 af3b040
Add py function PencilCode.read.allprocs_grid().
Adapted from read.proc_grid() to read grid.dat under allprocs/.
