### 2020-11-15 cff4f8f
Add keyword to py function PencilCode.read.grid().
Add keyword allprocs to read grid from under allprocs/.
