+++ python/PencilCode/read.py
@@ -355,10 +355,13 @@ def dimensions(datadir='./data'):
                       mvar=mvar, maux=maux, mglobal=mglobal, double_precision=double_precision,
                       nprocx=nprocx, nprocy=nprocy, nprocz=nprocz, procz_last=procz_last)
 #=======================================================================
-def grid(datadir='./data', interface=False, par=None, trim=True):
+def grid(allprocs=False, datadir='./data', interface=False, par=None,
+         trim=True):
     """Returns the coordinates and their derivatives of the grid.
 
     Keyword Arguments:
+        allprocs
+            If True, the grid is read under allprocs/.
         datadir
             Name of the data directory.
         interface
@@ -374,68 +377,80 @@ def grid(datadir='./data', interface=False, par=None, trim=True):
     """
     # Author: Chao-Chin Yang
     # Created: 2014-11-02
-    # Last Modified: 2017-07-03
+    # Last Modified: 2020-11-15
     from collections import namedtuple
     import numpy as np
 
     # Get the dimensions.
     dim = dimensions(datadir=datadir)
 
-    # Allocate arrays.
-    x = np.zeros(dim.mxgrid,)
-    y = np.zeros(dim.mygrid,)
-    z = np.zeros(dim.mzgrid,)
-    dx, dy, dz, Lx, Ly, Lz = 0.0, 0.0, 0.0, 0.0, 0.0, 0.0
-    dx_1 = np.zeros(dim.mxgrid,)
-    dy_1 = np.zeros(dim.mygrid,)
-    dz_1 = np.zeros(dim.mzgrid,)
-    dx_tilde = np.zeros(dim.mxgrid,)
-    dy_tilde = np.zeros(dim.mygrid,)
-    dz_tilde = np.zeros(dim.mzgrid,)
+    if allprocs:
+        # Read grid from under allprocs/ and unpack.
+        grid = allprocs_grid(datadir=datadir, dim=dim)
+        x, y, z = grid.x, grid.y, grid.z
+        dx, dy, dz = grid.dx, grid.dy, grid.dz
+        Lx, Ly, Lz = grid.Lx, grid.Ly, grid.Lz
+        dx_1, dy_1, dz_1 = grid.dx_1, grid.dy_1, grid.dz_1
+        dx_tilde = grid.dx_tilde
+        dy_tilde = grid.dy_tilde
+        dz_tilde = grid.dz_tilde
 
-    # Define functions for assigning local coordinates to global.
-    def assign(l1, l2, xg, xl, label, proc):
-        indices = xg[l1:l2] != 0.0
-        if any(xg[l1:l2][indices] != xl[indices]):
-            print("Warning: inconsistent ", label, " for process", proc)
-        xg[l1:l2][~indices] = xl[~indices]
-        return xg
-    def assign1(old, new, label, proc):
-        if old != 0.0 and old != new:
-            print("Warning: inconsistent ", label, " for process", proc)
-        return new
+    else:
+        # Allocate arrays.
+        x = np.zeros(dim.mxgrid,)
+        y = np.zeros(dim.mygrid,)
+        z = np.zeros(dim.mzgrid,)
+        dx, dy, dz, Lx, Ly, Lz = 0.0, 0.0, 0.0, 0.0, 0.0, 0.0
+        dx_1 = np.zeros(dim.mxgrid,)
+        dy_1 = np.zeros(dim.mygrid,)
+        dz_1 = np.zeros(dim.mzgrid,)
+        dx_tilde = np.zeros(dim.mxgrid,)
+        dy_tilde = np.zeros(dim.mygrid,)
[…]
