### 2020-11-15 e0f51b9
Remove keyword from PencilCode.read.grid().
Revert commit 8b1faf5c and automatically determine where to read
grid.dat from.
