### 2020-11-15 dc08211
Fix Python function PencilCode.read.slices().
The slice data was not recorded as numpy arrays.
