### 2020-11-15 b905219
Adapt Python function pc.animate._slices2d().
Change the method of accessing the slice data due to the change of the
data type.
