+++ python/PencilCode/animate.py
@@ -382,8 +382,11 @@ def _slices2d(field, t, slices, dim, par, grid, **kwarg):
         **kwarg
             Keyword arguments passed to matplotlib.pyplot.figure().
     """
-    # Chao-Chin Yang, 2015-05-27
+    # Author: Chao-Chin Yang
+    # Created: 2015-04-22
+    # Last Modified: 2015-05-27
     import numpy as np
+
     # Determine the slice plane.
     if dim.nxgrid == 1:
         plane = 'yz'
@@ -394,17 +397,21 @@ def _slices2d(field, t, slices, dim, par, grid, **kwarg):
     else:
         plane = 'xy'
         xdir, ydir = 0, 1
+
     # Check the coordinate system.
     if par.coord_system != 'cartesian':
         raise NotImplementedError("2D, curvilinear model")
     xlabel, ylabel = "xyz"[xdir], "xyz"[ydir]
     x, y = getattr(grid, xlabel), getattr(grid, ylabel)
+
     # Replace the labels if in **kwarg.
     xlabel = kwarg.pop("xlabel", xlabel)
     ylabel = kwarg.pop("ylabel", ylabel)
     clabel = kwarg.pop("clabel", field)
+
     # Send to the animator.
-    _frame_rectangle(t, x, y, slices[:][plane], xlabel=xlabel, ylabel=ylabel, clabel=clabel, **kwarg)
+    _frame_rectangle(t, x, y, getattr(slices, plane),
+            xlabel=xlabel, ylabel=ylabel, clabel=clabel, **kwarg)
 #=======================================================================
 def _slices3d(field, t, slices, dim, par, grid, **kwarg):
     """Animates video slices from a 3D model.
