### 2020-11-16 0fe9c22
Cosmetics of Python function pc.read.dimensions().
