### 2020-11-16 776074a
Adapt samples/2d-tests/streaming_instability/.
Make the sample use IO=io_mpi2 so that MPI-IO is under regular
auto-tests.
