+++ samples/2d-tests/streaming_instability/video.in
@@ -1,0 +1,2 @@
+rho
+rhop
