### 2020-11-16 6e66734
Make allprocs=True default in module PencilCode.
This applies to read.var() and read.pvar().
