### 2020-11-16 cf304c5
Temporarily change the header of src/io_mpi2.f90.
The openmpi used by the hourly-test server is too old to support the
more robust way of accessing the MPI interface.
