+++ python/pencil/io/npfile.py
@@ -65,8 +65,10 @@ class npfile(object):
                  endian = 'dtype',
                  order = 'C'):
         if 'b' not in permission: permission += 'b'
-        if isinstance(file_name, basestring):
-            self.file = file(file_name, permission)
+        #if isinstance(file_name, basestring):
+        if isinstance(file_name, str):
+            #self.file = file(file_name, permission)
+            self.file = open(file_name, permission)
         else:
             try:
                 closed = file_name.closed
