### 2020-11-16 d0c168b
added "lroot.and." in ! !  save tdiagnos/myr, but do it only on root processor. !         if (lroot.and.idiag_tmyr/=1) call save_name(tdiagnos/myr,idiag_tmyr)
