+++ src/cparam.f90
@@ -141,7 +141,7 @@ module Cparam
 !  using mpi_reduce_sum;
 !  value 0 causes the value simply to be used from the root processor.
 !
-  integer, parameter :: ilabel_max=-1,ilabel_sum=1,ilabel_save=0
+  integer, parameter :: ilabel_max=-1,ilabel_sum=1,ilabel_save=11
   integer, parameter :: ilabel_max_sqrt=-2,ilabel_sum_sqrt=2
   integer, parameter :: ilabel_sum_log10=10
   integer, parameter :: ilabel_max_dt=-3,ilabel_max_neg=-4
