### 2020-11-17 7f2032c
added SX and hX terms for pt
