### 2020-11-17 e6f4d2b
Deleted duplicate python/TO_DO file and merging the content
