+++ python/TO_DO_LIST.txt
@@ -1,7 +1,0 @@
-Please, add to the list any issue you encounter during the usage of the python module:
-
-
--Check that every file works independently if you load the full pencil module or not.
--Check duplicate function names: for instance, I think having "io/npfile.py" locally, whereas you als have scipy.io.npfile is not a good idea.
--Check the manuals and wiki and update obsolete functions to the new version of the code.
--Adding new functions available in IDL
