+++ license/developers.txt
@@ -11,9 +11,9 @@ In such cases that person must expect phone calls even outside office hours.
 
 Name,                 email,                         svn-name,             office,            mobile/home/alternative,  skype-name,  institution
 ------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-Omer Anjum,           oanjum@illinois.edu,           oa98105,              ,                  ,                         ,            Aalto Univ. & Illinois, USA
+Omer Anjum,           oanjum/illinois.edu,           oa98105,              ,                  ,                         ,            Aalto Univ. & Illinois, USA
 Natalia Babkovskaia,  nbabkovskaia/gmail.com,        NBabkovskaia,         +358 8 553 1939,   +358 40 5286784,          ,            University of Helsinki
-Iouri Belokopytov,    iourib@kth.se,                 ,                     +46 8 553 784 29   ,                         ,            Nordita
+Iouri Belokopytov,    iourib/kth.se,                 ,                     +46 8 553 784 29   ,                         ,            Nordita
 Steve Berukoff,       steveb/physics.ucla.edu,       ,                     ,                  ,                         ,            National Solar Observatory Boulder
 Sven Bingert,         bingert/mps.mpg.de,            sven.bingert,         +49 5556 979 358,  +49 179 2126192,          ,            GWDG Gottingen
 Philippe Bourdin,     pc/bourdin.ch,                 PABourdin,            +43 316 4120 592,  +43 720 513 493,          angiebecker.ch, University of Graz
@@ -25,7 +25,7 @@ Fabio Del Sordo,      fabio/nordita.org,             fadiesis,             +46 8
 Boris Dintrans,       boris.dintrans/gmail.com,    	     boris.dintrans,       +33 4 67141400,                 , bdintrans,   CNRS/CINES
 Wolfgang Dobler,      wdobler/gmail.com,             wdobler,              +49 33201 43791,   +49 151 28123472,         wolfgangdobler,
 Thomas Gastine,       gastine/mps.mpg.de,            ,                     +49 5556 979 456,  ,                         ,            Max-Planck-Institut für Sonnensystemforschung Göttingen
-Fred A. Gent,         frederick.gent@aalto.fi,       fred.gent.ncl,        +358 50 344 9122,  +44 7779 364 893 (mob),   fred_gent,   Aalto University
+Fred A. Gent,         frederick.gent/aalto.fi,       fred.gent.ncl,        +358 50 344 9122,  +44 7779 364 893 (mob),   fred_gent,   Aalto University
 Gustavo Guerrero,     tavo.buk/gmail.com,            ,                     +46 8 5537 8718,   +46 76 089 1432,          ,            Universidade Federal de Minas Gerais
 Anders Johansen,      anders/astro.lu.se,            ,                     +46 46 22 21589,   +46 736 84 96 98,         ,            Lund Observatory
 Nils E.L. Haugen,     Nils.E.Haugen/sintef.no,       ,                     +47 986 08 683,    +47 986 08 683,           ,            Norges teknisk-naturvitenskapelige universitet
