### 2020-11-18 5e292f2
Fix a bug in py function PencilCode.read.grid().
Incorrect operator for logical equality.
