### 2020-11-18 9c8b205
MR: new function plegendre for (associated) Legendre polynomials     originally by Mariangela from NumRecipies
