### 2020-11-19 143f64e
added gT-gX output for pt and p2 as well
