### 2020-11-20 828ebce
The subroutine k_omega_spectra can now produce azimuthally averaged 2d spectra in (modified) polar coordinates (kr,kmu), where kr=sqrt ( kx^2 + ky^2 + kz^2 ) kmu=the cosine of the polar angle=kz/|k|
