+++ src/power_spectrum.f90
@@ -34,6 +34,8 @@ module power_spectrum
   real :: pdf_max=30., pdf_min=-30., pdf_max_logscale=3.0, pdf_min_logscale=-3.
   logical :: lintegrate_shell=.true., lintegrate_z=.true., lcomplex=.false.
   logical :: lhalf_factor_in_GW=.false., lcylindrical_spectra=.false.
+  logical :: lcyl_polar_spectra=.false.
+  integer :: legendre_lmax=1.
   integer :: firstout = 0
 !
   character (LEN=linelen) :: ckxrange='', ckyrange='', czrange=''
@@ -45,7 +47,8 @@ module power_spectrum
   namelist /power_spectrum_run_pars/ &
       lintegrate_shell, lintegrate_z, lcomplex, ckxrange, ckyrange, czrange, &
       lcylindrical_spectra, inz, n_segment_x, lhalf_factor_in_GW, &
-      pdf_max, pdf_min, pdf_min_logscale, pdf_max_logscale
+      pdf_max, pdf_min, pdf_min_logscale, pdf_max_logscale, &
+      lcyl_polar_spectra, legendre_lmax
 !
   contains
 !***********************************************************************
@@ -3988,6 +3991,7 @@ endsubroutine pdf
 !  in run.in under &hydro_run_pars
 !
 !  29-oct-20/hongzhe: added this subroutine
+!  20-nov-20/hongzhe: can now also compute Legendre coefficients
 !
     use Fourier, only: fft_xyz_parallel
     use Mpicomm, only: mpireduce_sum
@@ -3995,8 +3999,10 @@ endsubroutine pdf
     use Chiral, only: iXX_chiral, iYY_chiral
 !
   integer, parameter :: nk=nxgrid/2
-  integer :: i, k, ikx, iky, ikz, jkz, im, in, ivec, ivec_jj
-  real :: k2
+  integer :: i, k, ikx, iky, ikz, jkz, im, in, ivec, ivec_jj, ikr, ikmu
+  integer, dimension(nk) :: nmu
+  real, allocatable, dimension(:,:) :: kmu
+  real :: k2,mu
   real, dimension (mx,my,mz,mfarray) :: f
   real, dimension(nx,ny,nz) :: a_re,a_im,b_re,b_im
   real, dimension(nx) :: bbi, jji, b2, j2
@@ -4005,9 +4011,12 @@ endsubroutine pdf
   real, dimension(nk) :: k2m=0.,k2m_sum=0.,krms
   real, dimension(nk,nzgrid) :: cyl_spectrum, cyl_spectrum_sum
   real, dimension(nk,nzgrid) :: cyl_spectrumhel, cyl_spectrumhel_sum
+  real, allocatable, dimension(:,:) :: cyl_polar_spec, cyl_polar_spec_sum
+  real, allocatable, dimension(:,:) :: cyl_polar_spechel, cyl_polar_spechel_sum
   real, dimension(nxgrid) :: kx
   real, dimension(nygrid) :: ky
   real, dimension(nzgrid) :: kz
+  integer, dimension(1) :: temp
   character (len=3) :: sp
   logical, save :: lwrite_krms=.true., lwrite_krms_GWs=.false.
 !
@@ -4037,6 +4046,31 @@ endsubroutine pdf
     cyl_spectrum_sum=0.
     cyl_spectrumhel=0.
     cyl_spectrumhel_sum=0.
+  ! allow for polar representation
+    if (lcyl_polar_spectra) then
+      do ikr=1,nk
+        nmu(ikr)=2*FLOOR(0.84815*ikr/2)+1
+      enddo
+      !
[…]
