### 2020-11-21 2d37ea4
MR: removed tentative polar grid code;     marked all (?) places where changes for polar grid are needed;     some simplifications and corrections;
