+++ src/deriv.f90
@@ -129,17 +129,19 @@ module Deriv
       integer, intent(in) :: j, k
       logical, intent(in), optional :: ignoredx
 !
-      real, parameter :: a = 1.0 / 60.0
+      real, parameter :: a = 1.0/60.0
       real, dimension(nx) :: fac
+      real :: facs
       logical :: withdx
-      integer :: nphi,n1n,n1p,n2p,n2n,n3p,n3n
 !
 !debug      if (loptimise_ders) der_call_count(k,icount_der,j,1) = & !DERCOUNT
-!debug                            der_call_count(k,icount_der,j,1)+1 !DERCOUNT
+!debug                          der_call_count(k,icount_der,j,1)+1 !DERCOUNT
 !
       if (present(ignoredx)) then
         withdx = .not. ignoredx
-        if (ignoredx) fac = a
+        if (ignoredx) then
+          fac = a; facs = a
+        endif
       else
         withdx = .true.
       endif
@@ -156,49 +158,29 @@ module Deriv
         endif
       elseif (j==2) then
         if (nygrid/=1) then
-          if (withdx) fac = a * dy_1(m)
-          df=fac*(+ 45.0*(f(l1:l2,m+1,n,k)-f(l1:l2,m-1,n,k)) &
-                  -  9.0*(f(l1:l2,m+2,n,k)-f(l1:l2,m-2,n,k)) &
-                  +      (f(l1:l2,m+3,n,k)-f(l1:l2,m-3,n,k)))
-          if (withdx .and. lspherical_coords  ) df = df * r1_mn
-          if (withdx .and. lcylindrical_coords) df = df * rcyl_mn1
+          if (withdx) facs = a * dy_1(m)
+          df=facs*(+ 45.0*(f(l1:l2,m+1,n,k)-f(l1:l2,m-1,n,k)) &
+                   -  9.0*(f(l1:l2,m+2,n,k)-f(l1:l2,m-2,n,k)) &
+                   +      (f(l1:l2,m+3,n,k)-f(l1:l2,m-3,n,k)))
+          if (withdx) then
+            if (lspherical_coords) then
+              df = df * r1_mn
+            elseif (lcylindrical_coords) then
+              df = df * rcyl_mn1
+            endif
+          endif
         else
           df=0.
           if (ip<=5) print*, 'der_main: Degenerate case in y-direction'
         endif
       elseif (j==3) then
         if (nzgrid/=1) then
-          if (lpole(2) .and. lcoarse) then
-            if (lfirst_proc_y .and. m<m1+1.5*ncoarse .and. m>=m1) then
-              nphi = max(mod(int(ncoarse/(m-m1+1)),ncoarse+1),1)
-            elseif (llast_proc_y .and. m>m2-1.5*ncoarse .and. m<=m2) then
-              nphi = max(mod(int(ncoarse/(m2-m+1)),ncoarse+1),1)
-            else
-              nphi = 1
-            endif
-            fac = a * dz_1(n)/nphi * r1_mn * sin1th(m)
-            n1p = n+1*nphi
-            if (n1p > n2+nghost) n1p = n1p - n2 - nghost
-            n2p = n+2*nphi
-            if (n2p > n2+nghost) n2p = n2p - n2 - nghost
-            n3p = n+3*nphi
-            if (n3p > n2+nghost) n3p = n3p - n2 - nghost
-            n1n = n-1*nphi
-            if (n1n < 1) n1n = n1n + n2 + nghost
[…]
