### 2020-11-21 d12493e
Added script for writing (non-var) 3D arrays in VTK format. Adapted from arrays2d_to_vtk.py.
