+++ python/pencil/export/__init__.py
@@ -3,6 +3,7 @@ Mostly routines for exporting data into different formats.
 '''
 
 from .arrays2d_to_vtk import arrays2d_to_vtk
+from .arrays3d_to_vtk import arrays3d_to_vtk
 from .particles_to_vtk import particles_to_vtk, ParticlesVtk
 from .create_h5 import create_aver_sph, fvars
 from .pc2vtk import var2vtk, slices2vtk#, aver2vtk, power2vtk
