### 2020-11-22 bda8ab0
Update quick_start.tex
Added about verification of results.
