+++ doc/quick-guide/quick_start.tex
@@ -219,10 +219,25 @@ If everything worked well, your output should contain the lines
 \begin{verbatim}
   start.x has completed successfully
 \end{verbatim}
-and
+after initiallizing everything successfully, it would then start running,   
+printing on the prompt the quantities specified in \file{print.in}, for 
+instance,
 \begin{verbatim}
+---it--------t-------dt------rhom------urms------uxpt-----uypt-----uzpt-----
+       0      0.00 4.9E-03 1.000E+00  1.414E+00 2.00E+00 0.00E+00 0.00E+00
+      10      0.05 4.9E-03 1.000E+00  1.401E+00 1.98E+00 0.00E+00 0.00E+00
+      20      0.10 4.9E-03 1.000E+00  1.361E+00 1.88E+00 0.00E+00 0.00E+00 
+\end{verbatim}
+ending with
+\begin{verbatim}
+  Simulation finished after        xxxx  time-steps
+  .....
   Wall clock time/timestep/meshpoint [microsec] = ...
 \end{verbatim}
+The file \file{COMPLETED} will appear in your run directory once the run is 
+finished.
+
+
 If you work with one of the samples or an identical setup in a new working directory, you can verify the correctness of the results
 by checking against reference data, delivered with each sample:
 \begin{verbatim}
