+++ doc/quick-guide/quick_start.tex
@@ -219,7 +219,7 @@ If everything worked well, your output should contain the lines
 \begin{verbatim}
   start.x has completed successfully
 \end{verbatim}
-after initializing everything successfully, it would then start running,   
+after initializing everything successfully. It would then start running,   
 printing in the console the quantities specified in \file{print.in}, for 
 instance,
 \begin{verbatim}
