### 2020-11-23 ce07491
changing obsolete python name functions in the quick start guide and adding python modules requirements
