+++ doc/quick-guide/quick_start.tex
@@ -365,7 +365,7 @@ been updated yet.
  
 
 In this example we use the modules: \code{numpy} and \code{matplotlib}.
-A complete list of requiered module is included in 
+A complete list of required module is included in 
 \file{pencil-code/python/pencil/README}.
 
 
