### 2020-11-23 202b33a
The subroutine k_omega_spectra now computes legendre coefficients of azimuthally averaged spectra
