+++ src/power_spectrum.f90
@@ -3999,7 +3999,8 @@ endsubroutine pdf
     use Chiral, only: iXX_chiral, iYY_chiral
 !
   integer, parameter :: nk=nxgrid/2
-  integer :: i, k, ikx, iky, ikz, jkz, im, in, ivec, ivec_jj, ikr, ikmu
+  integer :: i, k, ikx, iky, ikz, jkz, im, in, ivec, ivec_jj
+  integer :: ikr, ikmu
   integer, dimension(nk) :: nmu
   real, allocatable, dimension(:,:) :: kmu
   real :: k2,mu
@@ -4013,6 +4014,9 @@ endsubroutine pdf
   real, dimension(nk,nzgrid) :: cyl_spectrumhel, cyl_spectrumhel_sum
   real, allocatable, dimension(:,:) :: cyl_polar_spec, cyl_polar_spec_sum
   real, allocatable, dimension(:,:) :: cyl_polar_spechel, cyl_polar_spechel_sum
+  real, allocatable, dimension(:,:,:) :: legendre_p
+  real, dimension(legendre_lmax+1,nk) :: legendre_al, legendre_al_sum
+  real, dimension(legendre_lmax+1,nk) :: legendre_alhel, legendre_alhel_sum
   real, dimension(nxgrid) :: kx
   real, dimension(nygrid) :: ky
   real, dimension(nzgrid) :: kz
@@ -4046,10 +4050,12 @@ endsubroutine pdf
     cyl_spectrum_sum=0.
     cyl_spectrumhel=0.
     cyl_spectrumhel_sum=0.
-  ! allow for polar representation
+!
+! allow for polar representation
+!
     if (lcyl_polar_spectra) then
       do ikr=1,nk
-        nmu(ikr)=2*FLOOR(0.84815*ikr/2)+1
+        nmu(ikr)=2*(ikr-1)+1
       enddo
       !
       allocate( kmu(nk,nmu(nk)) )
@@ -4059,17 +4065,37 @@ endsubroutine pdf
       allocate( cyl_polar_spechel_sum(nk,nmu(nk)) )
       !
       kmu=0.
-      do ikr=3,nk
+      do ikr=2,nk
         do ikmu=1, nmu(ikr)
           kmu(ikr,ikmu) = -1+2.*(ikmu-1)/(nmu(ikr)-1)
         enddo
       enddo
-      !
+      ! 
       cyl_polar_spec=0.
       cyl_polar_spec_sum=0.
       cyl_polar_spechel=0.
       cyl_polar_spechel_sum=0.
+      legendre_al=0.
+      legendre_al_sum=0.
+      legendre_alhel=0.
+      legendre_alhel_sum=0.
       !
+      allocate( legendre_p(legendre_lmax+1,nk,nmu(nk)) )
+      legendre_p=0.
+      do ikr=1,nk
+      do ikmu=1, nmu(ikr)
+        ! legendre_p(i,:,:) is the (i-1)th order
+        legendre_p(1,ikr,ikmu)=1
+        legendre_p(2,ikr,ikmu)=kmu(ikr,ikmu)
+        if (legendre_lmax>=2) then
+          do i=3,legendre_lmax+1
+            legendre_p(i,ikr,ikmu)=1./i*( &
+              (2*i-1)*kmu(ikr,ikmu)*legendre_p(i-1,ikr,ikmu) &
+              -(i-1)*legendre_p(i-2,ikr,ikmu) )
+          enddo
+        endif
+      enddo
+      enddo
     endif
   endif
   !
@@ -4103,48 +4129,56 @@ endsubroutine pdf
[…]
