### 2020-11-23 ff56876
Rewrote subroutine corrfunc_3d and renamed it corfunc_cyl
