+++ src/power_spectrum.f90
@@ -3718,36 +3718,40 @@ endsubroutine pdf
   !
   endsubroutine anisoq_diag
 !***********************************************************************
-  subroutine corrfunc_3d(f)
+  subroutine corfunc_cyl(f,sp)
 !
-!  Calculate correlation functions <u_i^* u_j>, for i,j=1,2,3
-!  as functions of the vector k.
-!  I call the symmetric, real part of it vij,
-!  and (the antisymmetric, imaginary part / i) = wij.
-!  For the moment only u=velocity field,
+!  Calculate azimuthally averaged correlation functions v11,v33,v12.
+!  Here v11=(w11+w22)/2, v33=w33, v12=w12/sqrt(-1),
+!  where wij=<u_i^* u_j>, for i,j=1,2,3.
+!  They are functions kr=norm(kx,ky,kz) and kz/kr.
+!  For the moment u=velocity field only,
 !  but possible to extend using 'sp'
 !
-!  2020-Oct-14/hongzhe added this subroutine
+!  2020-Oct-14/hongzhe:  added this subroutine
+!  2020-Nov-23/hongzhe:  removed 3d outputs
     use Fourier, only: fft_xyz_parallel
     use Mpicomm, only: mpireduce_sum
     use Sub, only: del2vi_etc, del2v_etc, cross, grad, curli, curl, dot2
     use Chiral, only: iXX_chiral, iYY_chiral
 !
   integer, parameter :: nk=nxgrid/2
-  integer :: i, k, ikx, iky, ikz, jkx,jky,jkz, im, in, ivec, ivec_jj
+  integer :: i, ikx, iky, ikz, ikr, ikmu
+  integer, dimension(1) :: temploc
+  integer, dimension(nk) :: nmu
   real, dimension (mx,my,mz,mfarray) :: f
+  real, allocatable, dimension(:,:) :: kmu
   real, dimension(nx,ny,nz) :: ux_re, ux_im
   real, dimension(nx,ny,nz) :: uy_re, uy_im
   real, dimension(nx,ny,nz) :: uz_re, uz_im
-  real, dimension(nxgrid,nygrid,nzgrid) :: vxx, vxy, vyy, vxz, vyz, vzz
-  real, dimension(nxgrid,nygrid,nzgrid) :: wxy, wyz, wzx
-  real, dimension(nxgrid,nygrid,nzgrid) :: vxx_sum, vxy_sum, vyy_sum
-  real, dimension(nxgrid,nygrid,nzgrid) :: vxz_sum, vyz_sum, vzz_sum
-  real, dimension(nxgrid,nygrid,nzgrid) :: wxy_sum, wyz_sum, wzx_sum
-  real :: k2
-  real, dimension(nk,nzgrid) :: cyl_vxx, cyl_vxx_sum
-  real, dimension(nk,nzgrid) :: cyl_vzz, cyl_vzz_sum
-  real, dimension(nk,nzgrid) :: cyl_wxy, cyl_wxy_sum
+  real, allocatable, dimension(:,:) :: vxx, vxy, vzz
+  real, allocatable, dimension(:,:) :: vxx_sum, vxy_sum, vzz_sum
+  real, allocatable, dimension(:,:) :: coeff_a, coeff_b, coeff_c
[…]
