### 2020-11-23 3056398
Results of meeting
