+++ python/pencil/TO_DO_LIST.txt
@@ -59,4 +59,17 @@ Put simulations in dictionary to:
 * sort them
 * group them
 
+November 2020 meeting:
+Naming of files differ to functions/classes (Wolfgang - check reload options) 
+imports to be inside functions rather than file header (Wolfgang)
+use epydoc/sphinx and increase (Illa) Do we need to rewrite doc strings?
+Update to the Pencil Python wiki (Illa)
+Use black to run reformatting on Python ? (Wolfgang)
+Add tutorial examples to wiki and also the pencil/tutorials (Simon/Illa)
+Duplicate jupyter with python and ipython and update README (Simon)
+
+
+
+List here files that need to be reformatted according to Pencil style guide:
+
 
