### 2020-11-23 3c24c3c
Fix a bug from commit ff568769.
The interface to subroutine corfunc_cyl() was not complete.
