### 2020-11-23 1a31a72
Fix a bug in src/io_mpi2.
The informational output for how many slices are truncated was incorrect.

Also took the chance to compute a more precise cutoff time.
