### 2020-11-23 0d9aac7
Revert commit cf304c5d.
Since no auto-test is using this module yet, adopt a more modern
approach.
