### 2020-11-23 2329158
MR: corrected: lrmv is not only relevant for 1D-averages output
