### 2020-11-23 fbca756
MR: added initialization of email address of user
