### 2020-11-23 bcf7893
Added Kelvin-Helmholz time scale as output. Petri, please check!
