### 2020-11-24 513bd24
Align proc[xyz]_bound, xyz[01]_loc, and xyz[01].
Centralize the computation of

    proc[xyz]_bound
    xyz[01]_loc
    Lxyz_loc

under grid.f90, and align the two ends with global xyz[01] specified in
start.in.
