+++ src/grid.f90
@@ -97,9 +97,9 @@ module Grid
       real, dimension(my) :: g2,g2der1,g2der2,xi2,yprim2
       real, dimension(mz) :: g3,g3der1,g3der2,xi3,zprim2
 !
-      real, dimension(0:2*nprocx+1) :: xi1proc,g1proc
-      real, dimension(0:2*nprocy+1) :: xi2proc,g2proc
-      real, dimension(0:2*nprocz+1) :: xi3proc,g3proc
+      real, dimension(0:nprocx) :: xi1proc, g1proc
+      real, dimension(0:nprocy) :: xi2proc, g2proc
+      real, dimension(0:nprocz) :: xi3proc, g3proc
 !
       real :: a,b,c
       integer :: i
@@ -154,23 +154,26 @@ module Grid
 !  particle migration (see redist_particles_bounds). The select cases
 !  should use these arrays to set g{2,3}proc using the grid function.
 !
-      if (lparticles .or. lsolid_cells) then
-        do i=0,nprocx
-          xi1proc(2*i)  =i*nx-1
-          xi1proc(2*i+1)=i*nx
-        enddo
-        do i=0,nprocy
-          xi2proc(2*i)  =i*ny-1
-          xi2proc(2*i+1)=i*ny
-        enddo
-        do i=0,nprocz
-          xi3proc(2*i)  =i*nz-1
-          xi3proc(2*i+1)=i*nz
-        enddo
-        if (lperi(1)) xi1proc = xi1proc + 0.5
-        if (lperi(2)) xi2proc = xi2proc + 0.5
-        if (lperi(3)) xi3proc = xi3proc + 0.5
-      endif
+      xi1proc = (/ (real(i * nx), i = 0, nprocx) /)
+      xi1p: if (lactive_dimension(1) .and. .not. lperi(1)) then
+        xi1proc = xi1proc - 0.5
+        xi1proc(0) = 0.0
+        xi1proc(nprocx) = real(nxgrid - 1)
+      endif xi1p
+!
+      xi2proc = (/ (real(i * ny), i = 0, nprocy) /)
+      xi2p: if (lactive_dimension(2) .and. .not. (lperi(2) .or. lpole(2))) then
+        xi2proc = xi2proc - 0.5
+        xi2proc(0) = 0.0
+        xi2proc(nprocy) = real(nygrid - 1)
+      endif xi2p
+!
+      xi3proc = (/ (real(i * nz), i = 0, nprocz) /)
+      xi3p: if (lactive_dimension(3) .and. .not. lperi(3)) then
+        xi3proc = xi3proc - 0.5
+        xi3proc(0) = 0.0
+        xi3proc(nprocz) = real(nzgrid - 1)
+      endif xi3p
 !
 !  The following is correct for periodic and non-periodic case
 !    Periodic: x(xi=0) = x0 and x(xi=N) = x1
@@ -196,7 +199,8 @@ module Grid
         xprim2 = 0.
         dx_1 = 0.
         dx_tilde = 0.
-        g1proc=x00
+        g1proc(0) = x00
+        g1proc(1) = x00 + Lx
       else
 !
         select case (grid_func(1))
@@ -207,15 +211,12 @@ module Grid
           call grid_profile(a*(xi1  -xi1star),grid_func(1),g1,g1der1,g1der2)
           call grid_profile(a*(xi1lo-xi1star),grid_func(1),g1lo)
           call grid_profile(a*(xi1up-xi1star),grid_func(1),g1up)
+          call grid_profile(a * (xi1proc - xi1star), grid_func(1), g1proc)
 !
           x     =x00+Lx*(g1  -  g1lo)/(g1up-g1lo)
           xprim =    Lx*(g1der1*a   )/(g1up-g1lo)
[…]

[…]
