### 2020-11-24 8adb9ab
revised to use plegendre in general.f90
