+++ src/power_spectrum.f90
@@ -3732,7 +3732,7 @@ endsubroutine pdf
     use Fourier, only: fft_xyz_parallel
     use Mpicomm, only: mpireduce_sum
     use Sub, only: del2vi_etc, del2v_etc, cross, grad, curli, curl, dot2
-    use Chiral, only: iXX_chiral, iYY_chiral
+    use General, only: plegendre
 !
   integer, parameter :: nk=nxgrid/2
   integer :: i, ikx, iky, ikz, ikr, ikmu
@@ -3747,7 +3747,6 @@ endsubroutine pdf
   real, allocatable, dimension(:,:) :: vxx_sum, vxy_sum, vzz_sum
   real, allocatable, dimension(:,:) :: coeff_a, coeff_b, coeff_c
   real, allocatable, dimension(:,:) :: coeff_a_sum, coeff_b_sum, coeff_c_sum
-  real, allocatable, dimension(:,:,:) :: legendre_p
   real, dimension(legendre_lmax+1,nk) :: legendre_al_a, legendre_al_a_sum
   real, dimension(legendre_lmax+1,nk) :: legendre_al_b, legendre_al_b_sum
   real, dimension(legendre_lmax+1,nk) :: legendre_al_c, legendre_al_c_sum
@@ -3821,22 +3820,6 @@ endsubroutine pdf
   legendre_al_c=0.
   legendre_al_c_sum=0.
   !
-  allocate( legendre_p(legendre_lmax+1,nk,nmu(nk)) )
-  legendre_p=0.
-  do ikr=1,nk
-    do ikmu=1, nmu(ikr)
-      legendre_p(1,ikr,ikmu)=1  ! legendre_p(i,:,:) is the (i-1)th order
-      legendre_p(2,ikr,ikmu)=kmu(ikr,ikmu)
-      if (legendre_lmax>=2) then
-        do i=3,legendre_lmax+1
-          legendre_p(i,ikr,ikmu)=1./i*( &
-              (2*i-1)*kmu(ikr,ikmu)*legendre_p(i-1,ikr,ikmu) &
-              -(i-1)*legendre_p(i-2,ikr,ikmu) )
-        enddo
-      endif
-    enddo
-  enddo
-  !
   !  calculate each components
   !
   if (sp=='kin') then
@@ -3896,11 +3879,14 @@ endsubroutine pdf
       endif
       do i=1,legendre_lmax+1
         legendre_al_a(i,ikr)=legendre_al_a(i,ikr)+ &
-            1./nmu(ikr)*(2*i-1)/2*coeff_a(ikr,ikmu)*legendre_p(i,ikr,ikmu)
+            1./nmu(ikr)*(2*i-1)/2*coeff_a(ikr,ikmu)* &
+            sqrt(4*pi/(2.*i-1))*plegendre(i-1,0,kmu(ikr,ikmu))
         legendre_al_b(i,ikr)=legendre_al_b(i,ikr)+ &
-            1./nmu(ikr)*(2*i-1)/2*coeff_a(ikr,ikmu)*legendre_p(i,ikr,ikmu)
+            1./nmu(ikr)*(2*i-1)/2*coeff_a(ikr,ikmu)* &
+            sqrt(4*pi/(2.*i-1))*plegendre(i-1,0,kmu(ikr,ikmu))
         legendre_al_c(i,ikr)=legendre_al_c(i,ikr)+ &
-            1./nmu(ikr)*(2*i-1)/2*coeff_a(ikr,ikmu)*legendre_p(i,ikr,ikmu)
+            1./nmu(ikr)*(2*i-1)/2*coeff_a(ikr,ikmu)* &
[…]
