### 2020-11-24 749f090
MR: changes for proper notification after "file repair"
