### 2020-11-24 de70ff9
MR: initialize messages after read*pars
