### 2020-11-24 1847f08
Make MAUX CONTRIBUTION consistent in samples.
The following samples declared more auxiliaries than requested while
lwrite_aux = .true., which makes snapshots contain garbage data.

    samples/2d-tests/selfgravdisk-logspirals
    samples/debris-photoelectric-fluid
    samples/2d-tests/cylinder_ogrid_thermo_chem
    samples/2d-tests/WhiteDwarfDisk
    samples/kin-dynamo
    samples/2d-tests/debris-photoeletric-streaming
