+++ src/particles_dust_blocks.f90
@@ -112,7 +112,8 @@ module Particles
       remove_particle_at_time, remove_particle_criteria, remove_particle_criteria_size, &
       remove_particle_criteria_edtog, rad_sphere, pos_sphere, rhopmat, &
       a_ellipsoid, b_ellipsoid, c_ellipsoid, pos_ellipsoid, &
-      lrandom_particle_pencils, lnocalc_np, lnocalc_rhop, it1_loadbalance, &
+      lrandom_particle_pencils, it1_loadbalance, &
+      lnocalc_np, lnocalc_rhop, lcommunicate_rhop, lcommunicate_np, &
       np_const, rhop_const, lrandom_particle_blocks, lreblock_particles_run, &
       lbrick_partition, ldraglaw_variable, ladopt_own_light_bricks, &
       xp1, yp1, zp1, vpx1, vpy1, vpz1, xp2, yp2, zp2, vpx2, vpy2, vpz2, &
