+++ samples/2d-tests/WhiteDwarfDisk/run.in
@@ -63,5 +63,4 @@
   ldraglaw_variable=T
   gravr_profile='newtonian'
   lcylindrical_gravity_par=T
-  lcommunicate_rhop=T
 /
