+++ samples/2d-tests/WhiteDwarfDisk/start.in
@@ -46,6 +46,7 @@
   dustdensity_powerlaw=0.0
   bcpx='rmv',bcpy='p',bcpz='p'
 !                                                                                                
+  lcommunicate_rhop=T
   lparticlemesh_tsc=T
   tausp=1.0
   eps_dtog=1.
