+++ samples/2d-tests/debris-photoeletric-streaming/run.in
@@ -66,5 +66,4 @@
   ldraglaw_variable=T
   gravr_profile='newtonian'
   lcylindrical_gravity_par=T
-  lcommunicate_rhop=T
 /
