+++ samples/kin-dynamo/start.in
@@ -6,6 +6,7 @@
   cvsid='$Id$',
   random_gen='nr_f90'
   lwrite_aux=T
+  lkinflow_as_aux=T
 /
 &hydro_init_pars
 /
