### 2020-11-25 e24ee34
Explicitly check MAUX with number of auxiliaries.
If lwrite_aux == .true. and maux > naux, garbage data would be written
to each snapshot var.dat, VAR0, VAR1, etc..  In this case, fatal error
is issued to trap early problems of ill declaration of MAUX CONTRIBUTION.

If lwrite_aux == .false. and maux > naux, only a warning is printed.
