### 2020-11-25 0c93dec
Add keyword to py function pc.animate.slices().
Add keyword tmin to truncate slices with time earlier than tmin.
