### 2020-11-25 7dd36fb
minor changes to subroutines corfunc_cyl and k_omega_spectra
