+++ src/power_spectrum.f90
@@ -3779,6 +3779,7 @@ endsubroutine pdf
   !
   k2m=0.
   nks=0.
+  nmu=1
   do ikr=1,nk
     nmu(ikr)=2*(ikr-1)+1
   enddo
@@ -3868,7 +3869,7 @@ endsubroutine pdf
   do ikr=1,nk
     do ikmu=1,nmu(ikr)
       kmu2=kmu(ikr,ikmu)**2
-      coeff_c(ikr,ikmu)=vxy(ikr,ikmu)/(2*pi*kmu(ikr,ikmu))
+      if (kmu2/=0.) coeff_c(ikr,ikmu)=vxy(ikr,ikmu)/(2*pi*kmu(ikr,ikmu))
       if (kmu2==1.) then
         coeff_a(ikr,ikmu)=vxx(ikr,ikmu)/(pi*2.)
       else
@@ -3878,15 +3879,17 @@ endsubroutine pdf
             ( pi*(1-kmu2)**2*(2+kmu2) )
       endif
       do i=1,legendre_lmax+1
-        legendre_al_a(i,ikr)=legendre_al_a(i,ikr)+ &
-            1./nmu(ikr)*(2*i-1)/2*coeff_a(ikr,ikmu)* &
-            sqrt(4*pi/(2.*i-1))*plegendre(i-1,0,kmu(ikr,ikmu))
-        legendre_al_b(i,ikr)=legendre_al_b(i,ikr)+ &
-            1./nmu(ikr)*(2*i-1)/2*coeff_b(ikr,ikmu)* &
-            sqrt(4*pi/(2.*i-1))*plegendre(i-1,0,kmu(ikr,ikmu))
-        legendre_al_c(i,ikr)=legendre_al_c(i,ikr)+ &
-            1./nmu(ikr)*(2*i-1)/2*coeff_c(ikr,ikmu)* &
-            sqrt(4*pi/(2.*i-1))*plegendre(i-1,0,kmu(ikr,ikmu))
+        if (i<=nmu(ikr)) then  !  only meaningful when legendre order <= nmu-1
+          legendre_al_a(i,ikr)=legendre_al_a(i,ikr)+ &
+              1./nmu(ikr)*(2*i-1)/2*coeff_a(ikr,ikmu)* &
+              sqrt(4*pi/(2.*i-1))*plegendre(i-1,0,kmu(ikr,ikmu))
+          legendre_al_b(i,ikr)=legendre_al_b(i,ikr)+ &
+              1./nmu(ikr)*(2*i-1)/2*coeff_b(ikr,ikmu)* &
+              sqrt(4*pi/(2.*i-1))*plegendre(i-1,0,kmu(ikr,ikmu))
+          legendre_al_c(i,ikr)=legendre_al_c(i,ikr)+ &
+              1./nmu(ikr)*(2*i-1)/2*coeff_c(ikr,ikmu)* &
+              sqrt(4*pi/(2.*i-1))*plegendre(i-1,0,kmu(ikr,ikmu))
+        endif
       enddo
     enddo
   enddo
@@ -3910,19 +3913,39 @@ endsubroutine pdf
   if (lroot) then
     if (ip<10) print*,'Writing two point correlations to'  &
         ,trim(datadir)//'/cor2_.dat files'
-    open(1,file=trim(datadir)//'/cor2_la_a_'//trim(sp)//'.dat',position='append')
+    !  ABC coefficients
+    open(1,file=trim(datadir)//'/cor2_a_'//trim(sp)//'.dat',position='append')
+    write(1,*) t
+    do ikr=1,nk; do ikmu=1,nmu(ikr)
[…]
