### 2020-11-25 ead88d1
MR: added missing fnlen; minor; early setting of lcoarse
