+++ doc/quick-guide/quick_start.tex
@@ -215,7 +215,7 @@ It is now time to run the code:
 \begin{verbatim}
   pc_run
 \end{verbatim}
-If everything worked well, your output should contain the lines
+If everything worked well, your output should contain the line
 \begin{verbatim}
   start.x has completed successfully
 \end{verbatim}
