### 2020-11-25 72d16fb
MR: added  mailcmd
