### 2020-11-25 1e9cee0
MR: minor; adaptations for mailcmd
