+++ src/messages.f90
@@ -921,7 +921,7 @@ module Messages
 !
         if ( .not.lopen.and..not.lread ) &        ! send mail to user
           call system_cmd( &
-               'echo '//trim(errormsg)//'|'//trim(mailcmd)//"-s 'PencilCode Message' "//trim(mailaddress) )
+               'echo '//trim(errormsg)//'|'//trim(mailcmd)//"-s 'PencilCode Message' "//trim(mailaddress)//' >& /dev/null')
       endif
     endif
 !
