### 2020-11-26 9ce56af
Revemp map_nearest_grid() of particles_map.f90.
* Delegate the computation to module Grid.

* Speed up by using array arithmetic.
