### 2020-11-26 83c313c
Clean up particles_map.f90.
* Remove unused variables.

* Keep compiler quiet for dummy arguments.

* Initialize potentially uninitialized variables.

TODO: Deal with unused procedures interpolate_fourth() and
    invert_ineargrid_list().
