### 2020-11-26 92f3d80
Removed reding of jobid in simulation objects.
