### 2020-11-26 e1478c6
Converted jupyter notebook to script.
