### 2020-11-26 10fdef1
MR: avoid unnecessary writing of particle bounds
