### 2020-11-27 cf527a2
Add debug output to subroutine map_nearest_grid().
