### 2020-11-27 636fad2
Minor change in subroutine map_nearest_grid().
Attach the INTENT attributes to the argument declaration, instead of a
separate statement.
