### 2020-11-28 c70651a
Revise subroutine migrate_particles().
Use the spacing of xyz0 to quantumize the distance to the left boundary
of each process domain.  This spacing is what is used as the reference
point in subroutine inverse_grid() to find the nearest grid point to a
particle.

TODO: A more robust way to determine boundary crossing should be to
self-consistently use the index space between map_nearest_grid() and
migrate_particles().
