### 2020-11-28 84195ad
Revise the I/O of proc_bounds.dat.
In response to commit 10fdef1 and commit 4aeb457.

* The arrays proc[xyz]_bounds are the same irrespective of processes, so
  only one global file data/proc_bounds.dat is written and being read.

* For backward compatability, the grid is reconstructed and written
  whenever data/proc_bounds.dat does not exist.

* Note that the values of xyz[01]_loc and Lxyz_loc are aligned with and
  depends on proc[xyz]_bounds (see commit 513bd24).  These variables may
  become useful for diagnostic purposes, so they are constructed even
  when particles are not activated.
