### 2020-11-28 3079554
Minor change.
Some compiler complains the same if-label with an existing module name.
