+++ doc/citations/ref.bib
@@ -11134,3 +11134,74 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{2020ApJ...903..148L,
+       author = {{Li}, Xiang-Yu and {Mattsson}, Lars},
+        title = "{Dust Growth by Accretion of Molecules in Supersonic Interstellar Turbulence}",
+      journal = {\apj},
+     keywords = {Interstellar dust processes, Shocks, Interplanetary turbulence, 838, 2086, 830, Astrophysics - Astrophysics of Galaxies},
+         year = 2020,
+        month = nov,
+       volume = {903},
+       number = {2},
+          eid = {148},
+        pages = {148},
+          doi = {10.3847/1538-4357/abb9ad},
+archivePrefix = {arXiv},
+       eprint = {2009.00151},
+ primaryClass = {astro-ph.GA},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020ApJ...903..148L},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{2020arXiv201105556K,
+       author = {{Kahniashvili}, Tina and {Brandenburg}, Axel and {Gogoberidze}, Grigol and
+         {Mandal}, Sayan and {Roper Pol}, Alberto},
+        title = "{Circular Polarization of Gravitational Waves from Early-Universe Helical Turbulence}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics, General Relativity and Quantum Cosmology, High Energy Physics - Phenomenology, Physics - Fluid Dynamics},
+         year = 2020,
+        month = nov,
+          eid = {arXiv:2011.05556},
+        pages = {arXiv:2011.05556},
+archivePrefix = {arXiv},
+       eprint = {2011.05556},
+ primaryClass = {astro-ph.CO},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv201105556K},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{2020arXiv201107849K,
+       author = {{Klahr}, Hubert and {Schreiber}, Andreas},
+        title = "{Testing the Jeans, Toomre and Bonnor-Ebert concepts for planetesimal formation: 3D streaming instability simulations of diffusion regulated formation of planetesimals}",
+      journal = {arXiv e-prints},
+     keywords = {Astrophysics - Earth and Planetary Astrophysics},
+         year = 2020,
+        month = nov,
+          eid = {arXiv:2011.07849},
+        pages = {arXiv:2011.07849},
+archivePrefix = {arXiv},
+       eprint = {2011.07849},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020arXiv201107849K},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{2020ApJ...898..112P,
+       author = {{Park}, Kiwan},
+        title = "{Negative Magnetic Diffusivity {\ensuremath{\beta}} Replacing the {\ensuremath{\alpha}} Effect in the Helical Dynamo}",
+      journal = {\apj},
+     keywords = {Solar dynamo, Magnetohydrodynamics, Solar magnetic fields, Plasma astrophysics, 2001, 1964, 1503, 1261, Astrophysics - Solar and Stellar Astrophysics, Physics - Geophysics, Physics - Plasma Physics},
+         year = 2020,
+        month = aug,
+       volume = {898},
[…]
