+++ doc/citations/notes.tex
@@ -105,6 +105,7 @@ In addition, 80 papers reference it for code comparison or other purposes %(2020
 2020arXiv200710696K,%Klahr+Schreiber "Turbulence sets the length scale for planetesimal formation: Local 2D simulations of ..."
 2020arXiv200707325B,%Bhat+Zhou+Loureiro "Inverse energy transfer in decaying, three dimensional, nonhelical magnetic ..."
 2020MNRAS.496.4749B,%Brandenburg+Furuya "Application of a helicity proxy to edge-on galaxies"
+2020arXiv201105556K,%Kahniashvili+ "Circular Polarization of Gravitational Waves from Early-Universe Helical Turbulence"
 2020JPlPh..86c9018R,%Rudiger+Kuker+Kapyla "Electrodynamics of turbulent fluids with fluctuating electric conductivity"
 2020A&A...638A..69W,%Willamo+ "Shapes of stellar activity cycles"
 2020ApJ...896...86C,%Candelaresi+Del Sordo "Stabilizing Effect of Magnetic Helicity on Magnetic Cavities in the Intergalactic..."
@@ -115,7 +116,7 @@ Zhang_etal_2020comb,%Zhang+ "Drag force for a burning particle"
 2020arXiv200203638B,%Brandenburg "Piecewise quadratic growth during the 2019 novel coronavirus epidemic"
 2020arXiv200604426V,%Viviani+ "Physically motivated heat conduction treatment in simulations of solar-like stars: ..."
 2020arXiv200602971J,%Jakab+Brandenburg "The effect of a dynamo-generated field on the Parker wind"
-2020arXiv200900151L,%Li+Mattsson "Dust growth by accretion of molecules in supersonic interstellar turbulence"
+2020ApJ...903..148L,%Li+Mattsson "Dust growth by accretion of molecules in supersonic interstellar turbulence"
 2020arXiv200212172L,%Li+Mattsson "Coagulation of inertial particles in supersonic turbulence"
 2020A&A...633A.113A,%Adrover-Gonzalez+Terradas "3D numerical simulations of oscillations in solar prominences"
 2020ApJ...896L..14B,%Brandenburg+Bruggen "Hemispheric Handedness in the Galactic Synchrotron Polarization Foreground"
@@ -720,7 +721,7 @@ The {\sc Pencil Code} has been used for the following research topics
 {\em Interstellar and intercluster medium} \citep{
 2020arXiv201001833G,%Gent+ "Small-Scale Dynamo in Supernova-Driven Interstellar Turbulence"
 2020ApJ...896...86C,%Candelaresi+Del Sordo "Stabilizing Effect of Magnetic Helicity on Magnetic Cavities in the Intergalactic Medium
-2020arXiv200900151L,%Li+Mattsson "Dust growth by accretion of molecules in supersonic interstellar turbulence"
+2020ApJ...903..148L,%Li+Mattsson "Dust growth by accretion of molecules in supersonic interstellar turbulence"
 2020arXiv200212172L,%Li+Mattsson "Coagulation of inertial particles in supersonic turbulence"
 2020MNRAS.496.4749B,%Brandenburg+Furuya "Application of a helicity proxy to edge-on galaxies"
 2020ApJ...896L..14B,%Brandenburg+Bruggen "Hemispheric Handedness in the Galactic Synchrotron Polarization Foreground"
