+++ doc/citations/notes.tex
@@ -102,6 +102,7 @@ In addition, 80 papers reference it for code comparison or other purposes %(2020
 2020arXiv200908231B,%Brandenburg+ "{The Pencil Code, a modular MPI code for partial differential ..."
 2020ApJ...898..112P,%Park "Negative Magnetic Diffusivity {\ensuremath{\beta}} Replacing the {\ensuremath{\alpha}} Effect in the Helical Dynamo"
 2020arXiv200801119Z,%Zhu+Yang "Streaming Instability with Multiple Dust Species: I. Favourable Conditions for the Linear Growth"
+2020arXiv201107849K,%Klahr+ "Testing the Jeans, Toomre and Bonnor-Ebert concepts for planetesimal formation: 3D streaming ..."
 2020arXiv200710696K,%Klahr+Schreiber "Turbulence sets the length scale for planetesimal formation: Local 2D simulations of ..."
 2020arXiv200707325B,%Bhat+Zhou+Loureiro "Inverse energy transfer in decaying, three dimensional, nonhelical magnetic ..."
 2020MNRAS.496.4749B,%Brandenburg+Furuya "Application of a helicity proxy to edge-on galaxies"
@@ -829,45 +830,47 @@ The {\sc Pencil Code} has been used for the following research topics
 %d u s t  &  p l a n e t PLANERESIMAL
 \item
 {\em Planet formation} \citep{
+2020arXiv201107849K,%Klahr+ "Testing the Jeans, Toomre and Bonnor-Ebert concepts for planetesimal formation: 3D streaming ..."
+2020arXiv200710696K,%Klahr+Schreiber "Turbulence sets the length scale for planetesimal formation: Local 2D simulations of ..."
 2020MNRAS.491.4702Y,%Yang+Zhu "Morphological signatures induced by dust back reaction in discs with an embedded planet"
 2020A&A...635A.110E,%Eriksson+Johansen+Liu "Pebble drift and planetesimal formation in protoplanetary discs with embedded planets"
 2020ApJ...895...91G,%Gerbig+ "Requirements for Gravitational Collapse in Planetesimal Formation - The Impact of Scales ..."
 2019ApJ...887....6C,%Castrejon, Lyra,+ "Disentangling Planets from Photoelectric Instability in Gas-Rich Optically Thin Dusty Disks"
-2019ApJ...881..162B,% Baehr & Klahr "The Concentration and Growth of Solids in Fragmenting Circumstellar Disks"
+2019ApJ...881..162B,%Baehr & Klahr "The Concentration and Growth of Solids in Fragmenting Circumstellar Disks"
 2018JPhCS1031a2007M,%McNally+ "Low mass planet migration in Hall-affected disks"
-2018ApJ...861...47S,% Schreiber et al. Azimuthal and Vertical Streaming Instability at High Dust-to-gas Ratios..."
+2018ApJ...861...47S,%Schreiber et al. Azimuthal and Vertical Streaming Instability at High Dust-to-gas Ratios..."
 2019AAS...23336918H,%Hernandez+ "Migration of Embedded Black Holes in Active Galactic Nucleus Disk Simulations"
 2019Sci...364...66M,%Manser+...+Lyra+ "A planetesimal orbiting within the debris disc around a white dwarf star"
 2018ApJ...868...27Y,%Yang+ "Diffusion and Concentration of Solids in the Dead Zone of a Protoplanetary Disk"
-2018MNRAS.475..921R,% Rice & Nayakshin "On fragmentation of turbulent self-gravitating discs in the long cooling time regime"
[…]
