+++ doc/citations/ref.bib
@@ -11186,22 +11186,3 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@ARTICLE{2020ApJ...898..112P,
-       author = {{Park}, Kiwan},
-        title = "{Negative Magnetic Diffusivity {\ensuremath{\beta}} Replacing the {\ensuremath{\alpha}} Effect in the Helical Dynamo}",
-      journal = {\apj},
-     keywords = {Solar dynamo, Magnetohydrodynamics, Solar magnetic fields, Plasma astrophysics, 2001, 1964, 1503, 1261, Astrophysics - Solar and Stellar Astrophysics, Physics - Geophysics, Physics - Plasma Physics},
-         year = 2020,
-        month = aug,
-       volume = {898},
-       number = {2},
-          eid = {112},
-        pages = {112},
-          doi = {10.3847/1538-4357/ab9b89},
-archivePrefix = {arXiv},
-       eprint = {1911.01039},
- primaryClass = {astro-ph.SR},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2020ApJ...898..112P},
-      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
-}
-
