+++ doc/citations/ref.bib
@@ -10383,8 +10383,6 @@ archivePrefix = {arXiv},
        volume = {5},
         pages = {681-690},
           doi = {10.1016/j.idm.2020.08.014},
-          eid = {arXiv:2002.03638},
-        pages = {arXiv:2002.03638},
 archivePrefix = {arXiv},
        eprint = {2002.03638},
  primaryClass = {q-bio.PE},
@@ -11031,7 +11029,7 @@ archivePrefix = {arXiv},
 @ARTICLE{2020arXiv201007046B,
        author = {{Brandenburg}, Axel and {Das}, Upasana},
         title = "{Turbulent radiative diffusion and turbulent Newtonian cooling}",
-      journal = {arXiv e-prints},
+      journal = {$\!\!\!$},
      keywords = {Physics - Fluid Dynamics, Astrophysics - Solar and Stellar Astrophysics},
          year = 2020,
         month = oct,
@@ -11048,7 +11046,7 @@ archivePrefix = {arXiv},
        author = {{Gent}, Frederick A. and {Mac Low}, Mordecai-Mark and
          {Kapyla}, Maarit J. and {Singh}, Nishant K.},
         title = "{Small-Scale Dynamo in Supernova-Driven Interstellar Turbulence}",
-      journal = {arXiv e-prints},
+      journal = {$\!\!\!$},
      keywords = {Astrophysics - Astrophysics of Galaxies},
          year = 2020,
         month = oct,
@@ -11074,7 +11072,7 @@ archivePrefix = {arXiv},
          {Hubbard}, A. and {Guerrero}, G. and {Snodin}, A. and {Losada}, I.~R. and
          {Pekkil{\"a}}, J. and {Qian}, C.},
         title = "{The Pencil Code, a modular MPI code for partial differential equations and particles: multipurpose and multiuser-maintained}",
-      journal = {arXiv e-prints},
+      journal = {$\!\!\!$},
      keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Solar and Stellar Astrophysics, Physics - Fluid Dynamics},
          year = 2020,
         month = sep,
@@ -11089,7 +11087,7 @@ archivePrefix = {arXiv},
 
 @ARTICLE{2020ApJ...901...18B,
        author = {{Brandenburg}, Axel},
-        title = "{Hall Cascade with Fractional Magnetic Helicity in Neutron Star Crusts}",
+        title = "{Hall cascade with fractional magnetic helicity in neutron star crusts}",
       journal = {\apj},
      keywords = {Magnetohydrodynamical simulations, Magnetars, 1966, 992, Astrophysics - High Energy Astrophysical Phenomena, Physics - Fluid Dynamics},
          year = 2020,
@@ -11106,30 +11104,32 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-@MISC{2020zndo...3961647A,
-       author = {{Axel Brandenburg}, on behalf of the Pencil Code Collaboration},
-        title = "{Pencil Code}",
+@ARTICLE{2020zndo...3961647A,
+       author = {{Pencil Code Collaboration}},
+        title = "{The Pencil Code}",
+      journal = {$\!\!\!$},
      keywords = {Fluid dynamics, MHD, turbulent combustion, cloud microphysics, radiation, gravitational waves},
          year = 2020,
         month = aug,
-          eid = {10.5281/zenodo.3961647},
+        pages = {DOI:10.5281/zenodo.3961647},
+          eid = {DOI:10.5281/zenodo.3961647},
           doi = {10.5281/zenodo.3961647},
[…]
