### 2020-11-28 154d6df
modified according to the PCSC meeting
precising procedural aspects
