+++ doc/citations/ref.bib
@@ -11134,6 +11134,25 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+@ARTICLE{2020ApJ...901...54K,
+       author = {{Klahr}, Hubert and {Schreiber}, Andreas},
+        title = "{Turbulence Sets the Length Scale for Planetesimal Formation: Local 2D Simulations of Streaming Instability and Planetesimal Formation}",
+      journal = {\apj},
+     keywords = {Solar system formation, Protoplanetary disks, Planet formation, Planetesimals, Asteroids, Small solar system bodies, Classical Kuiper Belt objects, Trans-Neptunian objects, Comets, Hydrodynamical simulations, 1530, 1300, 1241, 1259, 72, 1469, 250, 1705, 280, 767, Astrophysics - Earth and Planetary Astrophysics},
+         year = 2020,
+        month = sep,
+       volume = {901},
+       number = {1},
+          eid = {54},
+        pages = {54},
+          doi = {10.3847/1538-4357/abac58},
+archivePrefix = {arXiv},
+       eprint = {2007.10696},
+ primaryClass = {astro-ph.EP},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2020ApJ...901...54K},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
 @ARTICLE{2020ApJ...903..148L,
        author = {{Li}, Xiang-Yu and {Mattsson}, Lars},
         title = "{Dust Growth by Accretion of Molecules in Supersonic Interstellar Turbulence}",
