+++ doc/citations/notes.tex
@@ -102,8 +102,8 @@ In addition, 80 papers reference it for code comparison or other purposes %(2020
 2020arXiv200908231B,%Brandenburg+ "{The Pencil Code, a modular MPI code for partial differential ..."
 2020ApJ...898..112P,%Park "Negative Magnetic Diffusivity {\ensuremath{\beta}} Replacing the {\ensuremath{\alpha}} Effect in the Helical Dynamo"
 2020arXiv200801119Z,%Zhu+Yang "Streaming Instability with Multiple Dust Species: I. Favourable Conditions for the Linear Growth"
+2020ApJ...901...54K,%Klahr+Schreiber "Turbulence sets the length scale for planetesimal formation: Local 2D simulations of ..."
 2020arXiv201107849K,%Klahr+ "Testing the Jeans, Toomre and Bonnor-Ebert concepts for planetesimal formation: 3D streaming ..."
-2020arXiv200710696K,%Klahr+Schreiber "Turbulence sets the length scale for planetesimal formation: Local 2D simulations of ..."
 2020arXiv200707325B,%Bhat+Zhou+Loureiro "Inverse energy transfer in decaying, three dimensional, nonhelical magnetic ..."
 2020MNRAS.496.4749B,%Brandenburg+Furuya "Application of a helicity proxy to edge-on galaxies"
 2020arXiv201105556K,%Kahniashvili+ "Circular Polarization of Gravitational Waves from Early-Universe Helical Turbulence"
@@ -831,7 +831,7 @@ The {\sc Pencil Code} has been used for the following research topics
 \item
 {\em Planet formation} \citep{
 2020arXiv201107849K,%Klahr+ "Testing the Jeans, Toomre and Bonnor-Ebert concepts for planetesimal formation: 3D streaming ..."
-2020arXiv200710696K,%Klahr+Schreiber "Turbulence sets the length scale for planetesimal formation: Local 2D simulations of ..."
+2020ApJ...901...54K,%Klahr+Schreiber "Turbulence sets the length scale for planetesimal formation: Local 2D simulations of ..."
 2020MNRAS.491.4702Y,%Yang+Zhu "Morphological signatures induced by dust back reaction in discs with an embedded planet"
 2020A&A...635A.110E,%Eriksson+Johansen+Liu "Pebble drift and planetesimal formation in protoplanetary discs with embedded planets"
 2020ApJ...895...91G,%Gerbig+ "Requirements for Gravitational Collapse in Planetesimal Formation - The Impact of Scales ..."
