### 2020-12-01 f4aa53f
Removed two parallel communications for checking the existence of a file.
