### 2020-12-01 ab9cdb6
Simplified code and do not override a given "ltruncate=.false." setting.
