### 2020-12-01 ae776a2
Disable a warning in io_mpi2.
Disable a warning from subroutine output_part_snap() of src/io_mpi2 when
the optional argument ltruncate is present.  This argument is needed for
src/io_hdf5 but has no obvious ways to be encapsulated therein.
