### 2020-12-01 639824a
Added comment regarding the ltruncate flag, as dicussed with Chao-Chin Yang.
