### 2020-12-03 e37fd12
Follow commit c70651a.
* Scale the spacing with Lxyz instead of xyz0.  This proves to be more
  accurate in my debugging process.

* Also allow a spacing to the right boundary to make ineargrid and
  boundary crossing numerically consistent.

* Make one sanity check more concise.
