### 2020-12-03 09b709d
MR: new: homogeneous part of test problems if njtest=6, diagnostic idiag_bhrms for that;          damp_uxb (<1) for damping the (uxb)' term in the test problems     minor polishments
