+++ src/test_methods/testfield_compress_z.f90
@@ -73,9 +73,9 @@ module Testfield
   character (len=labellen) :: itestfield='B11-B21',itestfield_method='(i)'
   real :: ktestfield=1., ktestfield1=1.
   real :: lin_testfield=0.,lam_testfield=0.,om_testfield=0.,delta_testfield=0.
-  real :: delta_testfield_next=0., delta_testfield_time=0.
   integer :: naainit
   real :: bamp=1.,bamp1=1.,bamp12=1.
+  real :: damp_uxb=0.
 
   namelist /testfield_init_pars/ &
        Btest_ext,zextent,initaatest, &
@@ -88,7 +88,7 @@ module Testfield
        lugutest, lfprestest, lSghtest
 
   ! run parameters
-  real :: etatest=0.,etatest1=0.,nutest=0.,nutest1=0.
+  real :: etatest=0.,etatest1=0.,nutest=0.,nutest1=0.,eta_hyper3_test=0.
   real :: ampl_fcont_aatest=1.,ampl_fcont_uutest=1.
   real, dimension(njtest) :: rescale_aatest=0.,rescale_uutest=0.,rescale_hhtest=0.
   logical :: ltestfield_newz=.true.,leta_rank2=.true.
@@ -113,7 +113,7 @@ module Testfield
        rescale_aatest,rescale_uutest, rescale_hhtest, rho0test, &
        lupw_uutest, lupw_hhtest, luse_main_run, lvisc_simplified_testfield, Omega, &
        lugutest, lfprestest, lSghtest, &
-       lremove_meanaa0x_test, lremove_meanaa0y_test
+       lremove_meanaa0x_test, lremove_meanaa0y_test, damp_uxb
 
   ! other variables (needs to be consistent with reset list below)
   integer :: idiag_alp11=0      ! DIAG_DOC: $\alpha_{11}$
@@ -196,6 +196,7 @@ module Testfield
   integer :: idiag_u0max=0      ! DIAG_DOC: $\operatorname{max}\left|\boldsymbol{u}_{0}\right|$
   integer :: idiag_b0max=0      ! DIAG_DOC: $\operatorname{max}\left|\boldsymbol{b}_{0}\right|$
   integer :: idiag_h0max=0      ! DIAG_DOC: $\operatorname{max}h_{0}$
+  integer :: idiag_bhrms=0      ! DIAG_DOC: $\left<b_{h}^2\right>^{1/2}$
   integer :: idiag_jb0m=0       ! DIAG_DOC: $\left<jb_{0}\right>$
   integer :: idiag_E11rms=0     ! DIAG_DOC: $\left<{\cal E}_{11}^2\right>^{1/2}$
   integer :: idiag_E21rms=0     ! DIAG_DOC: $\left<{\cal E}_{21}^2\right>^{1/2}$
@@ -421,7 +422,7 @@ module Testfield
 !  Override iE0 if njtest is big enough.
 !  This method of identifying the location of the reference field is not very elegant.
 !
-      if (njtest==5) iE0=5
+      if (njtest>=5) iE0=njtest
 !
 !  rescale the testfield
 !  (in future, could call something like init_aa_simple)
@@ -430,7 +431,7 @@ module Testfield
 !
 !  Don't rescale the reference field, so if existent rescale only to njtest-1
 !
-        if (iE0==5) then
+        if (iE0>=5) then
           nscal=4
         else
           nscal=njtest
@@ -574,7 +575,7 @@ module Testfield
 
       do j=1,ninit
 !
-!  select on initial condition for aatest
+!  select an initial condition for aatest
 !
         select case (initaatest(j))
[…]
