### 2020-12-05 d5df3c8
MR: corrected calc of lnrhorms
