### 2020-12-06 4138da0
MR: MPI_COMM_WORLD -> MPI_COMM_PENCIL
