### 2020-12-06 b5a4361
MR: MPI_COMM_WORLD -> MPI_COMM_PENCIL
