### 2020-12-07 b472108
added a logical ldensity_linearstart to start the linear density already in linear instead of in log and then taking the exponential
declared in cdata in set in density
