### 2020-12-07 28113df
Add Li-Ingersoll vortex in shallow water test (works in serial only).
