+++ src/initial_condition/shallow_water_init.f90
@@ -25,14 +25,14 @@ module InitialCondition
   real :: v_jet_peak,sigma_jet,r_jet_cent
   real :: Omega_vortex=1.
 ! Vortex parameters: 
-  real :: xv0,yv0, bv,rm,vm  
+  real :: xv0,yv0, bv,rm,vm,rout  
 !
   character (len=labellen), dimension(ninit) :: init_shallow_density='nothing'
   character (len=labellen), dimension(ninit) :: init_shallow_hydro='nothing'
 !
   namelist /initial_condition_pars/  eta0, k_eta, x0_drop, y0_drop, Omega_SB, &
        init_shallow_density,init_shallow_hydro,gamma_parameter,v_jet_peak, &
-       sigma_jet,r_jet_cent, Omega_vortex, xv0,yv0, bv,rm,vm
+       sigma_jet,r_jet_cent,Omega_vortex,xv0,yv0,bv,rm,vm,rout
 !
   real :: planetary_radius=impossible ! derives from gamma_parameter and Omega
 !
@@ -72,13 +72,16 @@ module InitialCondition
 !     h = eta + Lb
 !     rho is g*eta
 !
+      use General, only: notanumber
+!
       real, dimension (mx,my,mz,mfarray), intent(inout) :: f
       real, dimension (nx) :: eta,r2
-      integer :: j
+      integer :: j,i
 
       real, dimension(nx) :: rr,xx,ww
       real :: xs,ys,yy
       real :: b,rm,Ro,Bu
+      real :: phi
 
 !
       do j=1,ninit
@@ -114,12 +117,53 @@ module InitialCondition
             enddo
           enddo
 !
+        case('Li-Ingersoll')
+          do n=n1,n2
+             do m=m1,m2
+               xx = x(l1:l2)-xv0
+               yy = y(  m  )-yv0
+               rr = sqrt(xx**2 + yy**2)
+!
+               do i=l1,l2
+                 call calc_phi(rr(i-l1+1),phi)
+                 if (ldensity_linearstart) then
+                    f(i,m,n,irho) = - phi
+                 else
+                    call fatal_error("initial_condition_lnrho",&
+                         "switch ldensity_linearstart=T in density_init_pars")
+                 endif
+               enddo
+!               
+             enddo
+          enddo
+!
        endselect
 !
       enddo
 !
     endsubroutine initial_condition_lnrho
 !***********************************************************************
+    subroutine calc_phi(r,intgr)
+!
+      real :: Lr,dr,vr,tmp,rr
+      integer :: ir
+      real, intent(in) :: r
+      real, dimension(nx) :: a
+      real, intent(out) :: intgr
+!
+      Lr = rout - r
+      dr = Lr/nx
+      do ir=1,nx
+        rr = r + (ir-1)*dr
+        vr = vm * (rr/rm) * exp(1/bv * (1-(rr/rm)**bv))
+        tmp = vr**2/rr + 2*Omega_SB*vr 
+        a(ir) = tmp*dr
+      enddo
+!      
+      intgr = 0.5*(a(1)+a(nx)) + sum(a(2:nx-1))
+!      
+    endsubroutine calc_phi
+!***********************************************************************    
     subroutine initial_condition_uu(f)
 !
 !  Initial condition given by 
@@ -180,7 +224,8 @@ module InitialCondition
                vr = vm * (rr/rm) * exp(1/bv * (1-(rr/rm)**bv))
 !
                f(l1:l2,m,n,iux) = - vr * yy/rr
-               f(l1:l2,m,n,iuy) =   vr * xx/rr 
+               f(l1:l2,m,n,iuy) =   vr * xx/rr
+               
              enddo
           enddo
 !
