### 2020-12-08 767dd97
implemented gauss-legendre quadrature in power_spectrum.f90
