+++ src/power_spectrum.f90
@@ -34,9 +34,12 @@ module power_spectrum
   real :: pdf_max=30., pdf_min=-30., pdf_max_logscale=3.0, pdf_min_logscale=-3.
   logical :: lintegrate_shell=.true., lintegrate_z=.true., lcomplex=.false.
   logical :: lhalf_factor_in_GW=.false., lcylindrical_spectra=.false.
-  logical :: lcyl_polar_spectra=.false.
-  integer :: legendre_lmax=1.
+  logical :: lread_gauss_quadrature=.false.
+  integer :: legendre_lmax=1
   integer :: firstout = 0
+  logical :: lglq_dot_dat_exists=.false.
+  integer :: n_glq=1
+  real, allocatable, dimension(:,:) :: legendre_zeros,glq_weight
 !
   character (LEN=linelen) :: ckxrange='', ckyrange='', czrange=''
   integer, dimension(3,nk_max) :: kxrange=0, kyrange=0
@@ -48,12 +51,14 @@ module power_spectrum
       lintegrate_shell, lintegrate_z, lcomplex, ckxrange, ckyrange, czrange, &
       lcylindrical_spectra, inz, n_segment_x, lhalf_factor_in_GW, &
       pdf_max, pdf_min, pdf_min_logscale, pdf_max_logscale, &
-      lcyl_polar_spectra, legendre_lmax
+      lread_gauss_quadrature, legendre_lmax
 !
   contains
 !***********************************************************************
     subroutine initialize_power_spectrum
 !
+      use Messages
+      integer :: ikr, ikmu
       !!! the following warnings should become fatal errors
       if (nxgrid > nx) call warning ('power_spectrum', &
           "Part of the high-frequency spectrum are lost because nxgrid/= nx.")
@@ -61,6 +66,27 @@ module power_spectrum
           ((dx /= dz) .and. ((nxgrid-1)*(nzgrid-1) /= 0))) &
           call warning ('power_spectrum', &
           "Shell-integration will be wrong; set dx=dy=dz to fix this.")
+      !
+      !  07-dec-20/hongzhe: import gauss-legendre quadrature from gauss_legendre_quadrature.dat
+      !
+      if(lread_gauss_quadrature) then
+        inquire(FILE="gauss_legendre_quadrature.dat", EXIST=lglq_dot_dat_exists)
+        if (lglq_dot_dat_exists) then
+          open(9,file='gauss_legendre_quadrature.dat',status='old')
+          read(9,*) n_glq
+          if (n_glq<=legendre_lmax) call inevitably_fatal_error( &
+              'k_omega_spectra','either smaller lmax or larger gauss_legendre_quadrature.dat required')
+          allocate( legendre_zeros(n_glq,n_glq) )
+          allocate( glq_weight(n_glq,n_glq) )
+          do ikr=1,n_glq; do ikmu=1,n_glq
+            read(9,*) legendre_zeros(ikr,ikmu)
+            read(9,*) glq_weight(ikr,ikmu)
+          enddo; enddo
+          close(9)
+        else
[…]
