+++ src/power_spectrum.f90
@@ -3757,7 +3757,6 @@ endsubroutine pdf
 !  2020-Nov-23/hongzhe:  removed 3d outputs
     use Fourier, only: fft_xyz_parallel
     use Mpicomm, only: mpireduce_sum
-    use Sub, only: del2vi_etc, del2v_etc, cross, grad, curli, curl, dot2
     use General, only: plegendre
 !
   integer, parameter :: nk=nxgrid/2
@@ -3785,10 +3784,6 @@ endsubroutine pdf
   character (len=3) :: sp
   logical, save :: lwrite_krms=.true., lwrite_krms_GWs=.false.
 !
-!  passive scalar contributions (hardwired for now)
-!
-  real, dimension(nx,3) :: gtmp1,gtmp2
-!
 !  identify version
 !
   if (lroot .AND. ip<10) call svn_id("$Id$")
@@ -3805,25 +3800,43 @@ endsubroutine pdf
   !
   k2m=0.
   nks=0.
-  nmu=1
-  do ikr=2,nk
-    nmu(ikr)=2*(ikr-1)+3
-  enddo
-  allocate( kmu(nk,nmu(nk)) )
-  allocate( dmu(nk,nmu(nk)) )
-  kmu=0.
-  mu_offset=0.01
-  dmu=2-2*mu_offset
-  do ikr=1,nk
-    if (nmu(ikr)>=2) then
-      do ikmu=1, nmu(ikr)
-        kmu(ikr,ikmu) = -1+mu_offset+(ikmu-1)*(2-2*mu_offset)/(nmu(ikr)-1)
-        dmu(ikr,ikmu)=(2-2*mu_offset)/(nmu(ikr)-1)
+!
+! mesh for polar representation
+!
+  if (lread_gauss_quadrature) then  !  use gauss-legendre quadrature
+    allocate( kmu(nk,n_glq) )
+    allocate( dmu(nk,n_glq) )
+    do ikr=1,nk
+      nmu(ikr)=min( n_glq,max(1,3*(ikr-1)) )
+      do ikmu=1,nmu(ikr)
+        kmu(ikr,ikmu)=legendre_zeros(nmu(ikr),ikmu)
+        dmu(ikr,ikmu)=glq_weight(nmu(ikr),ikmu)
       enddo
-      dmu(ikr,1)=dmu(ikr,1)/2+mu_offset
-      dmu(ikr,nmu(ikr))=dmu(ikr,1)/2+mu_offset
-    endif
-  enddo
+    enddo
+  else  !  otherwise generate mesh by hand
+    nmu=1
+    do ikr=2,nk
+      nmu(ikr)=2*(ikr-1)+3
+    enddo
+    allocate( kmu(nk,nmu(nk)) )
+    allocate( dmu(nk,nmu(nk)) )
+    kmu=0.
+    mu_offset=0.01
+    dmu=2-2*mu_offset
+    do ikr=1,nk
+      if (nmu(ikr)>=2) then
+        do ikmu=1, nmu(ikr)
+          kmu(ikr,ikmu) = -1+mu_offset+(ikmu-1)*(2-2*mu_offset)/(nmu(ikr)-1)
+          dmu(ikr,ikmu)=(2-2*mu_offset)/(nmu(ikr)-1)
+        enddo
+        dmu(ikr,1)=dmu(ikr,1)/2+mu_offset
+        dmu(ikr,nmu(ikr))=dmu(ikr,nmu(ikr))/2+mu_offset
+      endif
+    enddo
+  endif
+!
+!  initialize spectra
+!
   allocate( vxx(nk,nmu(nk)) )
   allocate( vxx_sum(nk,nmu(nk)) )
   allocate( vzz(nk,nmu(nk)) )
@@ -4006,7 +4019,6 @@ endsubroutine pdf
[…]
