### 2020-12-08 7540898
MR: new switch  lzero_only for integration of the 0-problem only
