### 2020-12-09 883fb0a
made the resolution integration of the geopotential allocatable
