### 2020-12-09 98373a6
remove proc_bounds from the proc_dirs list
