+++ python/pencil/read/var.py
@@ -275,6 +275,7 @@ class DataCube(object):
                 proc_dirs = self.__natural_sort(
                     filter(lambda s: s.startswith('proc'),
                            os.listdir(datadir)))
+                proc_dirs.remove("proc_bounds.dat")
                 if param.lcollective_io:
                     # A collective IO strategy is being used
                     proc_dirs = ['allprocs']
