### 2020-12-09 118789b
add new lower limit for SLD characteristic speed from the sound speed
