### 2020-12-10 1d9e1a3
Merged three subroutines with similar functionalities: corfunc_cyl, anisoq_diag, k_omega_spectra -> polar_spectrum The new subroutine polar_spectrum computes, in k space, azimuthally averaged spectra in polar coordinates, and then performs a Legendre decomposition.
