+++ src/power_spectrum.f90
@@ -75,7 +75,8 @@ module power_spectrum
           open(9,file='gauss_legendre_quadrature.dat',status='old')
           read(9,*) n_glq
           if (n_glq<=legendre_lmax) call inevitably_fatal_error( &
-              'k_omega_spectra','either smaller lmax or larger gauss_legendre_quadrature.dat required')
+              'initialize_power_spectrum', &
+              'either smaller lmax or larger gauss_legendre_quadrature.dat required')
           allocate( legendre_zeros(n_glq,n_glq) )
           allocate( glq_weight(n_glq,n_glq) )
           do ikr=1,n_glq; do ikmu=1,n_glq
@@ -84,7 +85,8 @@ module power_spectrum
           enddo; enddo
           close(9)
         else
-          call inevitably_fatal_error('k_omega_spectra', 'you must give an input gauss_legendre_quadrature.dat file')
+          call inevitably_fatal_error('initialize_power_spectrum', &
+              'you must give an input gauss_legendre_quadrature.dat file')
         endif
       endif
 !
@@ -3520,519 +3522,50 @@ endsubroutine pdf
   !
   endsubroutine power_vec
 !***********************************************************************
-  subroutine anisoq_diag(f)
+  subroutine polar_spectrum(f,sp)
 !
-!  Anisotropic alpha2 dynamo diagostics.
-!  Calculate azimuthally averaged spectra of
-!  fluid (v.u), kinetic (u.o), magnetic (a.b), and current (b.j) helicities.
-!  Here v is the vector potential of u if u is incompressible.
-!
-!  16-sep-20/hongzhe: added this subroutine, modified from powerhel subroutine
-    use Fourier, only: fft_xyz_parallel
-    use Mpicomm, only: mpireduce_sum
-    use Sub, only: del2vi_etc, del2v_etc, cross, grad, curli, curl, dot2
-    use Chiral, only: iXX_chiral, iYY_chiral
-!
-  integer, parameter :: nk=nxgrid/2
-  integer :: i, k, ikx, iky, ikz, jkz, im, in, ivec, ivec_jj
-  real :: k2, ktot2
-  real, dimension (mx,my,mz,mfarray) :: f
-  real, dimension(nx,ny,nz) :: u_re, u_im, o_re, o_im
-  real, dimension(nx,ny,nz) :: a_re, a_im, b_re, b_im
-  real, dimension(nx) :: bbi, jji, b2, j2
-  real, dimension(nx,3) :: bb, bbEP, jj
-  real, dimension(nk) :: nks=0.,nks_sum=0.
-  real, dimension(nk) :: k2m=0.,k2m_sum=0.,krms
-  real, dimension(nk,nzgrid) :: vu_spec, vu_spec_sum
-  real, dimension(nk,nzgrid) :: uo_spec, uo_spec_sum
-  real, dimension(nk,nzgrid) :: ab_spec, ab_spec_sum
-  real, dimension(nk,nzgrid) :: bj_spec, bj_spec_sum
-  real, dimension(nxgrid) :: kx
-  real, dimension(nygrid) :: ky
-  real, dimension(nzgrid) :: kz
[…]
