### 2020-12-11 48d6ed4
added idiag_EEK and also inserted "oxu2; oxu(3);", but did not compute them yet
