### 2020-12-12 b381a08
MR: sanity check for positive definiteness after regularization of beta
