+++ src/special/meanfield_e_tensor.f90
@@ -35,7 +35,7 @@ module Special
   use Diagnostics
   use General, only: keep_compiler_quiet,numeric_precision
   use Messages, only: svn_id, fatal_error, warning
-  use Mpicomm, only: mpibarrier,MPI_COMM_WORLD,MPI_INFO_NULL,mpireduce_min, mpireduce_max
+  use Mpicomm, only: mpibarrier,MPI_COMM_WORLD,MPI_INFO_NULL,mpireduce_min, mpireduce_max,mpibarrier
 
   use Sub, only: dot_mn, dot_mn_vm, curl_mn, cross_mn, vec_dot_3tensor,dot2_mn
   use HDF5
@@ -50,7 +50,7 @@ module Special
   ! HDF debug parameters:
 
   integer :: hdferr
-  logical :: hdf_exists
+  logical :: hdf_grid_exists
 
   ! Main HDF object ids and variables
 
@@ -284,6 +284,7 @@ module Special
 !  6-oct-03/tony: coded
 !
       integer :: i
+      logical :: hdf_exists
 !
       if (lroot) call svn_id( &
            "$Id$")
@@ -323,7 +324,7 @@ module Special
 !
         call H5Fopen_F(hdf_emftensors_filename, H5F_ACC_RDONLY_F, hdf_emftensors_file, hdferr, access_prp = hdf_emftensors_plist)
 !
-! Checks whether in HDF5 file there is a link /emftensor/.
+! Checks whether in HDF5 file there is a group /emftensor/.
 !
         call H5Lexists_F(hdf_emftensors_file,'/emftensor/', hdf_exists, hdferr)
 
@@ -362,45 +363,53 @@ module Special
 
         if (trim(dataset) == 'time-series' .or. trim(dataset) == 'time-crop') then
           lread_time_series=.true.
-        else
-        !  call fatal_error('initialize_special','Unknown dataset chosen!')
-        end if
+        elseif (trim(dataset) /= 'mean') then
+          call fatal_error('initialize_special','Unknown dataset chosen!')
+        endif
+
+        if (lalt_decomp.or.lreconstruct_tensors) then
+          lacoef=.true.; lacoef=.true.
+        endif
 
         if (.not.lreloading) then
 
-          call H5Lexists_F(hdf_emftensors_file,'/grid/', hdf_exists, hdferr)
-          if (.not. hdf_exists) then
+          call H5Lexists_F(hdf_emftensors_file,'/grid/', hdf_grid_exists, hdferr)
+          if (.not. hdf_grid_exists) then
             call H5Fclose_F(hdf_emftensors_file, hdferr)
             call H5Pclose_F(hdf_emftensors_plist, hdferr)
             call H5close_F(hdferr)
-            call fatal_error('initialize_special','group /grid/ does not exist!')
-          end if
-
-          call H5Gopen_F(hdf_emftensors_file, 'grid', hdf_grid_group, hdferr)
-          if (hdferr /= 0) call fatal_error('initialize_special','error while opening /grid/')
-
-          call openDataset_grid(time_id)
-        
-          if (hdferr /= 0) call fatal_error('initialize special','cannot select grid/t')
-
-          call H5Dget_type_F(scalar_id_D(time_id), datatype_id, hdferr)
-          call H5Tequal_f(datatype_id, hdf_memtype, flag, hdferr)
-          if (.not.flag.and.lroot) &
-            call information('initialize_special','Type of stored HDF5 data different from type in memory - converting while reading!')
-
-          if (lread_time_series) then 
-
[…]
