+++ samples/dust-accretion/video.in
@@ -1,0 +1,5 @@
+lnrho
+nd
+nd2
+nd3
+nd4
