+++ samples/dust-accretion/print.in
@@ -1,0 +1,18 @@
+it(i9)
+t(e10.3)
+dt(e10.3)
+urms(e11.3)
+umax(e11.3)
+epsK(e11.3)
+rhom(e11.3)
+rho2m(e11.3)
+rhomin(e11.3)
+rhomax(e11.3)
+shockmax(e11.3)
+vprms(e11.3)
+vpm(e11.3)
+vpmax(e11.3)
+apm(e11.3)
+ap2m(e11.3)
+apmin(e11.3)
+apmax(e11.3)
