+++ samples/helical-MHDturb/GAUSS_LEGENDRE_QUAD/GLQ.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python
+#!/usr/bin/env python2
 
 import numpy as np
 nmax=36
