### 2020-12-14 0910b3c
resovled the MAUX problem for the sample "superparticle-condensation-coagulation"
