### 2020-12-14 498d20c
refined SN injection
