### 2020-12-14 3b30c12
Added gaussian noise random magentic field proportional to density
