### 2020-12-14 33fd54b
now correctly taking into consideration the jacobian during remeshing
