+++ src/power_spectrum.f90
@@ -3566,6 +3566,7 @@ endsubroutine pdf
   real, dimension(legendre_lmax+1,nk) :: legendre_al_c, legendre_al_c_sum
   !
   real, dimension(nx) :: bbi, jji
+  real, allocatable, dimension(:,:) :: jac !  jacobian when doing remeshing
   real, allocatable, dimension(:,:) :: polar_spec, polar_spec_sum
   real, allocatable, dimension(:,:) :: polar_spechel, polar_spechel_sum
   real, dimension(legendre_lmax+1,nk) :: legendre_al, legendre_al_sum
@@ -3593,13 +3594,16 @@ endsubroutine pdf
   if (lread_gauss_quadrature) then  !  use gauss-legendre quadrature
     allocate( kmu(nk,n_glq) )
     allocate( dmu(nk,n_glq) )
+    allocate( jac(nk,n_glq) )
     kmu=0.
     dmu=2.
     do ikr=1,nk
+      jac(ikr,:)=1./max(1,ikr-1)/2.
       nmu(ikr)=min( n_glq,max(1,3*(ikr-1)) )
       do ikmu=1,nmu(ikr)
         kmu(ikr,ikmu)=legendre_zeros(nmu(ikr),ikmu)
         dmu(ikr,ikmu)=glq_weight(nmu(ikr),ikmu)
+        jac(ikr,ikmu)=1./max(1,ikr-1)/dmu(ikr,ikmu)
       enddo
     enddo
   else  !  otherwise generate mesh by hand
@@ -3609,10 +3613,12 @@ endsubroutine pdf
     enddo
     allocate( kmu(nk,nmu(nk)) )
     allocate( dmu(nk,nmu(nk)) )
+    allocate( jac(nk,nmu(nk)) )
     kmu=0.
     mu_offset=0.01
     dmu=2-2*mu_offset
     do ikr=1,nk
+      jac(ikr,:)=1./max(1,ikr-1)/(2-2*mu_offset)
       if (nmu(ikr)>=2) then
         do ikmu=1, nmu(ikr)
           kmu(ikr,ikmu) = -1+mu_offset+(ikmu-1)*(2-2*mu_offset)/(nmu(ikr)-1)
@@ -3620,6 +3626,9 @@ endsubroutine pdf
         enddo
         dmu(ikr,1)=dmu(ikr,1)/2+mu_offset
         dmu(ikr,nmu(ikr))=dmu(ikr,nmu(ikr))/2+mu_offset
+        do ikmu=1, nmu(ikr)
+          jac(ikr,ikmu)=1./max(1,ikr-1)/dmu(ikr,ikmu)
+        enddo
       endif
     enddo
   endif
@@ -3654,31 +3663,31 @@ endsubroutine pdf
             temploc=minloc(abs(kmu(ikr+1,:)-mu))
             ikmu=temploc(1)
             if (mu==0. .and. mod(nmu(ikr+1),2)==0) then
-              vxx(ikr+1,ikmu)=vxx(ikr+1,ikmu)+0.5*0.5*( &
+              vxx(ikr+1,ikmu)=vxx(ikr+1,ikmu)+ jac(ikr+1,ikmu)*0.5*0.5*( &
                   +ux_re(ikx,iky,ikz)**2+ux_im(ikx,iky,ikz)**2 &
                   +uy_re(ikx,iky,ikz)**2+uy_im(ikx,iky,ikz)**2 )
-              vxx(ikr+1,ikmu+1)=vxx(ikr+1,ikmu+1)+0.5*0.5*( &
[…]
