### 2020-12-14 7614ae9
Remove files, which have been replaced with extended and documented counterparts.
