+++ src/particles_dust.f90
@@ -384,13 +384,13 @@ module Particles
         iapn = iapn(1) + indgen(ndustrad) - 1
         call farray_index_append('n_np_ap',ndustrad)
       end if
-      call put_shared_variable('iapn', iapn, caller='register_particles')
+      call put_shared_variable('iapn', iapn)
 !
 !  Rhop
 !
       if (.not. lnocalc_rhop) call farray_register_auxiliary('rhop',irhop, &
           communicated = lparticles_sink .or. lcommunicate_rhop)
-      call put_shared_variable('irhop', irhop, caller='register_particles')
+      call put_shared_variable('irhop', irhop)
       if (lcalc_uup .or. ldragforce_stiff) then
         call farray_register_auxiliary('uup',iuup,communicated=.true.,vector=3)
         iupx = iuup
@@ -561,7 +561,7 @@ module Particles
 !  Check if shear advection is on and decide if it needs to be included in the timestep condition.
 !
       if (lshear) then
-        call get_shared_variable('lshearadvection_as_shift', lshearadvection_as_shift)
+        call get_shared_variable('lshearadvection_as_shift', lshearadvection_as_shift, caller='initialize_particles')
         lcdtp_shear = .not. lshearadvection_as_shift
         nullify(lshearadvection_as_shift)
       endif
@@ -626,7 +626,7 @@ module Particles
 !  Share friction time (but only if Epstein drag regime!).
 !
       if (ldraglaw_epstein .or. ldraglaw_simple) then
-        call put_shared_variable( 'tausp_species', tausp_species)
+        call put_shared_variable( 'tausp_species', tausp_species, caller='initialize_particles')
         call put_shared_variable('tausp1_species',tausp1_species)
       endif
 !
@@ -662,7 +662,7 @@ module Particles
             rhom = mean_density(f)
             if (lreference_state) then
               call get_shared_variable('reference_state_mass',reference_state_mass, &
-                  caller = 'initialize_particles')
+                                       caller = 'initialize_particles')
               rhom = rhom+reference_state_mass/box_volume
             endif
           endif
@@ -697,9 +697,7 @@ module Particles
 !
 !  Share Keplerian gravity.
 !
-      call put_shared_variable('gravr',gravr,ierr)
-      if (ierr /= 0) call fatal_error('initialize_particles', &
-          'there was a problem when sharing gravr')
+      call put_shared_variable('gravr',gravr,caller = 'initialize_particles')
 !
 !  Inverse of minimum gas friction time (time-step control).
 !
@@ -761,6 +759,8 @@ module Particles
             'both. Stop and choose only one.'
         call fatal_error('initialize_particles','')
       endif
+      if (ldraglaw_steadystate .and. mean_free_path_gas==0. .and. .not.lstocunn1) &
+        call fatal_error('initialize_particles','mean_free_path_gas==0. for ldraglaw_steadystate=T, lstocunn1=F')
 !
 !  Stiff drag force approximation.
 !
@@ -946,10 +946,10 @@ module Particles
 !
       if (l_shell) then
         if ( k_shell < 0) call fatal_error('initialize_particles','Set k_shell')
-        call put_shared_variable('uup_shared',uup_shared,ierr)
-        call put_shared_variable('vel_call',vel_call,ierr)
-        call put_shared_variable('turnover_call',turnover_call,ierr)
-        call put_shared_variable('turnover_shared',turnover_shared,ierr)
+        call put_shared_variable('uup_shared',uup_shared)
[…]
