### 2020-12-15 ce0b3e6
set mean_free_path_gas for the superparticle simulation samples. Thank Matthias for spotting this
