### 2020-12-15 bc36bc1
added the mean_fre_path_gas for the superparticles simulations as spotted by Matthias.
