+++ samples/superparticle-coagulation/start.in
@@ -31,6 +31,7 @@
   lparticlemesh_tsc=F
   lnocalc_np=T
   lnocalc_rhop=T
+  mean_free_path_gas=6.8e-8 ! mean free path of air
 /
 &particles_radius_init_pars
   initap='lognormal', a0_initdist=10e-6, sigma_initdist=0.2
