### 2020-12-15 8d79221
MR: suppressed fatal error for naux/=maux in start
