+++ src/farray.f90
@@ -832,7 +832,7 @@ module FArrayManager
 !
 !  24-nov-2020/ccyang: coded
 !
-      use Cdata, only: lwrite_aux
+      use Cdata, only: lstart, lwrite_aux
 !
       character(len=*), parameter :: NEWLINE = achar(10) // achar(9) // "| "
       character(len=*), parameter :: ROUTINE = "farray_check_maux"
@@ -846,7 +846,7 @@ module FArrayManager
 !
       neq: if (naux /= maux) then
         if (lroot) print *, "farray_check_maux: naux, maux = ", naux, maux
-        if (lwrite_aux) then
+        if (.not.lstart.and.lwrite_aux) then
           call fatal_error(ROUTINE, MESSAGE)
         else
           call warning(ROUTINE, MESSAGE)
