### 2020-12-15 31c10cc
MR: added lrepair_snap; data strucrures for polar grid
