### 2020-12-16 c1412d3
MR: reinstated MAUX CONTRIBUTION 2 to fix the autotest for the moment
