### 2020-12-17 a93587a
corrected error in Bij for spherical coordinates
