### 2020-12-17 dfb7a0f
MR: fixes to make the auotest run
