+++ samples/superparticle-condensation-coagulation/run.in
@@ -47,6 +47,7 @@
   lcondensation_rate=T
   G_condensation = 1.17e-10
   loutput_psize_dist=T, log_ap_min_dist=-6., log_ap_max_dist=-3., nbin_ap_dist=50
+  mean_free_path_gas=6.8e-8 ! mean free path of air
 /
 &particles_radius_run_pars
   lcondensation_par=T
