### 2020-12-21 1e5bc16
added EEK(e9.2) and oxurms(e11.2) to print.in, and dsound=5 to run.in
