+++ samples/kin-dynamo/print.in
@@ -1,6 +1,7 @@
 it(i6)
 t(f8.3)
 dt(f8.3)
+EEK(e9.2)
 urms(e9.2)
 brms(e12.4)
 jrms(e11.2)
@@ -8,3 +9,5 @@ jmax(e11.2)
 abm(e11.2)
 jbm(e11.2)
 bmz(e11.2)
+oum(e11.2)
+oxurms(e11.2)
