### 2020-12-21 a41241d
added computation of oxurms
