+++ samples/kin-dynamo/print.in
@@ -9,5 +9,4 @@ jmax(e11.2)
 abm(e11.2)
 jbm(e11.2)
 bmz(e11.2)
-oum(e11.2)
-oxurms(e11.2)
+ourms(e11.2)
