### 2020-12-27 ba1f387
Revise the assignment of particle species.
Use the newly implemented function assign_species() in Particles_sub
module, which avoids potential overflow of integer mulplications.
