### 2020-12-27 154a3fc
Minor change.
Remove a redundant assignment of rhopswarm.
