### 2020-12-27 3f80162
Follow commit 6be7ab2.
Fortran 95 does not even allow intrinsic functions to be used to
initialize a parameter.  Resort to an inflexible, hard-coded interger
size to hold the computation, and leave the more flexible syntax as a
comment.
