+++ src/particles_sub.f90
@@ -44,7 +44,10 @@ module Particles_sub
 !
       integer, intent(in) :: ipar
 !
-      integer, parameter :: KIND = selected_int_kind(int(log10(real(npar_species)) + log10(real(npar))) + 1)
+!     Fortran 2003 syntax: flexible
+!      integer, parameter :: KIND = selected_int_kind(int(log10(real(npar_species)) + log10(real(npar))) + 1)
+!     Fortran 95 syntax: not flexible
+      integer, parameter :: KIND = selected_int_kind(9)
 !
       p = int(int(npar_species, kind=KIND) * int(ipar - 1, kind=KIND) / int(npar, kind=KIND)) + 1
 !
