### 2020-12-27 ee31878
Follow commit 3f80162.
Use a more conservative kind.
