+++ src/particles_sub.f90
@@ -47,7 +47,7 @@ module Particles_sub
 !     Fortran 2003 syntax: flexible
 !      integer, parameter :: KIND = selected_int_kind(int(log10(real(npar_species)) + log10(real(npar))) + 1)
 !     Fortran 95 syntax: not flexible
-      integer, parameter :: KIND = selected_int_kind(9)
+      integer, parameter :: KIND = selected_int_kind(12)
 !
       p = int(int(npar_species, kind=KIND) * int(ipar - 1, kind=KIND) / int(npar, kind=KIND)) + 1
 !
