### 2020-12-28 5b4f89b
update of prefactor in stress
