### 2020-12-30 e743bed
Need to remove proc_bounds.dat from proc_dirs also here.
