### 2020-12-30 ba640e3
Changes in the calculation of the Kelvin-Helmholtz timescale, Jörn please take a look.
