### 2021-01-03 6e7d0a0
added ydep_profile cos(y) profile
