### 2021-01-04 0a60961
MR: added switches lremove_E0, lremove_F0
