### 2021-01-04 ea28891
MR: removed unneeded ig[gxyz]; some simplifications
