+++ src/cdata.f90
@@ -332,7 +332,7 @@ module Cdata
   integer :: ibbt=0,ibxt=0,ibyt=0,ibzt=0,ijjt=0,ijxt=0,ijyt=0,ijzt=0, &
              ijxb=0, ijxbx=0, ijxby=0, ijxbz=0
   integer :: ibxf=0,ibyf=0,ibzf=0,ibbf=0
-  integer :: igg=0,igx=0,igy=0,igz=0,ipotself=0
+  integer :: ipotself=0
   integer :: iaa=0,iax=0,iay=0,iaz=0
   integer :: ispx=0,ispy=0,ispz=0
   integer :: ifcr=0,ifcrx=0,ifcry=0,ifcrz=0
