### 2021-01-04 5effd17
MR: safer allocation of KS_*
