### 2021-01-07 627c1a1
MR: new switch lconvergence_only (default: True)
