### 2021-01-07 cf9a884
Added diagnostics for horizontally averaged vertical force from the pressure gradient.
