### 2021-01-07 0aa305b
Added diagnostics for the vertical component of horizontally averaged viscous force.
