### 2021-01-07 93c6718
Added diagnostics for horizontal averages of the vertical component of the total acceleration and force.
