### 2021-01-09 bfcb3da
put wavenumber to zero if no extent
