### 2021-01-13 8232420
Streamlined the streamlines python routine.
