### 2021-01-13 4f09174
Improved syntax in fixed point finding python routine.
