+++ python/pencil/math/interpolation.py
@@ -38,50 +38,62 @@ def vec_int(xyz, field, dxyz, oxyz, nxyz, interpolation='trilinear'):
 
     import numpy as np
 
-    # Find the adjacent indices.
-    i = (xyz[0] - oxyz[0])/dxyz[0]
-    if i < 0:
-        i = 0
-    if i > nxyz[0] - 1:
-        i = nxyz[0] - 1
-    ii = np.array([int(np.floor(i)), int(np.ceil(i))])
+    if (interpolation == 'mean') or (interpolation == 'trilinear'):
+        # Find the adjacent indices.
+        i = (xyz[0]-oxyz[0])/dxyz[0]
+        ii = np.array([int(np.floor(i)), int(np.ceil(i))])
+        if i < 0:
+            i = 0
+        if i > nxyz[0]-1:
+            i = nxyz[0]-1
+        ii = np.array([int(np.floor(i)), int(np.ceil(i))])
 
-    j = (xyz[1] - oxyz[1])/dxyz[1]
-    if j < 0:
-        j = 0
-    if j > nxyz[1] - 1:
-        j = nxyz[1] - 1
-    jj = np.array([int(np.floor(j)), int(np.ceil(j))])
+        j = (xyz[1]-oxyz[1])/dxyz[1]
+        jj = np.array([int(np.floor(j)), int(np.ceil(j))])
+        if j < 0:
+            j = 0
+        if j > nxyz[1]-1:
+            j = nxyz[1]-1
+        jj = np.array([int(np.floor(j)), int(np.ceil(j))])
 
-    k = (xyz[2] - oxyz[2])/dxyz[2]
-    if k < 0:
-        k = 0
-    if k > nxyz[2] - 1:
-        k = nxyz[2] - 1
-    kk = np.array([int(np.floor(k)), int(np.ceil(k))])
+        k = (xyz[2]-oxyz[2])/dxyz[2]
+        kk = np.array([int(np.floor(k)), int(np.ceil(k))])
+        if k < 0:
+            k = 0
+        if k > nxyz[2]-1:
+            k = nxyz[2]-1
+        kk = np.array([int(np.floor(k)), int(np.ceil(k))])
 
     # Interpolate the field.
     if interpolation == 'mean':
-        return np.mean(field[:, kk[0]:kk[1]+1, jj[0]:jj[1]+1, ii[0]:ii[1]+1],
-                       axis=(1, 2, 3))
-    elif interpolation == 'trilinear':
+        sub_field = field[:, :, :, [ii[0], ii[1]]]
+        sub_field = sub_field[:, :, [jj[0], jj[1]], :]
+        sub_field = sub_field[:, [kk[0], kk[1]], :, :]
+        return np.mean(sub_field, axis=(1, 2, 3))
+
+    if interpolation == 'trilinear':
         if ii[0] == ii[1]:
             w1 = np.array([1, 1])
         else:
-            w1 = i - ii[::-1]
+            w1 = (i-ii[::-1])
+
         if jj[0] == jj[1]:
             w2 = np.array([1, 1])
         else:
-            w2 = j - jj[::-1]
+            w2 = np.array([nxyz[1]-j, j-jj[0]])
+
         if kk[0] == kk[1]:
             w3 = np.array([1, 1])
         else:
-            w3 = k - kk[::-1]
-        weight = abs(w1.reshape((2, 1, 1))*w2.reshape((1, 2, 1))*\
-                 w3.reshape((1, 1, 2)))
-        return np.sum(field[:, kk[0]:kk[1]+1, jj[0]:jj[1]+1,
-                            ii[0]:ii[1]+1]*weight, axis=(1, 2, 3)) \
[…]
