+++ python/pencil/read/var.py
@@ -275,7 +275,8 @@ class DataCube(object):
                 proc_dirs = self.__natural_sort(
                     filter(lambda s: s.startswith('proc'),
                            os.listdir(datadir)))
-                proc_dirs.remove("proc_bounds.dat")
+                if (proc_dirs.index("proc_bounds.dat") > 0):
+                    proc_dirs.remove("proc_bounds.dat")
                 if param.lcollective_io:
                     # A collective IO strategy is being used
                     proc_dirs = ['allprocs']
