### 2021-01-31 #2807 @rtfisher
I'm done with the review of most of the items. Overall I am impressed with the code framework, test suite, and documentation. I have just two issues.

One is that I do not have access to IDL, and so cannot directly confirm the graphical output presented for the test case. I'm not sure how much of an issue that is for JOSS.

Additionally, I don't agree with the statement, "While some SPH and PPM codes for hydrodynamics and magnetohydrodynamics are publicly available, this does not generally seem to be the case for higher order finite-difference or spectral codes." This may have been true at one point in time, but I know of at least SpECTRE (https://arxiv.org/abs/1609.00098) which is a DG code with exponential convergence on smooth problems, and therefore falls into precisely this category. There may be others. It seems to me they should be mentioned.
