### 2021-02-06 #2807 @AxelBrandenburg
Dear Arfon and Robert,

> I'm done with the review of most of the items. Overall I am impressed
> with the code framework, test suite, and documentation. I have just
> two issues.

Thanks, Robert, for your positive assessment.

> One is that I do not have access to IDL, and so cannot directly confirm
> the graphical output presented for the test case. I'm not sure how much
> of an issue that is for JOSS.

On page 9 of the manual,
https://github.com/pencil-code/website/blob/master/doc/manual.pdf
we have now added:

  The same can be achieved using Python (see Sect. 5.19.5 for details) with
    unix > python
    python > from matplotlib import pylab as plt
    python > var = pc.read.var(trimall=True)
    python > plt.imshow(var.uu[0, 0, :, : ].T, origin=’lower’)
  where we also make sure that the axis are shown in a natural way.

We also recall that in
https://github.com/pencil-code/pencil-code/blob/master/README.md
under "Documentation", item 3, it says:
Information about PythonCodingStyle and the PythonCodingStyle can be
found on the wiki.
The latter refers to https://github.com/pencil-code/pencil-code/wiki

> Additionally, I don't agree with the statement, "While some SPH
> and PPM codes for hydrodynamics and magnetohydrodynamics are publicly
> available, this does not generally seem to be the case for higher order
> finite-difference or spectral codes." This may have been true at one point
> in time, but I know of at least SpECTRE (https://arxiv.org/abs/1609.00098)
> which is a DG code with exponential convergence on smooth problems, and
> therefore falls into precisely this category. There may be others. It
> seems to me they should be mentioned.

We fully agree. This refers to the foreword of the manual, page vii,
lines 2-6, where we now write:

  While some SPH and PPM codes for hydrodynamics and magnetohydrodynamics
  were publicly available, this did not seem to be generally the case for
  higher order finite-difference or spectral codes.  This has changed
  since 2001; examples are the SpECTRE code, which is a discontinuous
  Galerkin code, and there are also the Snoopy and Dedalus codes, which
  are spectral.

In the actual JOSS submission (paper.md), we have now added under "Key references"
Another example is SpECTRE, a task-based discontinuous Galerkin code
for relativistic astrophysics [@Kidder_2017].

We hope that with this our JOSS submission satisfies now the
publication requirements.
Cheers, Axel
