### 2021-02-06 #2807 @AxelBrandenburg
PS: there was a small typo in our README.md and reply below:
"PythonCodingStyle and the PythonCodingStyle" ->
"Python with the Pencil Code and the Python Coding Style",
which is now fixed.
