### 2021-02-20 #2807 @AxelBrandenburg
Thanks, Arfon, for the promising news. I have now made a new release and uploaded it on zenodo; see
https://zenodo.org/record/4553325
Let me know if anything is not quite right yet.
Cheers, Axel

Sent from my iPad

> On 20 Feb 2021, at 12:44, Arfon Smith <notifications@github.com> wrote:
> Zenodo/figshare archive
