### 2021-02-21 #2807 @rtfisher
Thank _you_ Arfon and Axel both for being extraordinarily patient with my
review. I look forward to working with you again in the future.

Best wishes,

Bob

On Sun, Feb 21, 2021 at 7:17 AM Arfon Smith <notifications@github.com>
wrote:

> @zingale <https://github.com/zingale>, @rtfisher
> <https://github.com/rtfisher> - many thanks for your reviews here. JOSS
> relies upon the volunteer efforts of folks like yourselves, and we simply
> wouldn't be able to do this without you ✨
>
> @AxelBrandenburg <https://github.com/AxelBrandenburg> - your paper is now
> accepted and published in JOSS ⚡🚀💥
>
